### 2021-11-02 #3878 @jedbrown
@mscroggs :wave: Thanks for your submission to JOSS. This looks like a really useful package. Some notes:

1. It looks like the C++ code makes heavy use of xtensor, which thus deserves citation.
2. The paper doesn't mention quadrature, but the choice of quadrature is significant (especially for tetrahedra) so it would be useful to state/cite supported methods.
3. The documentation site lacks C++ examples. I think it would be useful to have at least one example that has equivalent C++ and Python, considering that low-dependency C++ is one of the selling points of Basix relative to FIAT.
4. The docs have a nice breakdown of element topology and associated FE families -- I think it would be nice to include that in the paper, versus what is currently summarized in one sentence of the paper.
5. Is there any attempt or interest in preserving tensor product structure for bases that have it?

Once I hear from you on these points, I'll move forward with finding reviewers.
