### 2021-11-03 #3878 @mscroggs
@jedbrown Thanks for these comments, I'll push some updates to the documentation/paper and let you know when these are done
