+++ joss/paper.bib
@@ -200,3 +200,29 @@
  HOWPUBLISHED = {\url{https://defelement.com}},
          NOTE = {[Online; accessed 26-October-2021]}
 }
+
+@COMMENT{
+  The authors are a list of all people with >=5 commits and >=100 lines of code committed to xtensor in alphabetical order.
+}
+@misc{xtensor,
+        TITLE = {{xtensor}: C++ tensors with broadcasting and lazy computing},
+        AUTHOR = {Thorsten Beier and Stuarte Berg and Patrick Bos and David Brochart and
+                  Sylvain Corlay and Tom {de Geus} and Adrien Delsalle and Mario Emenlauer
+                  and Victor Evangelista and Loic Gouarin and Serge Guelton and Kenneth Hanley
+                  and Ullrich Koethe and {GitHub user kolibri91} and Astra Kolomatskaia and
+                  Johan Mabille and Antoine Prouvost and Martin Renou and
+                  {GitHub user SoundDev} and Wolf Vollprecht and Ray Zhangand Jun Zhu},
+          URL = {https://github.com/xtensor-stack/xtensor},
+         YEAR = {2021}
+}
+
+@article{xiao-gimbutas,
+       AUTHOR = {Hong Xiao and Zydrunas Gimbutas},
+        TITLE = {A numerical algorithm for the construction of efficient quadrature rules in two and higher dimensions},
+      JOURNAL = {Computers & Mathematics with Applications},
+       VOLUME = {59},
+       NUMBER = {2},
+        PAGES = {{663--676}},
+         year = {2010},
+          doi = {10.1016/j.camwa.2009.10.027},
+}
