+++ doc/web/make_html.py
@@ -221,6 +221,13 @@ os.system(f"rm {path('../../demo/python/*.py.rst')}")
 system(f"cd {path('../../demo/python')} && python3 convert_to_rst.py")
 system(f"mkdir {path('python/source/demo')}")
 system(f"cp {path('../../demo/python/*.rst')} {path('python/source/demo')}")
+for file in os.listdir(path("python/source/demo")):
+    if file.endswith(".py.rst"):
+        with open(path(f"python/source/demo/{file}")) as f:
+            content = f.read()
+        content = content.replace("literalinclude:: ../", "literalinclude:: ../../../../../demo/")
+        with open(path(f"python/source/demo/{file}"), "w") as f:
+            f.write(content)
 
 with open(path("python/source/index.rst"), "a") as f:
     f.write("\n")
