### 2021-11-03 65d3544
put prints in the same order as Python
