+++ demo/cpp/demo_create_and_tabulate/main.cpp
@@ -38,9 +38,9 @@ int main(int argc, char* argv[])
 
   xt::xtensor<double, 4> tab = lagrange.tabulate(0, points);
 
-  std::cout << "\nTabulate data shape: " << xt::adapt(tab.shape());
   std::cout << "\nTabulate data: \n"
             << xt::view(tab, 0, xt::all(), xt::all(), 0);
+  std::cout << "\nTabulate data shape: " << xt::adapt(tab.shape());
 
   return 0;
 }
