### 2021-11-03 1a3f135
adjust author criteria, and edit paper authors to match
