+++ joss/paper.bib
@@ -32,33 +32,6 @@
        EPRINT = {2102.11901},
 }
 
-@COMMENT{
-  The authors are a list of all people with >=5 commits and >=100 lines of code committed to DOLFINx since 2018 in alphabetical order.
-}
-@misc{dolfinx,
-        TITLE = {DOLFINx: Next generation FEniCS problem solving environment},
-        AUTHOR = {Francesco Ballarin and Igor A. Baratta and Jan Blechta
-        and Joseph P. Dean and J{\o}rgen S. Dokken and Jack S. Hale and
-        Michal Habera and Massimiliano Leoni and Matthias Rambausek and
-        Chris N. Richardson and Matthew W. Scroggs and Nathan Sime and
-        Garth N. Wells},
-          URL = {https://github.com/FEniCS/dolfinx},
-         YEAR = {2021}
-}
-
-@COMMENT{
-  The authors are a list of all people with >=5 commits and >=100 lines
-  of code committed to FFCx since 2018 in alphabetical order.
-}
-@misc{ffcx,
-        TITLE = {FFCx: Next generation FEniCS Form Compiler},
-        AUTHOR = {Igor A. Baratta and Jan Blechta and J{\o}rgen S.
-        Dokken and Michal Habera and Jack S. Hale and Chris N.
-        Richardson and Matthew W. Scroggs and Garth N. Wells},
-          URL = {https://github.com/FEniCS/ffcx},
-         YEAR = {2021}
-}
-
 @article{ufl,
        AUTHOR = {Martin S. Aln{\ae}s and Anders Logg and Kristian B. {\O}lgaard and Marie E. Rognes and Garth N. Wells},
         TITLE = {{U}nified {F}orm {L}anguage: A domain-specific language for weak formulations of partial differential equations},
@@ -201,21 +174,6 @@
          NOTE = {[Online; accessed 26-October-2021]}
 }
 
-@COMMENT{
-  The authors are a list of all people with >=5 commits and >=100 lines of code committed to xtensor in alphabetical order.
-}
-@misc{xtensor,
-        TITLE = {{xtensor}: C++ tensors with broadcasting and lazy computing},
-        AUTHOR = {Thorsten Beier and Stuarte Berg and Patrick Bos and David Brochart and
-                  Sylvain Corlay and Tom {de Geus} and Adrien Delsalle and Mario Emenlauer
-                  and Victor Evangelista and Loic Gouarin and Serge Guelton and Kenneth Hanley
-                  and Ullrich Koethe and {GitHub user kolibri91} and Astra Kolomatskaia and
-                  Johan Mabille and Antoine Prouvost and Martin Renou and
-                  {GitHub user SoundDev} and Wolf Vollprecht and Ray Zhangand Jun Zhu},
-          URL = {https://github.com/xtensor-stack/xtensor},
-         YEAR = {2021}
-}
-
 @article{xiao-gimbutas,
        AUTHOR = {Hong Xiao and Zydrunas Gimbutas},
         TITLE = {A numerical algorithm for the construction of efficient quadrature rules in two and higher dimensions},
@@ -225,4 +183,44 @@
         PAGES = {{663--676}},
          year = {2010},
           doi = {10.1016/j.camwa.2009.10.027},
+}
+
+@COMMENT{
+  The authors are a list of all people with >=5 commits and >=1000 total lines of code committed to DOLFINx since 2018
+  with the user with most total lines first and everyone else in alphabetical order.
+}
+@misc{dolfinx,
+        TITLE = {DOLFINx: Next generation FEniCS problem solving environment},
+        AUTHOR = {Garth N. Wells and Francesco Ballarin and Igor A. Baratta
[…]
