### 2021-11-03 375a659
Move tdim into assert
