### 2021-11-03 ed58318
Cast tdim to size_t
