### 2021-11-03 eff45b8
add second cpp demo and some improvements in test.py
