### 2021-11-04 2b8c772
Add second cpp demo and some improvements to test.py (#325)
* add second cpp demo and some improvements in test.py

* include cpp demo at end of python demo on website

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>
