+++ demo/cpp/demo_dof_transformations/main.cpp
@@ -1,0 +1,212 @@
+// ====================================
+// DOF permutations and transformations
+// ====================================
+
+// When using high degree finite elements on general meshes, adjustments
+// may need to be made to correct for differences in the orientation of
+// mesh entities on the mesh and on the reference cell. For example, in
+// a degree 4 Lagrange element on a triangle, there are 3 degrees of
+// freedom (DOFs) associated with each edge. If two neighbouring cells in
+// a mesh disagree on the direction of the edge, they could put an
+// incorrectly combine the local basis functions to give the wrong global
+// basis function.
+
+// This issue and the use of permutations and transformations to correct
+// it is discussed in detail in `Construction of arbitrary order finite
+// element degree-of-freedom maps on polygonal and polyhedral cell
+// meshes (Scroggs, Dokken, Richardons, Wells,
+// 2021) <https://arxiv.org/abs/2102.11901>`_.
+
+// Functions to permute and transform high degree elements are
+// provided by Basix. In this demo, we show how these can be used from C++.
+
+#include <basix/finite-element.h>
+#include <basix/lattice.h>
+#include <iomanip>
+#include <xtensor/xio.hpp>
+
+static const std::map<basix::cell::type, std::string> type_to_name
+    = {{basix::cell::type::point, "point"},
+       {basix::cell::type::interval, "interval"},
+       {basix::cell::type::triangle, "triangle"},
+       {basix::cell::type::tetrahedron, "tetrahedron"},
+       {basix::cell::type::quadrilateral, "quadrilateral"},
+       {basix::cell::type::pyramid, "pyramid"},
+       {basix::cell::type::prism, "prism"},
+       {basix::cell::type::hexahedron, "hexahedron"}};
+
+int main(int argc, char* argv[])
+{
+  // Degree 5 Lagrange element
+  // =========================
+  {
+    std::cout << "Degree 5 Lagrange element \n";
+
+    // Create a degree 5 Lagrange element on a triangle
+    auto family = basix::element::family::P;
+    auto cell_type = basix::cell::type::triangle;
+    int degree = 5;
+    auto variant = basix::element::lagrange_variant::equispaced;
+
+    // Create the lagrange element
+    basix::FiniteElement lagrange
+        = basix::create_element(family, cell_type, degree, variant);
+
+    // Print bools as true/false instead of 0/1
+    std::cout << std::boolalpha;
+
+    // The value of `dof_transformations_are_identity` is `false`: this tells
+    // us that permutations or transformations are needed for this element.
+    std::cout << "Dof transformations are identity: "
+              << lagrange.dof_transformations_are_identity() << "\n";
+
+    // The value of `dof_transformations_are_permutations` is `true`: this
+    // tells us that for this element, all the corrections we need to apply
+    // permutations. This is the simpler case, and means we make the
+    // orientation corrections by applying permutations when creating the
+    // DOF map.
+    std::cout << "Dof transformations are permutations: "
+              << lagrange.dof_transformations_are_permutations() << "\n";
+
+    // We can apply permutations by using the matrices returned by the
+    // method `base_transformations`. This method will return one matrix
+    // for each edge of the cell (for 2D and 3D cells), and two matrices
+    // for each face of the cell (for 3D cells). These describe the effect
+    // of reversing the edge or reflecting and rotating the face.
+
+    // For this element, we know that the base transformations will be
+    // permutation matrices.
+    std::cout << "\nBase transformations: \n";
+    std::cout << lagrange.base_transformations() << "\n";
+
+    // The matrices returned by `base_transformations` are quite large, and
[…]
