+++ demo/cpp/demo_dof_transformations/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.16)
-project(demo_create_and_tabulate)
+project(demo_dof_transformations)
 
 # Use C++17
 set(CMAKE_CXX_STANDARD 17)
