### 2021-11-08 e50f1ef
Update JOSS paper (#324)
* Update JOSS paper

* Add Makefile

* Put C++ demo in a folder

* put prints in the same order as Python

* 3

* adjust author criteria, and edit paper authors to match

* remove duplicate name

* Add second cpp demo and some improvements to test.py (#325)

* add second cpp demo and some improvements in test.py

* include cpp demo at end of python demo on website

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>

Co-authored-by: Igor Baratta <igorbaratta@gmail.com>
