+++ joss/paper.md
@@ -7,12 +7,15 @@ tags:
   - basis functions
   - numerical analysis
 authors:
-  - name: Chris N. Richardson
-    orcid: 0000-0003-3137-1392
-    affiliation: 1
   - name: Matthew W. Scroggs
     orcid: 0000-0002-4658-2443
     affiliation: 2
+  - name: Igor A. Baratta
+    orcid:  0000-0003-4298-2973
+    affiliation: 2
+  - name: Chris N. Richardson
+    orcid: 0000-0003-3137-1392
+    affiliation: 1
   - name: Garth N. Wells
     orcid: 0000-0001-5291-7951
     affiliation: 2
@@ -55,7 +58,7 @@ basis functions of this space are shown in \autoref{fig:fe}.
 The uppper three functions arise from point evaluations at the vertices.
 The lower three arise from point evaluations at the midpoints of the
 edges. These diagrams are taken from DefElement
-[@defelement].\label{fig:fe}](basis-functions.png){ width=60% }
+[@defelement].\label{fig:fe}](img/basis-functions.png){ width=60% }
 
 The functionals in $\mathcal{L}$ are each associated with a degree of
 freedom (DOF) of the finite element space. Each functional (or DOF) is
@@ -66,12 +69,15 @@ element space the desired continuity properties.
 
 Basix is a C++ library that creates and tabulates a range finite
 elements on triangles, tetrahedra, quadrilaterals, hexahedra, pyramids,
-and prisms. Currently supported element types include Lagrange, Nédélec
-first kind [@nedelec1], Nédélec second kind [@nedelec2], Raviart--Thomas
-[@rt], Brezzi--Douglas--Marini [@bdm], Crouzeix--Raviart [@cr],
-serendipity [@serendipity; @sdivcurl], and Regge [@regge; @regge2]
-elements. The majority of Basix's functionality can be used via the
-library's Python interface.
+and prisms. A full list of currently supported elements is included
+below.
+
+For many elements, the functionals in $\mathcal{L}$ are defined to be integrals
+on a sub-entity of the cell. The compute these integrals, Basix provides a
+range of quadrature rules, including Gauss--Jacobi, Gauss--Lobatto--Legendre,
+and Xiao--Gimbutas [@xiao-gimbutas]. Internally, Basix uses xtensor [@xtensor]
+for matrix and tensor storage and manipulation. The majority of Basix's
+functionality can be used via the library's Python interface.
 
 Basix forms part of FEniCSx alongside DOLFINx [@dolfinx], FFCx [@ffcx],
 and UFL [@ufl]. FEniCSx is the latest development version of FEniCS, a
@@ -120,5 +126,106 @@ transforming DOFs for higher-order elements. As described in
 @dof-transformations, these operations are necessary when solving
 problems on arbitrary meshes, as differences in how neighbouring cells
 orient their sub-entities can otherwise cause issues.
+
+# Supported elements
+
+## Interval
+
+In Basix, the sub-entities of the reference interval are numbered as
+shown in \autoref{fig:interval}. The following elements are supported on a interval:
+
+- Lagrange
+- bubble
+- serendipity [@serendipity]
+
+![The numbering of a reference interval. \label{fig:interval}](img/interval_numbering.png){ width=39% }
+
+## Triangle
+
+In Basix, the sub-entities of the reference triangle are numbered as
+shown in \autoref{fig:triangle}.
+The following elements are supported on a triangle:
+
+- Lagrange
+- Nédélec first kind [@nedelec1]
+- Raviart--Thomas [@rt]
[…]
