+++ joss/paper.bib
@@ -32,33 +32,6 @@
        EPRINT = {2102.11901},
 }
 
-@COMMENT{
-  The authors are a list of all people with >=5 commits and >=100 lines of code committed to DOLFINx since 2018 in alphabetical order.
-}
-@misc{dolfinx,
-        TITLE = {DOLFINx: Next generation FEniCS problem solving environment},
-        AUTHOR = {Francesco Ballarin and Igor A. Baratta and Jan Blechta
-        and Joseph P. Dean and J{\o}rgen S. Dokken and Jack S. Hale and
-        Michal Habera and Massimiliano Leoni and Matthias Rambausek and
-        Chris N. Richardson and Matthew W. Scroggs and Nathan Sime and
-        Garth N. Wells},
-          URL = {https://github.com/FEniCS/dolfinx},
-         YEAR = {2021}
-}
-
-@COMMENT{
-  The authors are a list of all people with >=5 commits and >=100 lines
-  of code committed to FFCx since 2018 in alphabetical order.
-}
-@misc{ffcx,
-        TITLE = {FFCx: Next generation FEniCS Form Compiler},
-        AUTHOR = {Igor A. Baratta and Jan Blechta and J{\o}rgen S.
-        Dokken and Michal Habera and Jack S. Hale and Chris N.
-        Richardson and Matthew W. Scroggs and Garth N. Wells},
-          URL = {https://github.com/FEniCS/ffcx},
-         YEAR = {2021}
-}
-
 @article{ufl,
        AUTHOR = {Martin S. Aln{\ae}s and Anders Logg and Kristian B. {\O}lgaard and Marie E. Rognes and Garth N. Wells},
         TITLE = {{U}nified {F}orm {L}anguage: A domain-specific language for weak formulations of partial differential equations},
@@ -199,4 +172,55 @@
          YEAR = {{2021}},
  HOWPUBLISHED = {\url{https://defelement.com}},
          NOTE = {[Online; accessed 26-October-2021]}
+}
+
+@article{xiao-gimbutas,
+       AUTHOR = {Hong Xiao and Zydrunas Gimbutas},
+        TITLE = {A numerical algorithm for the construction of efficient quadrature rules in two and higher dimensions},
+      JOURNAL = {Computers & Mathematics with Applications},
+       VOLUME = {59},
+       NUMBER = {2},
+        PAGES = {{663--676}},
+         year = {2010},
+          doi = {10.1016/j.camwa.2009.10.027},
+}
+
+@COMMENT{
+  The authors are a list of all people with >=5 commits and >=1000 total lines of code committed to DOLFINx since 2018
+  with the user with most total lines first and everyone else in alphabetical order.
+}
+@misc{dolfinx,
+        TITLE = {DOLFINx: Next generation FEniCS problem solving environment},
+        AUTHOR = {Garth N. Wells and Francesco Ballarin and Igor A. Baratta
+        and Joseph P. Dean and J{\o}rgen S. Dokken and Jack S. Hale and
+        Michal Habera and Chris N. Richardson and Matthew W. Scroggs and Nathan Sime},
+          URL = {https://github.com/FEniCS/dolfinx},
+         YEAR = {2021}
+}
+
+@COMMENT{
+  The authors are a list of all people with >=5 commits and >=500 total lines of code committed to FFCx since 2018
+  with the user with most total lines first and everyone else in alphabetical order.
+}
+@misc{ffcx,
+        TITLE = {FFCx: Next generation FEniCS Form Compiler},
+        AUTHOR = {Garth N. Wells and Igor A. Baratta and Michal Habera and Jack S. Hale and Chris N.
+        Richardson and Matthew W. Scroggs},
+          URL = {https://github.com/FEniCS/ffcx},
[…]
