+++ doc/web/make_html.py
@@ -113,6 +113,7 @@ def insert_info(txt):
     if "{{SUPPORTED ELEMENTS}}" in txt:
         with open(path("../../README.md")) as f:
             info = "## Supported elements" + f.read().split("## Supported elements")[1].split("\n## ")[0]
+        info = info.replace("joss/img", "img")
         txt = txt.replace("{{SUPPORTED ELEMENTS}}", info)
     if "{{INSTALL}}" in txt:
         with open(path("../../INSTALL.md")) as f:
@@ -200,7 +201,7 @@ with open(path("cpp/Doxyfile"), "w") as f:
     f.write(content)
 
 # Copy images and assets
-system(f"cp -r {path('../../img')} {path('html')}/img")
+system(f"cp -r {path('../../joss/img')} {path('html')}/img")
 system(f"cp -r {path('assets')} {path('html')}/assets")
 
 # Convert markdown to html
@@ -220,6 +221,13 @@ os.system(f"rm {path('../../demo/python/*.py.rst')}")
 system(f"cd {path('../../demo/python')} && python3 convert_to_rst.py")
 system(f"mkdir {path('python/source/demo')}")
 system(f"cp {path('../../demo/python/*.rst')} {path('python/source/demo')}")
+for file in os.listdir(path("python/source/demo")):
+    if file.endswith(".py.rst"):
+        with open(path(f"python/source/demo/{file}")) as f:
+            content = f.read()
+        content = content.replace("literalinclude:: ../", "literalinclude:: ../../../../../demo/")
+        with open(path(f"python/source/demo/{file}"), "w") as f:
+            f.write(content)
 
 with open(path("python/source/index.rst"), "a") as f:
     f.write("\n")
