### 2021-11-08 #3878 @mscroggs
@jedbrown We've pushed some changes to the paper:

1. We've added a citation to xtensor.
2. We've added a sentence about quadrature to the paper, and referenced the Xiao-Gimbutas quadrature rule as this is what is currently used for triangles.
3. We've added two translations of demos to C++ to the website at https://docs.fenicsproject.org/basix/main/python/demo/demo_create_and_tabulate.py.html and https://docs.fenicsproject.org/basix/main/python/demo/demo_dof_transformations.py.html
4. We've added the element and topology info to the paper.
5. We have longer term plans to work on providing details of tensor product representations, but didn't write this in the paper as this is still in the very early stages of development.

Thanks again for your comments.
