+++ cpp/basix/e-raviart-thomas.cpp
@@ -9,7 +9,6 @@
 #include "moments.h"
 #include "polyset.h"
 #include "quadrature.h"
-#include <numeric>
 #include <vector>
 #include <xtensor/xbuilder.hpp>
 #include <xtensor/xmath.hpp>
