+++ cpp/basix/e-nce-rtc.cpp
@@ -10,10 +10,8 @@
 #include "moments.h"
 #include "polyset.h"
 #include "quadrature.h"
-#include <numeric>
 #include <vector>
 #include <xtensor/xbuilder.hpp>
-#include <xtensor/xpad.hpp>
 #include <xtensor/xtensor.hpp>
 #include <xtensor/xview.hpp>
 
