### 2021-11-10 9cefb75
Use MODULE not SHARED in CMakeLists.txt (#327)
