### 2021-11-15 21a4a6e
Remove redundant function.
