### 2021-11-15 f1292c3
Remove redundant function. (#329)
