### 2021-11-15 835c321
use std::decay_t<O> for && reference types
