### 2021-11-15 96415d7
Added documentation of interpolation_matrix (#330)
