### 2021-11-16 940c96a
remove compute_expansion_coefficients and move functionality inside FiniteElement constructor
