+++ cpp/basix/e-nedelec.cpp
@@ -394,9 +394,7 @@ FiniteElement basix::element::create_nedelec(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, transforms, tdim, tdim);
   }
 
-  const xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, wcoeffs, {M[1], M[2], M[3]}, {x[1], x[2], x[3]}, degree);
-  return FiniteElement(element::family::N1E, celltype, degree, {tdim}, coeffs,
+  return FiniteElement(element::family::N1E, celltype, degree, {tdim}, wcoeffs,
                        transforms, x, M, maps::type::covariantPiola,
                        discontinuous);
 }
@@ -436,9 +434,7 @@ FiniteElement basix::element::create_nedelec2(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, tdim);
   }
 
-  const xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, wcoeffs, {M[1], M[2], M[3]}, {x[1], x[2], x[3]}, degree);
-  return FiniteElement(element::family::N2E, celltype, degree, {tdim}, coeffs,
+  return FiniteElement(element::family::N2E, celltype, degree, {tdim}, wcoeffs,
                        entity_transformations, x, M, maps::type::covariantPiola,
                        discontinuous);
 }
