+++ cpp/basix/e-lagrange.cpp
@@ -139,12 +139,9 @@ FiniteElement create_d_lagrange(cell::type celltype, int degree,
     entity_transformations[cell::type::quadrilateral] = ft;
   }
 
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, xt::eye<double>(ndofs), {M[0], M[1], M[2], M[3]},
-      {x[0], x[1], x[2], x[3]}, degree);
-  return FiniteElement(element::family::P, celltype, degree, {1}, coeffs,
-                       entity_transformations, x, M, maps::type::identity,
-                       true);
+  return FiniteElement(element::family::P, celltype, degree, {1},
+                       xt::eye<double>(ndofs), entity_transformations, x, M,
+                       maps::type::identity, true);
 }
 //-----------------------------------------------------------------------------
 xt::xtensor<double, 2> vtk_triangle_points(int degree)
@@ -772,12 +769,9 @@ FiniteElement create_vtk_element(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, 1);
   }
 
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, xt::eye<double>(ndofs), {M[0], M[1], M[2], M[3]},
-      {x[0], x[1], x[2], x[3]}, degree);
-  return FiniteElement(element::family::P, celltype, degree, {1}, coeffs,
-                       entity_transformations, x, M, maps::type::identity,
-                       discontinuous);
+  return FiniteElement(element::family::P, celltype, degree, {1},
+                       xt::eye<double>(ndofs), entity_transformations, x, M,
+                       maps::type::identity, discontinuous);
 }
 //-----------------------------------------------------------------------------
 } // namespace
@@ -964,13 +958,9 @@ FiniteElement basix::element::create_lagrange(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, 1);
   }
 
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, xt::eye<double>(ndofs), {M[0], M[1], M[2], M[3]},
-      {x[0], x[1], x[2], x[3]}, degree);
-
-  return FiniteElement(element::family::P, celltype, degree, {1}, coeffs,
-                       entity_transformations, x, M, maps::type::identity,
-                       discontinuous);
+  return FiniteElement(element::family::P, celltype, degree, {1},
+                       xt::eye<double>(ndofs), entity_transformations, x, M,
+                       maps::type::identity, discontinuous);
 }
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_dpc(cell::type celltype, int degree,
@@ -1045,9 +1035,7 @@ FiniteElement basix::element::create_dpc(cell::type celltype, int degree,
         = xt::xtensor<double, 3>({2, 0, 0});
   }
 
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, wcoeffs, {M[tdim]}, {x[tdim]}, degree);
-  return FiniteElement(element::family::DPC, celltype, degree, {1}, coeffs,
+  return FiniteElement(element::family::DPC, celltype, degree, {1}, wcoeffs,
                        entity_transformations, x, M, maps::type::identity,
                        discontinuous);
 }
