### 2021-11-16 6aca833
Fix Lagrange at a point
