### 2021-11-16 410f2e8
Correct Lagrange on a point
