+++ cpp/basix/e-lagrange.cpp
@@ -788,12 +788,12 @@ FiniteElement basix::element::create_lagrange(cell::type celltype, int degree,
 
     std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
     std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-    x[0].push_back({{0}});
+    x[0].push_back(xt::zeros<double>({1, 0}));
     M[0].push_back({{{1}}});
     std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
-    xt::xtensor<double, 3> coeffs = {{{1}}};
+    xt::xtensor<double, 2> wcoeffs = {{1}};
 
-    return FiniteElement(element::family::P, cell::type::point, 0, {1}, coeffs,
+    return FiniteElement(element::family::P, cell::type::point, 0, {1}, wcoeffs,
                          entity_transformations, x, M, maps::type::identity,
                          discontinuous);
   }
