### 2021-11-16 77e856c
Add coefficient_matrix function
