### 2021-11-17 d6d3094
Remove `compute_expansion_coefficients` and store dual matrix (#333)
* remove compute_expansion_coefficients and move functionality inside FiniteElement constructor

* Fix Lagrange at a point

* Correct Lagrange on a point

* Add coefficient_matrix function

* docs
