+++ cpp/basix/finite-element.h
@@ -23,162 +23,6 @@ namespace basix
 namespace element
 {
 
-/// Calculates the basis functions of the finite element, in terms of the
-/// polynomial basis.
-///
-/// The below explanation uses Einstein notation.
-///
-/// The basis functions @f${\phi_i}@f$ of a finite element are represented
-/// as a linear combination of polynomials @f$\{p_j\}@f$ in an underlying
-/// polynomial basis that span the space of all d-dimensional polynomials up
-/// to order @f$k \ (P_k^d)@f$:
-/// \f[ \phi_i = c_{ij} p_j \f]
-///
-/// In some cases, the basis functions @f$\{\phi_i\}@f$ do not span the
-/// full space @f$P_k@f$, in which case we denote space spanned by the
-/// basis functions by @f$\{q_k\}@f$, which can be represented by:
-/// @f[  q_i = b_{ij} p_j. @f]
-///  This leads to
-/// @f[  \phi_i = c^{\prime}_{ij} q_j = c^{\prime}_{ij} b_{jk} p_k,  @f]
-/// and in matrix form:
-/// \f[
-/// \phi = C^{\prime} B p
-/// \f]
-///
-/// If the basis functions span the full space, then @f$ B @f$ is simply
-/// the identity.
-///
-/// The basis functions @f$\phi_i@f$ are defined by a dual set of functionals
-/// @f$\{f_i\}@f$. The basis functions are the functions in span{@f$q_k@f$} such
-/// that
-///   @f[ f_i(\phi_j) = \delta_{ij} @f]
-/// and inserting the expression for @f$\phi_{j}@f$:
-///   @f[ f_i(c^{\prime}_{jk}b_{kl}p_{l}) = c^{\prime}_{jk} b_{kl} f_i \left(
-///   p_{l} \right) @f]
-///
-/// Defining a matrix D given by applying the functionals to each
-/// polynomial @f$p_j@f$:
-///  @f[ [D] = d_{ij},\mbox{ where } d_{ij} = f_i(p_j), @f]
-/// we have:
-/// @f[ C^{\prime} B D^{T} = I @f]
-///
-/// and
-///
-/// @f[ C^{\prime} = (B D^{T})^{-1}. @f]
-///
-/// Recalling that @f$C = C^{\prime} B@f$, where @f$C@f$ is the matrix
-/// form of @f$c_{ij}@f$,
-///
-/// @f[ C = (B D^{T})^{-1} B @f]
-///
-/// This function takes the matrices B (span_coeffs) and D (dual) as
-/// inputs and returns the matrix C.
-///
-/// Example: Order 1 Lagrange elements on a triangle
-/// ------------------------------------------------
-/// On a triangle, the scalar expansion basis is:
-///  @f[ p_0 = \sqrt{2}/2 \qquad
-///   p_1 = \sqrt{3}(2x + y - 1) \qquad
-///   p_2 = 3y - 1 @f]
-/// These span the space @f$P_1@f$.
-///
-/// Lagrange order 1 elements span the space P_1, so in this example,
-/// B (span_coeffs) is the identity matrix:
-///   @f[ B = \begin{bmatrix}
-///                   1 & 0 & 0 \\
-///                   0 & 1 & 0 \\
-///                   0 & 0 & 1 \end{bmatrix} @f]
-///
-/// The functionals defining the Lagrange order 1 space are point
-/// evaluations at the three vertices of the triangle. The matrix D
-/// (dual) given by applying these to p_0 to p_2 is:
-///  @f[ \mbox{dual} = \begin{bmatrix}
-///              \sqrt{2}/2 &  -\sqrt{3} & -1 \\
[…]
