+++ cpp/basix/e-brezzi-douglas-marini.cpp
@@ -76,13 +76,8 @@ FiniteElement basix::element::create_bdm(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, tdim);
   }
 
-  // Create coefficients for order (degree-1) vector polynomials
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, xt::eye<double>(ndofs), {M[tdim - 1], M[tdim]},
-      {x[tdim - 1], x[tdim]}, degree);
-
-  return FiniteElement(element::family::BDM, celltype, degree, {tdim}, coeffs,
-                       entity_transformations, x, M,
+  return FiniteElement(element::family::BDM, celltype, degree, {tdim},
+                       xt::eye<double>(ndofs), entity_transformations, x, M,
                        maps::type::contravariantPiola, discontinuous);
 }
 //-----------------------------------------------------------------------------
