+++ python/wrapper.cpp
@@ -307,8 +307,20 @@ Interface to the Basix C++ library.
                                                           py::cast(self));
                              })
       .def_property_readonly(
-          "interpolation_matrix", [](const FiniteElement& self) {
+          "interpolation_matrix",
+          [](const FiniteElement& self) {
             const xt::xtensor<double, 2>& P = self.interpolation_matrix();
+            return py::array_t<double>(P.shape(), P.data(), py::cast(self));
+          })
+      .def_property_readonly(
+          "dual_matrix",
+          [](const FiniteElement& self) {
+            const xt::xtensor<double, 2>& P = self.dual_matrix();
+            return py::array_t<double>(P.shape(), P.data(), py::cast(self));
+          })
+      .def_property_readonly(
+          "coefficient_matrix", [](const FiniteElement& self) {
+            const xt::xtensor<double, 2>& P = self.coefficient_matrix();
             return py::array_t<double>(P.shape(), P.data(), py::cast(self));
           });
 
