+++ cpp/basix/e-crouzeix-raviart.cpp
@@ -70,10 +70,8 @@ FiniteElement basix::element::create_cr(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, 1);
   }
 
-  const xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, xt::eye<double>(ndofs), {M[tdim - 1]}, {x[tdim - 1]}, degree);
-  return FiniteElement(element::family::CR, celltype, 1, {1}, coeffs,
-                       entity_transformations, x, M, maps::type::identity,
-                       discontinuous);
+  return FiniteElement(element::family::CR, celltype, 1, {1},
+                       xt::eye<double>(ndofs), entity_transformations, x, M,
+                       maps::type::identity, discontinuous);
 }
 //-----------------------------------------------------------------------------
