+++ cpp/basix/e-regge.cpp
@@ -190,11 +190,8 @@ FiniteElement basix::element::create_regge(cell::type celltype, int degree,
     entity_transformations[cell::type::triangle] = face_trans;
   }
 
-  const xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, wcoeffs, {M[1], M[2], M[3]}, {x[1], x[2], x[3]}, degree);
-
   return FiniteElement(element::family::Regge, celltype, degree, {tdim, tdim},
-                       coeffs, entity_transformations, x, M,
+                       wcoeffs, entity_transformations, x, M,
                        maps::type::doubleCovariantPiola, discontinuous);
 }
 //-----------------------------------------------------------------------------
