+++ cpp/basix/e-bubble.cpp
@@ -159,9 +159,7 @@ FiniteElement basix::element::create_bubble(cell::type celltype, int degree,
   M[tdim].push_back(xt::xtensor<double, 3>({ndofs, 1, ndofs}));
   xt::view(M[tdim][0], xt::all(), 0, xt::all()) = xt::eye<double>(ndofs);
 
-  xt::xtensor<double, 3> coeffs = element::compute_expansion_coefficients(
-      celltype, wcoeffs, {M[tdim]}, {x[tdim]}, degree);
-  return FiniteElement(element::family::bubble, celltype, degree, {1}, coeffs,
+  return FiniteElement(element::family::bubble, celltype, degree, {1}, wcoeffs,
                        entity_transformations, x, M, maps::type::identity,
                        discontinuous);
 }

[…]
