### 2021-11-17 faf6e08
Remove interpolate() (#334)
* Remove interpolate()

* DOLFINx branch
