+++ cpp/basix/e-nedelec.cpp
@@ -254,108 +254,6 @@ create_nedelec_3d_entity_transforms(int degree)
   return entity_transformations;
 }
 //-----------------------------------------------------------------------------
-std::pair<std::array<std::vector<xt::xtensor<double, 2>>, 4>,
-          std::array<std::vector<xt::xtensor<double, 3>>, 4>>
-create_nedelec2_2d_interpolation(int degree)
-{
-  const int quad_deg = 5 * degree;
-
-  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
-
-  std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
-      element::create_lagrange(cell::type::interval, degree,
-                               element::lagrange_variant::equispaced, true),
-      cell::type::triangle, 2, quad_deg);
-  if (degree > 1)
-  {
-    std::tie(x[2], M[2]) = moments::make_dot_integral_moments(
-        element::create_rt(cell::type::triangle, degree - 1, true),
-        cell::type::triangle, 2, quad_deg);
-  }
-
-  return {x, M};
-}
-//-----------------------------------------------------------------------------
-std::map<cell::type, xt::xtensor<double, 3>>
-create_nedelec2_2d_entity_transformations(int degree)
-{
-  std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
-
-  xt::xtensor<double, 3> edge_transforms
-      = moments::create_tangent_moment_dof_transformations(
-          element::create_lagrange(cell::type::interval, degree,
-                                   element::lagrange_variant::equispaced,
-                                   true));
-  entity_transformations[cell::type::interval] = edge_transforms;
-
-  return entity_transformations;
-}
-//-----------------------------------------------------------------------------
-std::pair<std::array<std::vector<xt::xtensor<double, 2>>, 4>,
-          std::array<std::vector<xt::xtensor<double, 3>>, 4>>
-create_nedelec2_3d_interpolation(int degree)
-{
-  // Create quadrature scheme on the edge
-  const int quad_deg = 5 * degree;
-
-  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
-
-  // Integral representation for the boundary (edge) dofs
-  std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
-      element::create_lagrange(cell::type::interval, degree,
-                               element::lagrange_variant::equispaced, true),
-      cell::type::tetrahedron, 3, quad_deg);
-
-  if (degree > 1)
-  {
-    // Integral moments on faces
-    std::tie(x[2], M[2]) = moments::make_dot_integral_moments(
-        element::create_rt(cell::type::triangle, degree - 1, true),
-        cell::type::tetrahedron, 3, quad_deg);
-  }
-
-  if (degree > 2)
-  {
-    // Interior integral moment
-    std::tie(x[3], M[3]) = moments::make_dot_integral_moments(
-        element::create_rt(cell::type::tetrahedron, degree - 2, true),
-        cell::type::tetrahedron, 3, quad_deg);
-  }
-
-  return {x, M};
-}
-//-----------------------------------------------------------------------------
-std::map<cell::type, xt::xtensor<double, 3>>
-create_nedelec2_3d_entity_transformations(int degree)
-{
-  std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
-
-  const xt::xtensor<double, 3> edge_transforms
-      = moments::create_tangent_moment_dof_transformations(
-          element::create_lagrange(cell::type::interval, degree,
[…]
