+++ cpp/basix/e-nedelec.cpp
@@ -64,46 +64,6 @@ xt::xtensor<double, 2> create_nedelec_2d_space(int degree)
   return wcoeffs;
 }
 //-----------------------------------------------------------------------------
-std::pair<std::array<std::vector<xt::xtensor<double, 2>>, 4>,
-          std::array<std::vector<xt::xtensor<double, 3>>, 4>>
-create_nedelec_2d_interpolation(int degree)
-{
-  const int quad_deg = 5 * degree;
-
-  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
-
-  // Integral representation for the boundary (edge) dofs
-  std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
-      element::create_lagrange(cell::type::interval, degree - 1,
-                               element::lagrange_variant::equispaced, true),
-      cell::type::triangle, 2, quad_deg);
-  if (degree > 1)
-  {
-    std::tie(x[2], M[2]) = moments::make_integral_moments(
-        element::create_lagrange(cell::type::triangle, degree - 2,
-                                 element::lagrange_variant::equispaced, true),
-        cell::type::triangle, 2, quad_deg);
-  }
-
-  return {x, M};
-}
-//-----------------------------------------------------------------------------
-std::map<cell::type, xt::xtensor<double, 3>>
-create_nedelec_2d_entity_transforms(int degree)
-{
-  std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
-
-  xt::xtensor<double, 3> edge_transforms
-      = moments::create_tangent_moment_dof_transformations(
-          element::create_lagrange(cell::type::interval, degree - 1,
-                                   element::lagrange_variant::equispaced,
-                                   true));
-  entity_transformations[cell::type::interval] = edge_transforms;
-
-  return entity_transformations;
-}
-//-----------------------------------------------------------------------------
 xt::xtensor<double, 2> create_nedelec_3d_space(int degree)
 {
   // Reference tetrahedron
@@ -190,70 +150,6 @@ xt::xtensor<double, 2> create_nedelec_3d_space(int degree)
   return wcoeffs;
 }
 //-----------------------------------------------------------------------------
-std::pair<std::array<std::vector<xt::xtensor<double, 2>>, 4>,
-          std::array<std::vector<xt::xtensor<double, 3>>, 4>>
-create_nedelec_3d_interpolation(int degree)
-{
-  // Number of dofs and interpolation points
-  const int quad_deg = 5 * degree;
-
-  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
-
-  std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
-      element::create_lagrange(cell::type::interval, degree - 1,
-                               element::lagrange_variant::equispaced, true),
-      cell::type::tetrahedron, 3, quad_deg);
-
-  if (degree > 1)
-  {
-    std::tie(x[2], M[2]) = moments::make_integral_moments(
-        element::create_lagrange(cell::type::triangle, degree - 2,
-                                 element::lagrange_variant::equispaced, true),
-        cell::type::tetrahedron, 3, quad_deg);
-  }
-
-  if (degree > 2)
-  {
-    std::tie(x[3], M[3]) = moments::make_integral_moments(
-        element::create_lagrange(cell::type::tetrahedron, degree - 3,
-                                 element::lagrange_variant::equispaced, true),
-        cell::type::tetrahedron, 3, quad_deg);
-  }
-
-  return {x, M};
-}
-//-----------------------------------------------------------------------------
[…]
