### 2021-11-18 0603ab9
Use fixed-width xtensor arrays now that we don't support the legacy Intel compiler
