### 2021-11-18 539587e
Use fixed-width xtensor arrays now that we don't support the legacy I… (#337)
* Use fixed-width xtensor arrays now that we don't support the legacy Intel compiler

* Remove commented lines
