+++ cpp/basix/quadrature.cpp
@@ -4,6 +4,7 @@
 
 #include "quadrature.h"
 #include "math.h"
+#include <algorithm>
 #include <array>
 #include <cmath>
 #include <vector>
@@ -136,19 +137,20 @@ std::array<std::vector<double>, 2> lobatto(const std::vector<double>& alpha,
   return gauss(alpha_l, beta_l);
 }
 //-----------------------------------------------------------------------------
+
+/// Evaluate the nth Jacobi polynomial and derivatives with weight
+/// parameters (a, 0) at points x
+/// @param[in] a Jacobi weight a
+/// @param[in] n Order of polynomial
+/// @param[in] nderiv Number of derivatives (if zero, just compute
+/// polynomial itself)
+/// @param[in] x Points at which to evaluate
+/// @return Array of polynomial derivative values (rows) at points
+/// (columns)
 xt::xtensor<double, 2> compute_jacobi_deriv(double a, std::size_t n,
                                             std::size_t nderiv,
                                             const xtl::span<const double>& x)
 {
-  // Evaluate the nth Jacobi polynomial and derivatives with weight
-  // parameters (a, 0) at points x
-  // @param[in] a Jacobi weight a
-  // @param[in] n Order of polynomial
-  // @param[in] nderiv Number of derivatives (if zero, just compute
-  // polynomial itself)
-  // @param[in] x Points at which to evaluate
-  // @return Array of polynomial derivative values (rows) at points
-  // (columns)
 
   std::vector<std::size_t> shape = {x.size()};
   const auto _x = xt::adapt(x.data(), x.size(), xt::no_ownership(), shape);
@@ -495,32 +497,34 @@ make_strang_fix_quadrature(cell::type celltype, std::size_t m)
     if (m == 2)
     {
       // Scheme from Strang and Fix, 3 points, degree of precision 2
-      xt::xtensor<double, 2> x = {{1.0 / 6.0, 1.0 / 6.0},
-                                  {1.0 / 6.0, 2.0 / 3.0},
-                                  {2.0 / 3.0, 1.0 / 6.0}};
+      xt::xtensor_fixed<double, xt::xshape<3, 2>> x = {{1.0 / 6.0, 1.0 / 6.0},
+                                                       {1.0 / 6.0, 2.0 / 3.0},
+                                                       {2.0 / 3.0, 1.0 / 6.0}};
       return {x, {1.0 / 6.0, 1.0 / 6.0, 1.0 / 6.0}};
     }
     else if (m == 3)
     {
       // Scheme from Strang and Fix, 6 points, degree of precision 3
-      xt::xtensor<double, 2> x = {{0.659027622374092, 0.231933368553031},
-                                  {0.659027622374092, 0.109039009072877},
-                                  {0.231933368553031, 0.659027622374092},
-                                  {0.231933368553031, 0.109039009072877},
-                                  {0.109039009072877, 0.659027622374092},
-                                  {0.109039009072877, 0.231933368553031}};
+      xt::xtensor_fixed<double, xt::xshape<6, 2>> x
+          = {{0.659027622374092, 0.231933368553031},
+             {0.659027622374092, 0.109039009072877},
+             {0.231933368553031, 0.659027622374092},
+             {0.231933368553031, 0.109039009072877},
+             {0.109039009072877, 0.659027622374092},
+             {0.109039009072877, 0.231933368553031}};
       std::vector<double> w(6, 1.0 / 12.0);
       return {x, w};
     }
     else if (m == 4)
     {
       // Scheme from Strang and Fix, 6 points, degree of precision 4
-      xt::xtensor<double, 2> x = {{0.816847572980459, 0.091576213509771},
[…]
