+++ cpp/basix/e-serendipity.cpp
@@ -557,9 +557,6 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
       = cell::topology(celltype);
   const std::size_t tdim = cell::topological_dimension(celltype);
 
-  // Number of dofs and interpolation points
-  int quad_deg = 5 * degree;
-
   std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
   std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
 
@@ -580,8 +577,8 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
   {
     FiniteElement moment_space
         = element::create_dpc(cell::type::interval, degree - 2, true);
-    std::tie(x[1], M[1])
-        = moments::make_integral_moments(moment_space, celltype, 1, quad_deg);
+    std::tie(x[1], M[1]) = moments::make_integral_moments(
+        moment_space, celltype, 1, 2 * degree - 2);
     if (tdim > 1)
     {
       edge_transforms
@@ -593,8 +590,8 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
   {
     FiniteElement moment_space
         = element::create_dpc(cell::type::quadrilateral, degree - 4, true);
-    std::tie(x[2], M[2])
-        = moments::make_integral_moments(moment_space, celltype, 1, quad_deg);
+    std::tie(x[2], M[2]) = moments::make_integral_moments(
+        moment_space, celltype, 1, 2 * degree - 4);
     if (tdim > 2)
     {
       face_transforms
@@ -606,7 +603,7 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
   {
     std::tie(x[3], M[3]) = moments::make_integral_moments(
         element::create_dpc(cell::type::hexahedron, degree - 6, true), celltype,
-        1, quad_deg);
+        1, 2 * degree - 6);
   }
 
   xt::xtensor<double, 2> wcoeffs;
@@ -668,9 +665,6 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
   const cell::type facettype
       = (tdim == 2) ? cell::type::interval : cell::type::quadrilateral;
 
-  // Number of dofs and interpolation points
-  int quad_deg = 5 * degree;
-
   std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
   std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
 
@@ -679,7 +673,7 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
   FiniteElement facet_moment_space
       = element::create_dpc(facettype, degree, true);
   std::tie(x[tdim - 1], M[tdim - 1]) = moments::make_normal_integral_moments(
-      facet_moment_space, celltype, tdim, quad_deg);
+      facet_moment_space, celltype, tdim, 2 * degree);
   if (tdim > 1)
   {
     facet_transforms
@@ -691,7 +685,7 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
     FiniteElement cell_moment_space
         = element::create_dpc(celltype, degree - 2, true);
     std::tie(x[tdim], M[tdim]) = moments::make_integral_moments(
-        cell_moment_space, celltype, tdim, quad_deg);
+        cell_moment_space, celltype, tdim, 2 * degree - 2);
   }
 
   xt::xtensor<double, 2> wcoeffs;
@@ -760,7 +754,7 @@ FiniteElement basix::element::create_serendipity_curl(cell::type celltype,
       = element::create_dpc(cell::type::interval, degree, true);
 
   std::tie(x[1], M[1]) = moments::make_tangent_integral_moments(
[…]
