+++ python/basix/finite_element.py
@@ -65,6 +65,8 @@ def string_to_family(family: str, cell: str) -> _EF:
             "NCE": _EF.N1E,
             "BDMCF": _EF.BDM,
             "BDMCE": _EF.N2E,
+            "AAF": _EF.BDM,
+            "AAE": _EF.N2E,
         })
     # Family names that are valid for triangles and tetrahedra
     if cell in ["triangle", "tetrahedron"]:
