### 2021-11-19 9b4f1af
Reduce quadrature orders used in element creation (#338)
* reduce quadrature orders when defining nedelec

* quadrature orders in BDM and RT

* quadrature orders for serendipity

* Quadrature order for NCE/RTC

* Add AAE and AAF alternative names

* run more tests

* revert to DOLFINx main

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>
