### 2021-11-19 c1ceda1
remove transform from quadrature rules by including transformed numbers directly
