+++ cpp/basix/quadrature.cpp
@@ -526,10 +526,8 @@ make_strang_fix_quadrature(cell::type celltype, std::size_t m)
              {0.445948490915965, 0.108103018168070},
              {0.445948490915965, 0.445948490915965}};
       std::vector<double> w
-          = {0.109951743655322, 0.109951743655322, 0.109951743655322,
-             0.223381589678011, 0.223381589678011, 0.223381589678011};
-      std::transform(w.cbegin(), w.cend(), w.begin(),
-                     [](auto x) { return 0.5 * x; });
+          = {0.054975871827661,  0.054975871827661,  0.054975871827661,
+             0.1116907948390055, 0.1116907948390055, 0.1116907948390055};
       return {x, w};
     }
     else if (m == 5)
@@ -543,12 +541,13 @@ make_strang_fix_quadrature(cell::type celltype, std::size_t m)
              {0.05971587178976981, 0.47014206410511505},
              {0.47014206410511505, 0.05971587178976981},
              {0.47014206410511505, 0.47014206410511505}};
-      std::vector<double> w
-          = {0.22500000000000000, 0.12593918054482717, 0.12593918054482717,
-             0.12593918054482717, 0.13239415278850616, 0.13239415278850616,
-             0.13239415278850616};
-      std::transform(w.cbegin(), w.cend(), w.begin(),
-                     [](auto x) { return 0.5 * x; });
+      std::vector<double> w = {0.1125,
+                               0.06296959027241358,
+                               0.06296959027241358,
+                               0.06296959027241358,
+                               0.06619707639425308,
+                               0.06619707639425308,
+                               0.06619707639425308};
       return {x, w};
     }
     else if (m == 6)
@@ -568,12 +567,10 @@ make_strang_fix_quadrature(cell::type celltype, std::size_t m)
              {0.053145049844816, 0.636502499121399},
              {0.053145049844816, 0.310352451033785}};
       std::vector<double> w
-          = {0.050844906370207, 0.050844906370207, 0.050844906370207,
-             0.116786275726379, 0.116786275726379, 0.116786275726379,
-             0.082851075618374, 0.082851075618374, 0.082851075618374,
-             0.082851075618374, 0.082851075618374, 0.082851075618374};
-      std::transform(w.cbegin(), w.cend(), w.begin(),
-                     [](auto x) { return 0.5 * x; });
+          = {0.0254224531851035, 0.0254224531851035, 0.0254224531851035,
+             0.0583931378631895, 0.0583931378631895, 0.0583931378631895,
+             0.041425537809187,  0.041425537809187,  0.041425537809187,
+             0.041425537809187,  0.041425537809187,  0.041425537809187};
       return {x, w};
     }
     else
@@ -660,13 +657,11 @@ make_keast_quadrature(cell::type celltype, std::size_t m)
              {0.3143728734931922, 0.3143728734931922, 0.0568813795204234},
              {0.3143728734931922, 0.0568813795204234, 0.3143728734931922}};
       std::vector<double> w
-          = {0.0190476190476190, 0.0190476190476190, 0.0190476190476190,
-             0.0190476190476190, 0.0190476190476190, 0.0190476190476190,
-             0.0885898247429807, 0.0885898247429807, 0.0885898247429807,
-             0.0885898247429807, 0.1328387466855907, 0.1328387466855907,
-             0.1328387466855907, 0.1328387466855907};
-      std::transform(w.cbegin(), w.cend(), w.begin(),
-                     [](auto x) { return x / 6.0; });
+          = {0.003174603174603167, 0.003174603174603167, 0.003174603174603167,
[…]
