### 2021-11-19 06f31ec
Map interface refactoring.
