### 2021-11-19 58797ce
Remove map functions
