+++ cpp/basix/finite-element.cpp
@@ -746,8 +746,30 @@ xt::xtensor<double, 3> FiniteElement::map_pull_back(
 {
   const std::size_t reference_value_size = value_size();
   xt::xtensor<double, 3> U({u.shape(0), u.shape(1), reference_value_size});
-  map_pull_back_m(u, J, detJ, K, U);
-  return U;
+  using u_t = xt::xview<decltype(u)&, std::size_t, xt::xall<std::size_t>,
+                        xt::xall<std::size_t>>;
+  using U_t = xt::xview<decltype(U)&, std::size_t, xt::xall<std::size_t>,
+                        xt::xall<std::size_t>>;
+  using J_t = xt::xview<decltype(J)&, std::size_t, xt::xall<std::size_t>,
+                        xt::xall<std::size_t>>;
+  using K_t = xt::xview<decltype(K)&, std::size_t, xt::xall<std::size_t>,
+                        xt::xall<std::size_t>>;
+  auto map = this->push_forward_fn<U_t, u_t, K_t, J_t>();
+  for (std::size_t i = 0; i < u.shape(0); ++i)
+  {
+    auto _K = xt::view(K, i, xt::all(), xt::all());
+    auto _J = xt::view(J, i, xt::all(), xt::all());
+    auto _u = xt::view(u, i, xt::all(), xt::all());
+    auto _U = xt::view(U, i, xt::all(), xt::all());
+    map(_U, _u, _K, 1.0 / detJ[i], _J);
+  }
+
+  return u;
+
+  // const std::size_t reference_value_size = value_size();
+  // xt::xtensor<double, 3> U({u.shape(0), u.shape(1), reference_value_size});
+  // map_pull_back_m(u, J, detJ, K, U);
+  // return U;
 }
 //-----------------------------------------------------------------------------
 void FiniteElement::permute_dofs(const xtl::span<std::int32_t>& dofs,
