+++ cpp/basix/maps.h
@@ -55,13 +55,7 @@ void dot21(Vec&& r, const Mat0& A, const Mat1& B)
     r[i] = acc;
   }
 }
-
-template <typename Vec0, typename Vec1, typename Mat0, typename Mat1>
-void identity(Vec0&& r, const Vec1& U, const Mat0& /*J*/, double /*detJ*/,
-              const Mat1& /*K*/)
-{
-  r.assign(U);
-}
+} // namespace impl
 
 template <typename O, typename P, typename Q, typename R>
 void covariant_piola(O&& r, const P& U, const Q& /*J*/, double /*detJ*/,
@@ -72,7 +66,7 @@ void covariant_piola(O&& r, const P& U, const Q& /*J*/, double /*detJ*/,
   {
     auto r_p = xt::row(r, p);
     auto U_p = xt::row(U, p);
-    dot21(r_p, Kt, U_p);
+    impl::dot21(r_p, Kt, U_p);
   }
 }
 
@@ -84,7 +78,7 @@ void contravariant_piola(O&& r, const P& U, const Q& J, double detJ,
   {
     auto r_p = xt::row(r, p);
     auto U_p = xt::row(U, p);
-    dot21(r_p, J, U_p);
+    impl::dot21(r_p, J, U_p);
   }
   double inv_detJ = 1 / detJ;
   std::for_each(r.begin(), r.end(), [inv_detJ](auto& ri) { ri *= inv_detJ; });
@@ -100,7 +94,7 @@ void double_covariant_piola(O&& r, const P& U, const Q& J, double /*detJ*/,
     auto U_p = xt::row(U, p);
     auto _U = xt::reshape_view(U_p, {J.shape(1), J.shape(1)});
     auto _r = xt::reshape_view(r_p, {K.shape(1), K.shape(1)});
-    dot22(_r, xt::transpose(K), _U, K);
+    impl::dot22(_r, xt::transpose(K), _U, K);
   }
 }
 
@@ -115,46 +109,10 @@ void double_contravariant_piola(O&& r, const P& U, const Q& J, double detJ,
     auto U_p = xt::row(U, p);
     auto _U = xt::reshape_view(U_p, {J.shape(1), J.shape(1)});
     auto _r = xt::reshape_view(r_p, {J.shape(0), J.shape(0)});
-    dot22(_r, J, _U, Jt);
+    impl::dot22(_r, J, _U, Jt);
   }
   double inv_detJ = 1 / (detJ * detJ);
   std::for_each(r.begin(), r.end(), [inv_detJ](auto& ri) { ri *= inv_detJ; });
-}
-} // namespace impl
-
-/// Apply a map to data. Note that the required input arguments depends
-/// on the type of map.
-///
-/// @note This function is designed to be called at runtime, so its performance
-/// is critical.
-/// @todo Remove the switch from this function if possible
-///
-/// @param[out] u The field after mapping, flattened with row-major
-/// layout
-/// @param[in] U The field to be mapped, flattened with row-major layout
-/// @param[in] J Jacobian of the map
-/// @param[in] detJ Determinant of `J`
-/// @param[in] K The inverse of `J`
-/// @param[in] map_type The map type
-template <typename O, typename P, typename Mat0, typename Mat1>
-void apply_map(O&& u, const P& U, const Mat0& J, double detJ, const Mat1& K,
-               maps::type map_type)
-{
-  switch (map_type)
-  {
-  case maps::type::identity:
-    return impl::identity(u, U, J, detJ, K);
-  case maps::type::covariantPiola:
-    return impl::covariant_piola(u, U, J, detJ, K);
-  case maps::type::contravariantPiola:
-    return impl::contravariant_piola(u, U, J, detJ, K);
[…]
