### 2021-11-19 689782a
Simplifications
