### 2021-11-19 7f784c2
Simplify maps
