+++ python/wrapper.cpp
@@ -64,7 +64,8 @@ Interface to the Basix C++ library.
   m.def("topology", &cell::topology, basix::docstring::topology.c_str());
   m.def(
       "geometry",
-      [](cell::type celltype) {
+      [](cell::type celltype)
+      {
         xt::xtensor<double, 2> g = cell::geometry(celltype);
         auto strides = g.strides();
         for (auto& s : strides)
@@ -76,7 +77,8 @@ Interface to the Basix C++ library.
         basix::docstring::sub_entity_connectivity.c_str());
   m.def(
       "sub_entity_geometry",
-      [](cell::type celltype, int dim, int index) {
+      [](cell::type celltype, int dim, int index)
+      {
         xt::xtensor<double, 2> g
             = cell::sub_entity_geometry(celltype, dim, index);
         auto strides = g.strides();
@@ -100,7 +102,8 @@ Interface to the Basix C++ library.
 
   m.def(
       "create_lattice",
-      [](cell::type celltype, int n, lattice::type type, bool exterior) {
+      [](cell::type celltype, int n, lattice::type type, bool exterior)
+      {
         auto l = lattice::create(celltype, n, type, exterior,
                                  lattice::simplex_method::none);
         auto strides = l.strides();
@@ -113,7 +116,8 @@ Interface to the Basix C++ library.
   m.def(
       "create_lattice",
       [](cell::type celltype, int n, lattice::type type, bool exterior,
-         lattice::simplex_method method) {
+         lattice::simplex_method method)
+      {
         auto l = lattice::create(celltype, n, type, exterior, method);
         auto strides = l.strides();
         for (auto& s : strides)
@@ -152,14 +156,16 @@ Interface to the Basix C++ library.
       basix::docstring::cell_volume.c_str());
   m.def(
       "cell_facet_normals",
-      [](cell::type cell_type) {
+      [](cell::type cell_type)
+      {
         xt::xtensor<double, 2> normals = cell::facet_normals(cell_type);
         return py::array_t<double>(normals.shape(), normals.data());
       },
       basix::docstring::cell_facet_normals.c_str());
   m.def(
       "cell_facet_reference_volumes",
-      [](cell::type cell_type) {
+      [](cell::type cell_type)
+      {
         xt::xtensor<double, 1> volumes
             = cell::facet_reference_volumes(cell_type);
         return py::array_t<double>(volumes.shape(), volumes.data());
@@ -167,7 +173,8 @@ Interface to the Basix C++ library.
       basix::docstring::cell_facet_reference_volumes.c_str());
   m.def(
       "cell_facet_outward_normals",
-      [](cell::type cell_type) {
+      [](cell::type cell_type)
+      {
         xt::xtensor<double, 2> normals = cell::facet_outward_normals(cell_type);
         return py::array_t<double>(normals.shape(), normals.data());
       },
@@ -176,7 +183,8 @@ Interface to the Basix C++ library.
         basix::docstring::cell_facet_orientations.c_str());
   m.def(
       "cell_facet_jacobians",
-      [](cell::type cell_type) {
+      [](cell::type cell_type)
+      {
         xt::xtensor<double, 3> jacobians = cell::facet_jacobians(cell_type);
         return py::array_t<double>(jacobians.shape(), jacobians.data());
       },
@@ -198,42 +206,46 @@ Interface to the Basix C++ library.
       .def(
           "tabulate",
           [](const FiniteElement& self, int n,
-             const py::array_t<double, py::array::c_style>& x) {
+             const py::array_t<double, py::array::c_style>& x)
+          {
             auto _x = adapt_x(x);
             auto t = self.tabulate(n, _x);
             return py::array_t<double>(t.shape(), t.data());
           },
[…]
