### 2021-11-19 2bf27f4
Further modernise python/pybind11 steps in cmake.
