+++ python/setup.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import sys
+import sysconfig
 
 from setuptools import Extension, setup
 from setuptools.command.build_ext import build_ext
