+++ cpp/basix/quadrature.cpp
@@ -871,30 +871,29 @@ make_xiao_gimbutas_quadrature(cell::type celltype, int m)
   {
     if (m == 1)
     {
-      // Xiao Gimbutas, 3 points, degree 1
       xt::xtensor<double, 2> x = {{0.3333333333333333, 0.3333333333333333}};
       std::vector<double> w = {0.5};
       return {x, w};
     }
     else if (m == 2)
     {
-      // Xiao Gimbutas, 3 points, degree 2
-      xt::xtensor<double, 2> x = {{0.16666666666666666, 0.16666666666666666},
-                                  {0.16666666666666666, 0.6666666666666667},
-                                  {0.6666666666666667, 0.16666666666666666}};
+      xt::xtensor_fixed<double, xt::xshape<3, 2>> x
+          = {{0.16666666666666666, 0.16666666666666666},
+             {0.16666666666666666, 0.6666666666666667},
+             {0.6666666666666667, 0.16666666666666666}};
       std::vector<double> w
           = {0.16666666666666666, 0.16666666666666666, 0.16666666666666666};
       return {x, w};
     }
     else if (m == 3)
     {
-      // Xiao Gimbutas, 3 points, degree 3
-      xt::xtensor<double, 2> x = {{0.4459484909159649, 0.4459484909159649},
-                                  {0.09157621350977085, 0.09157621350977085},
-                                  {0.4459484909159649, 0.10810301816807022},
-                                  {0.09157621350977085, 0.8168475729804583},
-                                  {0.10810301816807022, 0.4459484909159649},
-                                  {0.8168475729804583, 0.09157621350977085}};
+      xt::xtensor_fixed<double, xt::xshape<6, 2>> x
+          = {{0.4459484909159649, 0.4459484909159649},
+             {0.09157621350977085, 0.09157621350977085},
+             {0.4459484909159649, 0.10810301816807022},
+             {0.09157621350977085, 0.8168475729804583},
+             {0.10810301816807022, 0.4459484909159649},
+             {0.8168475729804583, 0.09157621350977085}};
       std::vector<double> w
           = {0.11169079483900574, 0.05497587182766094, 0.11169079483900574,
              0.05497587182766094, 0.11169079483900574, 0.05497587182766094};
@@ -902,13 +901,13 @@ make_xiao_gimbutas_quadrature(cell::type celltype, int m)
     }
     else if (m == 4)
     {
-      // Xiao Gimbutas, 3 points, degree 4
-      xt::xtensor<double, 2> x = {{0.4459484909159649, 0.4459484909159649},
-                                  {0.09157621350977085, 0.09157621350977085},
-                                  {0.4459484909159649, 0.10810301816807022},
-                                  {0.09157621350977085, 0.8168475729804583},
-                                  {0.10810301816807022, 0.4459484909159649},
-                                  {0.8168475729804583, 0.09157621350977085}};
+      xt::xtensor_fixed<double, xt::xshape<6, 2>> x
+          = {{0.4459484909159649, 0.4459484909159649},
+             {0.09157621350977085, 0.09157621350977085},
+             {0.4459484909159649, 0.10810301816807022},
+             {0.09157621350977085, 0.8168475729804583},
+             {0.10810301816807022, 0.4459484909159649},
+             {0.8168475729804583, 0.09157621350977085}};
       std::vector<double> w
           = {0.11169079483900574, 0.05497587182766094, 0.11169079483900574,
              0.05497587182766094, 0.11169079483900574, 0.05497587182766094};
@@ -916,14 +915,14 @@ make_xiao_gimbutas_quadrature(cell::type celltype, int m)
     }
     else if (m == 5)
     {
-      // Xiao Gimbutas, 3 points, degree 5
[…]
