### 2021-11-19 ba0eb9a
Make skbuild pass up the same cmake args as setup.py
