+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=42", "wheel", "pybind11>=2.6.4", "cmake>=3.16", "scikit-build>=0.12"]
+requires = ["setuptools>=42", "wheel", "pybind11>=2.6.4", "cmake>=3.18", "scikit-build>=0.12"]
 
 build-backend = "setuptools.build_meta"
 
