### 2021-11-19 2acfc9c
Return map function for an element (#342)
* Add a function that return an appropriate 'map' function

* test on dolfinx

* some optimizations to map

* revert change

* use std::decay_t<O> for && reference types

* Formatting fix

* Reference updates

* Add missing include

* Map interface refactoring.

* Refactor map.

* Remove map functions

* Map refactor

* Simplifications

* Simplify maps

* Updates

* Revert a change

* Update DOLFINx branch

* Fix return value

* Small updates

* Undo changes to a file

* Revert changes

Co-authored-by: IgorBaratta <igorbaratta@gmail.com>
