### 2021-11-19 1ff6c00
Remove public map_type
