+++ cpp/basix/finite-element.h
@@ -408,7 +408,7 @@ public:
   template <typename O, typename P, typename Q, typename R>
   std::function<void(O&, const P&, const Q&, double, const R&)> map_fn() const
   {
-    switch (map_type)
+    switch (_map_type)
     {
     case maps::type::identity:
       return [](O& u, const P& U, const Q&, double, const R&) { u.assign(U); };
@@ -739,9 +739,6 @@ public:
   /// `FiniteElement()` constructor.
   /// @return The dual matrix
   xt::xtensor<double, 2> coefficient_matrix() const;
-
-  /// Element map type
-  maps::type map_type;
 
 private:
   // Cell type
