### 2021-11-20 2acf805
Remove public map_type (#344)
* Remove public map_type

* Test branch update

* Point to test branch

* Update test branch [skip ci]

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>
