### 2021-11-22 4e3a141
Move back to CMake 3.16
