+++ setup.py
@@ -14,7 +14,7 @@ setup(name="fenics-basix",
       license="MIT",
       packages=["basix"],
       cmake_args=[
-          '-DPython3_EXECUTABLE=' + sys.executable,
+          f'-DPython3_EXECUTABLE={sys.executable}',
           f'-DPython3_LIBRARIES={sysconfig.get_config_var("LIBDEST")}',
           f'-DPython3_INCLUDE_DIRS={sysconfig.get_config_var("INCLUDEPY")}'],
       package_dir={"": "python"},
