+++ cpp/CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.18)
+cmake_minimum_required(VERSION 3.16)
 
 # Set the version
 project(Basix VERSION "0.3.1.0" LANGUAGES CXX)
