+++ python/setup.py
@@ -49,8 +49,8 @@ class CMakeBuild(build_ext):
         extdir = os.path.abspath(os.path.dirname(self.get_ext_fullpath(ext.name)))
         cmake_args = ['-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=' + extdir,
                       '-DPython3_EXECUTABLE=' + sys.executable,
-                      f'-DPython3_LIBRARIES={sysconfig.get_config_var("LIBDEST")}',
-                      f'-DPython3_INCLUDE_DIRS={sysconfig.get_config_var("INCLUDEPY")}']
+                      '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST"),
+                      '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")}]
 
         cfg = 'Debug' if self.debug else 'Release'
         build_args = ['--config', cfg]
