+++ setup.py
@@ -14,8 +14,8 @@ setup(name="fenics-basix",
       license="MIT",
       packages=["basix"],
       cmake_args=[
-          f'-DPython3_EXECUTABLE={sys.executable}',
-          f'-DPython3_LIBRARIES={sysconfig.get_config_var("LIBDEST")}',
-          f'-DPython3_INCLUDE_DIRS={sysconfig.get_config_var("INCLUDEPY")}'],
+          '-DPython3_EXECUTABLE=' + sys.executable,
+          '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST")},
+          '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")}],
       package_dir={"": "python"},
       cmake_install_dir="python/basix/")
