+++ setup.py
@@ -15,7 +15,7 @@ setup(name="fenics-basix",
       packages=["basix"],
       cmake_args=[
           '-DPython3_EXECUTABLE=' + sys.executable,
-          '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST")},
-          '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")}],
+          '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST"),
+          '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")],
       package_dir={"": "python"},
       cmake_install_dir="python/basix/")
