### 2021-11-23 527195e
Use FindPython3 in CMake (#343)
if statement around Development.Module can be removed once CMake >= 3.18 is widely available
