+++ python/setup.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import sys
+import sysconfig
 
 from setuptools import Extension, setup
 from setuptools.command.build_ext import build_ext
@@ -47,7 +48,9 @@ class CMakeBuild(build_ext):
     def build_extension(self, ext):
         extdir = os.path.abspath(os.path.dirname(self.get_ext_fullpath(ext.name)))
         cmake_args = ['-DCMAKE_LIBRARY_OUTPUT_DIRECTORY=' + extdir,
-                      '-DPYTHON_EXECUTABLE=' + sys.executable]
+                      '-DPython3_EXECUTABLE=' + sys.executable,
+                      '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST"),
+                      '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")]
 
         cfg = 'Debug' if self.debug else 'Release'
         build_args = ['--config', cfg]
