### 2021-11-26 e833c87
Added highest degree function to finite element
