+++ cpp/basix/e-raviart-thomas.cpp
@@ -117,8 +117,8 @@ FiniteElement basix::element::create_rt(cell::type celltype, int degree,
         = element::make_discontinuous(x, M, entity_transformations, tdim, tdim);
   }
 
-  return FiniteElement(element::family::RT, celltype, degree, {tdim}, B,
-                       entity_transformations, x, M,
-                       maps::type::contravariantPiola, discontinuous);
+  return FiniteElement(
+      element::family::RT, celltype, degree, {tdim}, B, entity_transformations,
+      x, M, maps::type::contravariantPiola, discontinuous, degree, degree);
 }
 //-----------------------------------------------------------------------------
