### 2021-11-26 f69e3da
Added highest space that is subset of element
