+++ cpp/basix/e-serendipity.cpp
@@ -646,7 +646,8 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
 
   return FiniteElement(element::family::serendipity, celltype, degree, {1},
                        wcoeffs, entity_transformations, x, M,
-                       maps::type::identity, discontinuous, degree, degree);
+                       maps::type::identity, discontinuous, degree,
+                       degree < static_cast<int>(tdim) ? 1 : degree / tdim);
 }
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_serendipity_div(cell::type celltype,
@@ -718,7 +719,7 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
   return FiniteElement(element::family::BDM, celltype, degree + 1, {tdim},
                        wcoeffs, entity_transformations, x, M,
                        maps::type::contravariantPiola, discontinuous,
-                       degree + 1, degree);
+                       degree + 1, degree / tdim);
 }
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_serendipity_curl(cell::type celltype,
@@ -811,6 +812,6 @@ FiniteElement basix::element::create_serendipity_curl(cell::type celltype,
   return FiniteElement(element::family::N2E, celltype, degree + 1, {tdim},
                        wcoeffs, entity_transformations, x, M,
                        maps::type::covariantPiola, discontinuous, degree + 1,
-                       degree);
+                       (degree == 2 && tdim == 3) ? 1 : degree / tdim);
 }
 //-----------------------------------------------------------------------------
