+++ test/test_interpolation_between_elements.py
@@ -140,16 +140,16 @@ def test_interpolation_orders(cell_type, degree, element_type, element_args):
     element = basix.create_element(element_type, cell_type, degree, *element_args)
 
     points = basix.create_lattice(cell_type, 10, basix.LatticeType.equispaced, True)
+    tab = element.tabulate(0, points)[0]
 
     # Test that this element's basis functions are contained in Lagrange space with
     # degree element.highest_polynomial_degree
     coeffs = np.random.rand(element.dim)
-    tab = element.tabulate(0, points)[0]
     values = np.array([tab[:, :, i] @ coeffs
                        for i in range(element.value_size)])
 
     if element.highest_polynomial_degree >= 0:
-        # The element being test should be a subset of this Lagrange space
+        # The element being tested should be a subset of this Lagrange space
         lagrange = basix.create_element(
             basix.ElementFamily.P, cell_type, element.highest_polynomial_degree, basix.LagrangeVariant.equispaced, True)
         lagrange_coeffs = basix.compute_interpolation_operator(element, lagrange) @ coeffs
@@ -160,9 +160,10 @@ def test_interpolation_orders(cell_type, degree, element_type, element_args):
         assert np.allclose(values, lagrange_values)
 
     if element.highest_polynomial_degree >= 1:
-        # The element being test should be NOT a subset of this Lagrange space
+        # The element being tested should be NOT a subset of this Lagrange space
         lagrange = basix.create_element(
-            basix.ElementFamily.P, cell_type, element.highest_polynomial_degree - 1, basix.LagrangeVariant.equispaced, True)
+            basix.ElementFamily.P, cell_type, element.highest_polynomial_degree - 1,
+            basix.LagrangeVariant.equispaced, True)
         lagrange_coeffs = basix.compute_interpolation_operator(element, lagrange) @ coeffs
         lagrange_tab = lagrange.tabulate(0, points)[0]
         lagrange_values = np.array([lagrange_tab[:, :, 0] @ lagrange_coeffs[i::element.value_size]
@@ -170,34 +171,38 @@ def test_interpolation_orders(cell_type, degree, element_type, element_args):
 
         assert not np.allclose(values, lagrange_values)
 
-"""
-    # Test that the basis functions of lagrange space with degree element.highest_complete_polynomial_degree
+    # Test that the basis functions of Lagrange space with degree element.highest_complete_polynomial_degree
     # are contained in this space
+
     if element.highest_complete_polynomial_degree >= 0:
+        # This Lagrange space should be a subset to the element being tested
         lagrange = basix.create_element(
-            basix.ElementFamily.P, cell_type, element.highest_polynomial_degree, basix.LagrangeVariant.equispaced, True)
-        lagrange_coeffs = basix.compute_interpolation_operator(element, lagrange) @ coeffs
+            basix.ElementFamily.P, cell_type, element.highest_complete_polynomial_degree,
+            basix.LagrangeVariant.equispaced, True)
+        lagrange_coeffs = np.random.rand(lagrange.dim * element.value_size)
         lagrange_tab = lagrange.tabulate(0, points)[0]
         lagrange_values = np.array([lagrange_tab[:, :, 0] @ lagrange_coeffs[i::element.value_size]
                                     for i in range(element.value_size)])
 
+        coeffs = basix.compute_interpolation_operator(lagrange, element) @ lagrange_coeffs
+        values = np.array([tab[:, :, i] @ coeffs
+                           for i in range(element.value_size)])
+
         assert np.allclose(values, lagrange_values)
-    lagrange = basix.create_element(
[…]
