+++ cpp/basix/e-nce-rtc.cpp
@@ -154,7 +154,7 @@ FiniteElement basix::element::create_rtc(cell::type celltype, int degree,
   return FiniteElement(element::family::RT, celltype, degree, {tdim}, wcoeffs,
                        entity_transformations, x, M,
                        maps::type::contravariantPiola, discontinuous, degree,
-                       degree);
+                       degree - 1);
 }
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_nce(cell::type celltype, int degree,
@@ -341,6 +341,6 @@ FiniteElement basix::element::create_nce(cell::type celltype, int degree,
 
   return FiniteElement(element::family::N1E, celltype, degree, {tdim}, wcoeffs,
                        entity_transformations, x, M, maps::type::covariantPiola,
-                       discontinuous, degree, degree);
+                       discontinuous, degree, degree - 1);
 }
 //-----------------------------------------------------------------------------
