+++ test/test_interpolation_between_elements.py
@@ -136,7 +136,6 @@ def test_blocked_interpolation(cell_type, order):
 
 @parametrize_over_elements(5)
 def test_interpolation_orders(cell_type, degree, element_type, element_args):
-    print(cell_type.name, element_type.name, degree)
     element = basix.create_element(element_type, cell_type, degree, *element_args)
 
     points = basix.create_lattice(cell_type, 10, basix.LatticeType.equispaced, True)
@@ -190,7 +189,6 @@ def test_interpolation_orders(cell_type, degree, element_type, element_args):
 
         assert np.allclose(values, lagrange_values)
 
-    print("->", element.highest_complete_polynomial_degree)
     if element.highest_complete_polynomial_degree >= -1:
         # This Lagrange space should NOT be a subset to the element being tested
         lagrange = basix.create_element(
