### 2021-11-26 39e3761
seed np.random
