### 2021-11-26 1a751d3
Add information about highest degree of polynomials in a finite element (#346)
* Added highest degree function to finite element

* Added highest space that is subset of element

* remove prints from test

* docs

* seed np.random

* rename function

* Minor edit

* typos

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>
