+++ test/test_interpolation_between_elements.py
@@ -5,6 +5,7 @@
 import numpy as np
 import basix
 import pytest
+from .utils import parametrize_over_elements
 
 
 def run_test(lower_element, higher_element, power, value_size):
@@ -131,3 +132,77 @@ def test_blocked_interpolation(cell_type, order):
     # Test interpolation from blocked Lagrange to Nedelec
     i_m = basix.compute_interpolation_operator(lagrange, nedelec)
     assert np.allclose(n_coeffs, i_m @ l_coeffs)
+
+
+@parametrize_over_elements(5)
+def test_degree_bounds(cell_type, degree, element_type, element_args):
+    np.random.seed(13)
+
+    element = basix.create_element(element_type, cell_type, degree, *element_args)
+
+    points = basix.create_lattice(cell_type, 10, basix.LatticeType.equispaced, True)
+    tab = element.tabulate(0, points)[0]
+
+    # Test that this element's basis functions are contained in Lagrange space with
+    # degree element.degree_bounds[1]
+    coeffs = np.random.rand(element.dim)
+    values = np.array([tab[:, :, i] @ coeffs
+                       for i in range(element.value_size)])
+
+    if element.degree_bounds[1] >= 0:
+        # The element being tested should be a subset of this Lagrange space
+        lagrange = basix.create_element(
+            basix.ElementFamily.P, cell_type, element.degree_bounds[1], basix.LagrangeVariant.equispaced, True)
+        lagrange_coeffs = basix.compute_interpolation_operator(element, lagrange) @ coeffs
+        lagrange_tab = lagrange.tabulate(0, points)[0]
+        lagrange_values = np.array([lagrange_tab[:, :, 0] @ lagrange_coeffs[i::element.value_size]
+                                    for i in range(element.value_size)])
+
+        assert np.allclose(values, lagrange_values)
+
+    if element.degree_bounds[1] >= 1:
+        # The element being tested should be NOT a subset of this Lagrange space
+        lagrange = basix.create_element(
+            basix.ElementFamily.P, cell_type, element.degree_bounds[1] - 1,
+            basix.LagrangeVariant.equispaced, True)
+        lagrange_coeffs = basix.compute_interpolation_operator(element, lagrange) @ coeffs
+        lagrange_tab = lagrange.tabulate(0, points)[0]
+        lagrange_values = np.array([lagrange_tab[:, :, 0] @ lagrange_coeffs[i::element.value_size]
+                                    for i in range(element.value_size)])
+
+        assert not np.allclose(values, lagrange_values)
+
+    # Test that the basis functions of Lagrange space with degree element.degree_bounds[0]
+    # are contained in this space
+
+    if element.degree_bounds[0] >= 0:
+        # This Lagrange space should be a subset to the element being tested
+        lagrange = basix.create_element(
+            basix.ElementFamily.P, cell_type, element.degree_bounds[0],
+            basix.LagrangeVariant.equispaced, True)
+        lagrange_coeffs = np.random.rand(lagrange.dim * element.value_size)
+        lagrange_tab = lagrange.tabulate(0, points)[0]
+        lagrange_values = np.array([lagrange_tab[:, :, 0] @ lagrange_coeffs[i::element.value_size]
+                                    for i in range(element.value_size)])
+
+        coeffs = basix.compute_interpolation_operator(lagrange, element) @ lagrange_coeffs
+        values = np.array([tab[:, :, i] @ coeffs
+                           for i in range(element.value_size)])
+
+        assert np.allclose(values, lagrange_values)
+
+    if element.degree_bounds[0] >= -1:
+        # This Lagrange space should NOT be a subset to the element being tested
+        lagrange = basix.create_element(
[…]
