### 2021-11-29 faa4414
start adding tensor product representation
