+++ test/test_tensor_products.py
@@ -16,12 +16,8 @@ def tensor_product(*data):
         return tensor_product(tensor_product(data[0], data[1]), *data[2:])
 
     a, b = data
-    assert a.shape[0] == b.shape[0] == 1  # TODO: implement for derivatives too
-    assert a.shape[3] == b.shape[3] == 1  # TODO: implement for non-scalar elements
-    out = np.empty((1, a.shape[1] * b.shape[1], a.shape[2] * b.shape[2], 1))
-    for n, (i, j) in enumerate(product(range(a.shape[1]), range(b.shape[1]))):
-        out[0, n, :, 0] = np.outer(a[0, i, :, 0], b[0, j, :, 0]).reshape(-1)
-    return out
+    assert a.shape[1] == b.shape[1] == 1  # TODO: implement for non-scalar elements
+    return np.outer(a[:, 0], b[:, 0]).reshape(-1)
 
 
 @parametrize_over_elements(4)
@@ -36,18 +32,20 @@ def test_tensor_product_factorisation(cell_type, degree, element_type, element_a
 
     factors = element.get_tensor_product_representation()
 
-    lattice_1d = basix.create_lattice(basix.CellType.interval, 1, basix.LatticeType.equispaced, True)
-    lattice_nd = np.array(list(product(*[lattice_1d for i in range(tdim)])))
+    lattice = basix.create_lattice(cell_type, 1, basix.LatticeType.equispaced, True)
 
-    tab1 = element.tabulate(0, lattice_nd)
+    tab1 = element.tabulate(0, lattice)
 
-    tab2 = np.empty((1, lattice_nd.shape[0], element.dim, 1))
+    for i, point in enumerate(lattice):
+        values1 = tab1[0, i, :, :]
 
-    for fs, perm in factors:
-        evals = [e.tabulate(0, lattice_1d) for e in fs]
-        tab = tensor_product(*evals)
-        for i, j in enumerate(perm):
-            if j != -1:
-                tab2[:, :, j, :] = tab[:, :, i, :]
+        values2 = np.empty((element.dim, 1))
 
-    assert np.allclose(tab1, tab2)
+        for fs, perm in factors:
+            evals = [e.tabulate(0, [p]) for e, p in zip(fs, point)]
+            tab2 = tensor_product(*evals)
+            for i, j in enumerate(perm):
+                if j != -1:
+                    values2[j, 0] = tab2[i]
+
+        assert np.allclose(values1, values2)
