### 2021-11-29 d7afed2
Add tensor representation of Q on hexes
