+++ test/test_quadrature.py
@@ -126,7 +126,7 @@ def test_gll():
     # 2D quad
     pts, wts = basix.make_quadrature(basix.QuadratureType.gll, basix.CellType.quadrilateral, m+1)
     pts, wts = 2*pts-1, 4*wts
-    ref_pts2 = np.array([[x, y] for y in ref_pts for x in ref_pts])
+    ref_pts2 = np.array([[x, y] for x in ref_pts for y in ref_pts])
     assert (np.allclose(pts, ref_pts2))
     ref_wts2 = np.array([w1*w2 for w1 in ref_wts for w2 in ref_wts])
     assert (np.allclose(wts, ref_wts2))
@@ -136,7 +136,7 @@ def test_gll():
     # 3D hex
     pts, wts = basix.make_quadrature(basix.QuadratureType.gll, basix.CellType.hexahedron, m+1)
     pts, wts = 2*pts-1, 8*wts
-    ref_pts3 = np.array([[x, y, z] for z in ref_pts for y in ref_pts for x in ref_pts])
+    ref_pts3 = np.array([[x, y, z] for x in ref_pts for y in ref_pts for z in ref_pts])
     assert (np.allclose(pts, ref_pts3))
     ref_wts3 = np.array([w1*w2*w3 for w1 in ref_wts for w2 in ref_wts for w3 in ref_wts])
     assert (np.allclose(wts, ref_wts3))
