### 2021-12-01 588928f
Add more tensor product tests (#351)
* add more tests

* fix flake8

* uncommnet test
