### 2021-12-01 f2c54e4
Add tensor product representation of Lagrange on quads and hexes (#349)
* start adding tensor product representation

* Refactor test

* Add tensor representation of Q on hexes

* Update docs

* Add more tensor product tests (#351)

* add more tests

* fix flake8

* uncommnet test

Co-authored-by: Igor Baratta <igorbaratta@gmail.com>
