+++ cpp/basix/finite-element.cpp
@@ -307,13 +307,16 @@ FiniteElement::FiniteElement(
     const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
     const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
     maps::type map_type, bool discontinuous, int highest_degree,
-    int highest_complete_degree)
+    int highest_complete_degree,
+    std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
+        tensor_factors)
     : _cell_type(cell_type), _cell_tdim(cell::topological_dimension(cell_type)),
       _cell_subentity_types(cell::subentity_types(cell_type)), _family(family),
       _degree(degree), _map_type(map_type),
       _entity_transformations(entity_transformations), _x(x),
       _discontinuous(discontinuous),
-      _degree_bounds({highest_complete_degree, highest_degree})
+      _degree_bounds({highest_complete_degree, highest_degree}),
+      _tensor_factors(tensor_factors)
 {
   _dual_matrix = compute_dual_matrix(cell_type, wcoeffs, M, x, degree);
   xt::xtensor<double, 2> B_cmajor({wcoeffs.shape(0), wcoeffs.shape(1)});
@@ -898,6 +901,19 @@ xt::xtensor<double, 2> FiniteElement::coefficient_matrix() const
 std::array<int, 2> FiniteElement::degree_bounds() const
 {
   return _degree_bounds;
+}
+//-----------------------------------------------------------------------------
+bool FiniteElement::has_tensor_product_factorisation() const
+{
+  return _tensor_factors.size() > 0;
+}
+//-----------------------------------------------------------------------------
+std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
+FiniteElement::get_tensor_product_representation() const
+{
+  if (!has_tensor_product_factorisation())
+    throw std::runtime_error("Element has no tensor product representation.");
+  return _tensor_factors;
 }
 //-----------------------------------------------------------------------------
 std::string basix::version()
