+++ cpp/basix/finite-element.h
@@ -222,15 +222,21 @@ public:
   /// @param[in] highest_complete_degree The highest degree n such that a
   /// Lagrange (or vector Lagrange) element of degree n is a subspace of this
   /// element
-  FiniteElement(element::family family, cell::type cell_type, int degree,
-                const std::vector<std::size_t>& value_shape,
-                const xt::xtensor<double, 2>& wcoeffs,
-                const std::map<cell::type, xt::xtensor<double, 3>>&
-                    entity_transformations,
-                const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
-                const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
-                maps::type map_type, bool discontinuous, int highest_degree,
-                int highest_complete_degree);
+  /// @param[in] tensor_factors The factors in the tensor product representation
+  /// of this element
+  FiniteElement(
+      element::family family, cell::type cell_type, int degree,
+      const std::vector<std::size_t>& value_shape,
+      const xt::xtensor<double, 2>& wcoeffs,
+      const std::map<cell::type, xt::xtensor<double, 3>>&
+          entity_transformations,
+      const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
+      const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
+      maps::type map_type, bool discontinuous, int highest_degree,
+      int highest_complete_degree,
+      std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
+          tensor_factors
+      = {});
 
   /// Copy constructor
   FiniteElement(const FiniteElement& element) = default;
@@ -754,6 +760,22 @@ public:
   /// subspace of this element
   std::array<int, 2> degree_bounds() const;
 
+  /// Indicates whether or not this element has a tensor product representation.
+  bool has_tensor_product_factorisation() const;
+
+  /// Get the tensor product representation of this element, or throw an error
+  /// if no such factorisation exists.
+  ///
+  /// The tensor product representation will be a vector of tuples. Each tuple
+  /// contains a vector of finite elements, and a vector on integers. The vector
+  /// of finite elements gives the elements on an interval that appear in the
+  /// tensor product representation. The vector of integers gives the
+  /// permutation between the numbering of the tensor product DOFs and the
+  /// number of the DOFs of this Basix element.
+  /// @return The tensor product representation
+  std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
+  get_tensor_product_representation() const;
+
 private:
   // Cell type
   cell::type _cell_type;
@@ -863,12 +885,19 @@ private:
   // The dual matrix
   xt::xtensor<double, 2> _dual_matrix;
 
-
   // Polynomial degree bounds
   // [0]: highest degree n such that Lagrange order n is a subspace of
   // this space
   // [1]: highest polynomial degree
   std::array<int, 2> _degree_bounds;
+
+  // Tensor product representation
+  // Entries of tuple are (list of elements on an interval, permutation of DOF
+  // numbers)
+  // @todo: For vector-valued elements, a tensor product type and a scaling
+  // factor may additionally be needed.
+  std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
+      _tensor_factors;
 };
 
 /// Create an element using a given Lagrange variant
