### 2021-12-01 7f0623e
Add z direction for tensor product test
