### 2021-12-02 7bdec34
Start work on integral moments against polynomials
