### 2021-12-02 068d538
put correct points into lattice and rotate right
