+++ cpp/basix/polynomials.cpp
@@ -4,6 +4,7 @@
 
 #include "polynomials.h"
 #include "polyset.h"
+#include <xtensor/xview.hpp>
 
 using namespace basix;
 
@@ -12,16 +13,19 @@ xt::xtensor<double, 2> polynomials::tabulate(polynomials::type polytype,
                                              cell::type celltype, int d,
                                              const xt::xarray<double>& x)
 {
-  throw std::runtime_error("not implemented yet");
-  assert(polytype);
-  assert(celltype);
-  assert(d);
-  assert(x(0, 0));
+  switch (polytype)
+  {
+  case polynomials::type::legendre:
+    return xt::view(polyset::tabulate(celltype, d, 0, x), 0, xt::all(),
+                    xt::all());
+  default:
+    throw std::runtime_error("not implemented yet a");
+  }
 }
 //-----------------------------------------------------------------------------
 int polynomials::dim(polynomials::type polytype, cell::type cell, int d)
 {
-  throw std::runtime_error("not implemented yet");
+  throw std::runtime_error("not implemented yet b");
   assert(polytype);
   assert(cell);
   assert(d);
