### 2021-12-02 5148239
Reordering of GLL quad points and weights (#355)
* Add z direction for tensor product test

* Rotate GLL quad points and weights

* put correct points into lattice and rotate right

* Reorder test GLL points and weights

* Improve comment

Co-authored-by: Igor Baratta <igorbaratta@gmail.com>

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>
Co-authored-by: Igor Baratta <igorbaratta@gmail.com>
