+++ cpp/basix/moments.cpp
@@ -115,9 +115,6 @@ xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
   {
     std::array<std::size_t, 3> shape = {2, pts.shape(0), pts.shape(1)};
     tpts = xt::zeros<double>(shape);
-    J.resize({2, 2, 2});
-    K.resize({2, 2, 2});
-    xt::xtensor_fixed<double, xt::xshape<2, 2>> A = xt::zeros<double>({2, 2});
     for (std::size_t i = 0; i < pts.shape(0); ++i)
     {
       tpts(0, i, 0) = pts(i, 1);
@@ -125,28 +122,14 @@ xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
       tpts(1, i, 0) = pts(i, 1);
       tpts(1, i, 1) = pts(i, 0);
     }
-
-    A = {{0, 1}, {-1, -1}};
-    xt::view(J, 0, xt::all(), xt::all()) = A;
-    A = {{-1, -1}, {1, 0}};
-    xt::view(K, 0, xt::all(), xt::all()) = A;
-
-    A = {{0, 1}, {1, 0}};
-    xt::view(J, 1, xt::all(), xt::all()) = A;
-    A = {{0, 1}, {1, 0}};
-    xt::view(K, 1, xt::all(), xt::all()) = A;
-
+    J = {{{0.0, 1.0}, {-1.0, -1.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
+    K = {{{-1.0, -1.0}, {1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
     break;
   }
   case cell::type::quadrilateral:
   {
     std::array<std::size_t, 3> shape0 = {2, pts.shape(0), pts.shape(1)};
     tpts = xt::zeros<double>(shape0);
-
-    J.resize({2, 2, 2});
-    K.resize({2, 2, 2});
-    xt::xtensor_fixed<double, xt::xshape<2, 2>> A = xt::zeros<double>({2, 2});
-
     for (std::size_t i = 0; i < pts.shape(0); ++i)
     {
       tpts(0, i, 0) = pts(i, 1);
@@ -154,17 +137,8 @@ xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
       tpts(1, i, 0) = pts(i, 1);
       tpts(1, i, 1) = pts(i, 0);
     }
-
-    A = {{0, 1}, {-1, 0}};
-    xt::view(J, 0, xt::all(), xt::all()) = A;
-    A = {{0, -1}, {1, 0}};
-    xt::view(K, 0, xt::all(), xt::all()) = A;
-
-    A = {{0, 1}, {1, 0}};
-    xt::view(J, 1, xt::all(), xt::all()) = A;
-    A = {{0, 1}, {1, 0}};
-    xt::view(K, 1, xt::all(), xt::all()) = A;
-
+    J = {{{0.0, 1.0}, {-1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
+    K = {{{0.0, -1.0}, {1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
     break;
   }
   default:
@@ -226,6 +200,92 @@ xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
   return out;
 }
 //----------------------------------------------------------------------------
+xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
+    polynomials::type polytype, cell::type celltype, int degree, int qdeg)
+{
+  // This function can be dramatically simplified and made
+  // understandable by using tensors to give more logic to the objects
+
+  if (celltype == cell::type::point)
[…]
