### 2021-12-07 6a51786
Remove xtensor, make download optional
