+++ cpp/CMakeLists.txt
@@ -13,14 +13,6 @@ endif()
 
 include(FeatureSummary)
 
-# Enable SIMD with xtensor
-option(XTENSOR_USE_XSIMD "Enable SIMD with xtensor" OFF)
-add_feature_info(XTENSOR_USE_XSIMD XTENSOR_USE_XSIMD "Enable SIMD with xtensor.")
-
-# Enable xtensor with target-specific optimization, i.e. -march=native
-option(XTENSOR_OPTIMIZE "Enable xtensor target-specific optimization" OFF)
-add_feature_info(XTENSOR_OPTIMIZE XTENSOR_OPTIMIZE "Enable architecture-specific optimizations as defined by xtensor.")
-
 # Set the C++ standard
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -30,62 +22,45 @@ set(CMAKE_CXX_EXTENSIONS OFF)
 option(BUILD_SHARED_LIBS "Build Basix with shared libraries." ON)
 add_feature_info(BUILD_SHARED_LIBS BUILD_SHARED_LIBS "Build Basix with shared libraries.")
 
+# Download xtl and xtensor
+option(XLIBS_DOWNLOAD "Download xtl and xtensor. Requires git." OFF)
+add_feature_info(XLIBS_DOWNLOAD XLIBS_DOWNLOAD "Download xtl and xtensor. Requires git.")
+
+# Enable xtensor with target-specific optimization, i.e. -march=native
+option(XTENSOR_OPTIMIZE "Enable xtensor target-specific optimization" OFF)
+add_feature_info(XTENSOR_OPTIMIZE XTENSOR_OPTIMIZE "Enable architecture-specific optimizations as defined by xtensor.")
+
 # Find dependecies
 set(XTENSOR_MIN_VERSION 0.23.10)
 set(XTL_MIN_VERSION 0.7.0)
-set(XSIMD_MIN_VERSION 7.4.10)
 
 include(FetchContent)
 
-# xsimd (optional)
-if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.19)
-  find_package(xsimd ${XSIMD_MIN_VERSION}...<8.0 QUIET)
+if(NOT XLIBS_DOWNLOAD)
+  # xtl, xtensor (required)
+  if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.19)
+    find_package(xtl ${XTL_MIN_VERSION}...<0.8)
+    find_package(xtensor ${XTENSOR_MIN_VERSION}...<0.25)
+  else()
+    find_package(xtl ${XTL_MIN_VERSION})
+    find_package(xtensor ${XTENSOR_MIN_VERSION})
+  endif()
 else()
-  find_package(xsimd ${XSIMD_MIN_VERSION} QUIET)
-endif()
-if(XTENSOR_USE_XSIMD AND NOT xsimd_FOUND)
-  message("downloading xsimd source...")
-  FetchContent_Declare(
-    xsimd
-    GIT_REPOSITORY https://github.com/xtensor-stack/xsimd.git
-    GIT_TAG        7.4.10
-  )
-  FetchContent_MakeAvailable(xsimd)
-else()
-  message("found xsimd ${xsimd_VERSION}")
-endif()
-
-# xtl, xtensor (required)
-if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.19)
-  find_package(xtl ${XTL_MIN_VERSION}...<0.8 QUIET)
-  find_package(xtensor ${XTENSOR_MIN_VERSION}...<0.24 QUIET)
-else()
[…]
