+++ setup.py
@@ -16,6 +16,7 @@ setup(name="fenics-basix",
       cmake_args=[
           '-DPython3_EXECUTABLE=' + sys.executable,
           '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST"),
-          '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY")],
+          '-DPython3_INCLUDE_DIRS=' + sysconfig.get_config_var("INCLUDEPY"),
+          '-DXLIBS_DOWNLOAD=ON'],
       package_dir={"": "python"},
       cmake_install_dir="python/basix/")
