+++ cpp/CMakeLists.txt
@@ -39,11 +39,11 @@ include(FetchContent)
 if(NOT XLIBS_DOWNLOAD)
   # xtl, xtensor (required)
   if(CMAKE_VERSION VERSION_GREATER_EQUAL 3.19)
-    find_package(xtl ${XTL_MIN_VERSION}...<0.8)
-    find_package(xtensor ${XTENSOR_MIN_VERSION}...<0.25)
+    find_package(xtl ${XTL_MIN_VERSION}...<0.8 REQUIRED)
+    find_package(xtensor ${XTENSOR_MIN_VERSION}...<0.25 REQUIRED)
   else()
-    find_package(xtl ${XTL_MIN_VERSION})
-    find_package(xtensor ${XTENSOR_MIN_VERSION})
+    find_package(xtl ${XTL_MIN_VERSION} REQUIRED)
+    find_package(xtensor ${XTENSOR_MIN_VERSION} REQUIRED)
   endif()
 else()
   message("Downloading xtl source...")
