+++ test/test_lagrange.py
@@ -435,9 +435,9 @@ def test_vtk_element(celltype, degree):
 @pytest.mark.parametrize("degree", range(1, 5))
 def test_integral_lagrange(celltype, degree, variant):
     e = basix.create_element(basix.ElementFamily.P, celltype, degree, variant, True)
-
     for p in e.points:
         assert in_cell(celltype, p)
 
     e = basix.create_element(basix.ElementFamily.P, celltype, degree, variant)
-
+    for p in e.points:
+        assert in_cell(celltype, p)
