### 2021-12-08 #3982 @whedon
[REVIEW]: Basix: a runtime finite element basis evaluation library
**Submitting author:** @mscroggs (<a href="http://orcid.org/0000-0002-4658-2443">Matthew Scroggs</a>)
**Repository:** <a href="https://github.com/FEniCS/basix" target ="_blank">https://github.com/FEniCS/basix</a>
**Version:** v0.3.0
