+++ cpp/basix/polynomials.h
@@ -21,28 +21,13 @@ enum class type
 
 /// Tabulate a set of polynomials.
 ///
+/// @param[in] polytype Polynomial type
 /// @param[in] celltype Cell type
 /// @param[in] d Polynomial degree
 /// @param[in] x Points at which to evaluate the basis. The shape is
 /// (number of points, geometric dimension).
 /// @return Polynomial sets, for each derivative, tabulated at points.
-/// The shape is `(number of derivatives computed, number of points,
-/// basis index)`.
-///
-/// - The first index is the derivative. The first entry is the basis
-/// itself. Derivatives are stored in triangular (2D) or tetrahedral
-/// (3D) ordering, e.g. if `(p, q)` denotes `p` order dervative with
-/// repsect to `x` and `q` order derivative with respect to `y`, [0] ->
-/// (0, 0), [1] -> (1, 0), [2] -> (0, 1), [3] -> (2, 0), [4] -> (1, 1),
-/// [5] -> (0, 2), [6] -> (3, 0),...
-/// The function basix::indexing::idx maps tuples `(p, q, r)` to the array
-/// index.
-///
-/// - The second index is the point, with index `i` correspondign to the
-/// point in row `i` of @p x.
-///
-/// - The third index is the basis function index.
-/// @todo Does the order for the third index need to be documented?
+/// The shape is `(number of points, basis index)`.
 xt::xtensor<double, 2> tabulate(polynomials::type polytype, cell::type celltype,
                                 int d, const xt::xarray<double>& x);
 
