### 2021-12-09 4eae08a
Adjust constants so polyset is orthonormal without constants
