### 2021-12-09 231bb35
Pass in/out memory into polysets
