+++ cpp/basix/polyset.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 Chris Richardson
+// Copyright (c) 2020 Chris Richardson & Matthew Scroggs
 // FEniCS Project
 // SPDX-License-Identifier:    MIT
 
@@ -29,12 +29,14 @@ constexpr std::array<double, 3> jrc(int a, int n)
 //-----------------------------------------------------------------------------
 // At a point, only the constant polynomial can be used. This has value 1 and
 // derivative 0.
-xt::xtensor<double, 3>
-tabulate_polyset_point_derivs(std::size_t, std::size_t nderiv,
-                              const xt::xtensor<double, 2>& x)
+void tabulate_polyset_point_derivs(xt::xtensor<double, 3>& P, std::size_t,
+                                   std::size_t nderiv,
+                                   const xt::xtensor<double, 2>& x)
 {
   assert(x.shape(0) > 0);
-  xt::xtensor<double, 3> P({nderiv + 1, x.shape(0), 1});
+  assert(P.shape(0) == nderiv + 1);
+  assert(P.shape(1) == x.shape(0));
+  assert(P.shape(2) == 1);
   for (std::size_t k = 0; k <= nderiv; ++k)
   {
     auto result = xt::view(P, k, xt::all(), xt::all());
@@ -43,8 +45,6 @@ tabulate_polyset_point_derivs(std::size_t, std::size_t nderiv,
     else
       xt::col(result, 0) = 0.0;
   }
-
-  return P;
 }
 //-----------------------------------------------------------------------------
 // Compute the complete set of derivatives from 0 to nderiv, for all the
@@ -52,14 +52,16 @@ tabulate_polyset_point_derivs(std::size_t, std::size_t nderiv,
 // Legendre Polynomials, with the recurrence relation given by
 // n P(n) = (2n - 1) x P_{n-1} - (n - 1) P_{n-2} in the interval [-1, 1]. The
 // range is rescaled here to [0, 1].
-xt::xtensor<double, 3>
-tabulate_polyset_line_derivs(std::size_t degree, std::size_t nderiv,
-                             const xt::xtensor<double, 1>& x)
+void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
+                                  std::size_t nderiv,
+                                  const xt::xtensor<double, 1>& x)
 {
   assert(x.shape(0) > 0);
   const auto X = x * 2.0 - 1.0;
   const std::size_t m = (degree + 1);
-  xt::xtensor<double, 3> P({nderiv + 1, x.shape(0), m});
+  assert(P.shape(0) == nderiv + 1);
+  assert(P.shape(1) == x.shape(0));
+  assert(P.shape(2) == m);
   for (std::size_t k = 0; k <= nderiv; ++k)
   {
     // Get reference to this derivative
@@ -85,8 +87,6 @@ tabulate_polyset_line_derivs(std::size_t degree, std::size_t nderiv,
   for (std::size_t k = 0; k < nderiv + 1; ++k)
     for (std::size_t p = 0; p <= degree; ++p)
       xt::view(P, k, xt::all(), p) *= std::sqrt(2 * p + 1);
-
-  return P;
 }
 //-----------------------------------------------------------------------------
 // Compute the complete set of derivatives from 0 to nderiv, for all the
@@ -96,9 +96,9 @@ tabulate_polyset_line_derivs(std::size_t degree, std::size_t nderiv,
 // above, but with a change of variables. The polynomials are then
 // extended in the q direction, using the relation given in Sherwin and
[…]
