### 2021-12-09 5127d59
remove unused variable
