+++ cpp/basix/polyset.cpp
@@ -841,7 +841,7 @@ int polyset::nderivs(cell::type celltype, int n)
   case cell::type::interval:
     return n + 1;
   case cell::type::triangle:
-    return (n + 1) * (n + 1);
+    return (n + 1) * (n + 2) / 2;
   case cell::type::quadrilateral:
     return (n + 1) * (n + 2) / 2;
   case cell::type::tetrahedron:
