### 2021-12-09 e123493
Remove xtensor, make download via git explicit option (#357)
* Remove xtensor, make download optional

* scikit-learn download xlibs

* remove xsimd

* Add REQUIRED.

* Change variable name.

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>
