+++ python/wrapper.cpp
@@ -380,6 +380,8 @@ Interface to the Basix C++ library.
       .value("gl_isaac", element::lagrange_variant::gl_isaac)
       .value("gl_centroid", element::lagrange_variant::gl_centroid)
       .value("integral_legendre", element::lagrange_variant::integral_legendre)
+      .value("integral_chebyshev",
+             element::lagrange_variant::integral_chebyshev)
       .value("vtk", element::lagrange_variant::vtk);
 
   // Create FiniteElement
