+++ python/docs.h
@@ -237,16 +237,16 @@ Returns
 =======
 numpy.ndarray[numpy.float64]
     The basis functions (and derivatives). The shape is
-(derivative, point, basis fn index, value index).
-- The first index is the derivative, with higher derivatives are
-stored in triangular (2D) or tetrahedral (3D) ordering, i.e. for
-the (x,y) derivatives in 2D: (0,0), (1,0), (0,1), (2,0), (1,1),
-(0,2), (3,0)... The function basix::indexing::idx can be used to find the
-appropriate derivative.
-- The second index is the point index
-- The third index is the basis function index
-- The fourth index is the basis function component. Its has size
-one for scalar basis functions.
+    (derivative, point, basis fn index, value index).
+    - The first index is the derivative, with higher derivatives are
+    stored in triangular (2D) or tetrahedral (3D) ordering, i.e. for
+    the (x,y) derivatives in 2D: (0,0), (1,0), (0,1), (2,0), (1,1),
+    (0,2), (3,0)... The function basix::indexing::idx can be used to find the
+    appropriate derivative.
+    - The second index is the point index
+    - The third index is the basis function index
+    - The fourth index is the basis function component. Its has size
+    one for scalar basis functions.
 )";
 
 const std::string FiniteElement__push_forward = R"(
@@ -269,7 +269,7 @@ Returns
 =======
 numpy.ndarray[numpy.float64]
     The function values on the cell. The indices are [Jacobian
-index, point index, components].
+    index, point index, components].
 )";
 
 const std::string FiniteElement__pull_back = R"(
@@ -600,7 +600,7 @@ Returns
 =======
 numpy.ndarray[numpy.float64]
     Matrix operator that maps the 'from' degrees-of-freedom to
-the 'to' degrees-of-freedom
+    the 'to' degrees-of-freedom
 )";
 
 const std::string tabulate_polynomial_set = R"(
@@ -634,23 +634,23 @@ Returns
 =======
 numpy.ndarray[numpy.float64]
     Polynomial sets, for each derivative, tabulated at points.
-The shape is `(number of derivatives computed, number of points,
-basis index)`.
-
-- The first index is the derivative. The first entry is the basis
-itself. Derivatives are stored in triangular (2D) or tetrahedral
-(3D) ordering, e.g. if `(p, q)` denotes `p` order dervative with
-repsect to `x` and `q` order derivative with respect to `y`, [0] ->
-(0, 0), [1] -> (1, 0), [2] -> (0, 1), [3] -> (2, 0), [4] -> (1, 1),
-[5] -> (0, 2), [6] -> (3, 0),...
-The function basix::indexing::idx maps tuples `(p, q, r)` to the array
-index.
-
-- The second index is the point, with index `i` correspondign to the
-point in row `i` of @p x.
-
-- The third index is the basis function index.
-TODO: Does the order for the third index need to be documented?
+    The shape is `(number of derivatives computed, number of points,
+    basis index)`.
+    
+    - The first index is the derivative. The first entry is the basis
+    itself. Derivatives are stored in triangular (2D) or tetrahedral
+    (3D) ordering, e.g. if `(p, q)` denotes `p` order dervative with
+    repsect to `x` and `q` order derivative with respect to `y`, [0] ->
+    (0, 0), [1] -> (1, 0), [2] -> (0, 1), [3] -> (2, 0), [4] -> (1, 1),
+    [5] -> (0, 2), [6] -> (3, 0),...
+    The function basix::indexing::idx maps tuples `(p, q, r)` to the array
+    index.
+    
+    - The second index is the point, with index `i` correspondign to the
+    point in row `i` of @p x.
+    
[…]
