### 2021-12-10 c98f628
tidy test utils file
