+++ test/utils.py
@@ -3,71 +3,56 @@
 # SPDX-License-Identifier: MIT
 
 import pytest
-import basix
+from basix import ElementFamily, CellType, LagrangeVariant
 
 
 def parametrize_over_elements(degree, reference=None):
     elementlist = []
 
-    elementlist += [(c, basix.ElementFamily.P, o, [basix.LagrangeVariant.integral_legendre])
-                    for c in [basix.CellType.interval, basix.CellType.triangle,
-                              basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.P, o, [basix.LagrangeVariant.gll_isaac])
-                    for c in [basix.CellType.interval, basix.CellType.triangle,
-                              basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron,
-                              basix.CellType.prism]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.P, o, [basix.LagrangeVariant.gll_warped])
-                    for c in [basix.CellType.interval, basix.CellType.triangle,
-                              basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron,
-                              basix.CellType.prism]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.P, o, [basix.LagrangeVariant.equispaced])
-                    for c in [basix.CellType.interval, basix.CellType.triangle,
-                              basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron,
-                              basix.CellType.prism, basix.CellType.pyramid]
-                    for o in range(1, min(4, degree + 1))]
-    elementlist += [(c, basix.ElementFamily.N1E, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.RT, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.N2E, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.BDM, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron,
-                              basix.CellType.quadrilateral, basix.CellType.hexahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(c, basix.ElementFamily.CR, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron]
-                    for o in range(1, min(2, degree + 1))]
-    elementlist += [(c, basix.ElementFamily.Regge, o, [])
-                    for c in [basix.CellType.triangle, basix.CellType.tetrahedron]
-                    for o in range(1, degree + 1)]
-    elementlist += [(basix.CellType.interval, basix.ElementFamily.bubble, o, [])
-                    for o in range(2, degree + 1)]
[…]
