+++ test/test_dof_transformations.py
@@ -64,15 +64,15 @@ def test_apply_to_transpose(cell_type, element_type, degree, element_args):
         assert np.allclose(data1.transpose(), data2)
 
 
-@parametrize_over_elements(5, "interval")
+@parametrize_over_elements(5, basix.CellType.interval)
 def test_interval_transformation_size(element_type, degree, element_args):
-    e = basix.create_element(element_type, "interval", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.interval, degree, *element_args)
     assert len(e.base_transformations()) == 0
 
 
-@parametrize_over_elements(5, "triangle")
+@parametrize_over_elements(5, basix.CellType.triangle)
 def test_triangle_transformation_degrees(element_type, degree, element_args):
-    e = basix.create_element(element_type, "triangle", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.triangle, degree, *element_args)
     bt = e.base_transformations()
     assert len(bt) == 3
     identity = np.identity(e.dim)
@@ -82,9 +82,9 @@ def test_triangle_transformation_degrees(element_type, degree, element_args):
             identity)
 
 
-@parametrize_over_elements(5, "tetrahedron")
+@parametrize_over_elements(5, basix.CellType.tetrahedron)
 def test_tetrahedron_transformation_degrees(element_type, degree, element_args):
-    e = basix.create_element(element_type, "tetrahedron", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.tetrahedron, degree, *element_args)
     bt = e.base_transformations()
     assert len(bt) == 14
     identity = np.identity(e.dim)
@@ -94,9 +94,9 @@ def test_tetrahedron_transformation_degrees(element_type, degree, element_args):
             identity)
 
 
-@parametrize_over_elements(5, "quadrilateral")
+@parametrize_over_elements(5, basix.CellType.quadrilateral)
 def test_quadrilateral_transformation_degrees(element_type, degree, element_args):
-    e = basix.create_element(element_type, "quadrilateral", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.quadrilateral, degree, *element_args)
     bt = e.base_transformations()
     assert len(bt) == 4
     identity = np.identity(e.dim)
@@ -106,9 +106,9 @@ def test_quadrilateral_transformation_degrees(element_type, degree, element_args
             identity)
 
 
-@parametrize_over_elements(5, "hexahedron")
+@parametrize_over_elements(5, basix.CellType.hexahedron)
 def test_hexahedron_transformation_degrees(element_type, degree, element_args):
-    e = basix.create_element(element_type, "hexahedron", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.hexahedron, degree, *element_args)
     bt = e.base_transformations()
     assert len(bt) == 24
     identity = np.identity(e.dim)
@@ -119,9 +119,9 @@ def test_hexahedron_transformation_degrees(element_type, degree, element_args):
             identity)
 
 
-@parametrize_over_elements(5, "prism")
+@parametrize_over_elements(5, basix.CellType.prism)
 def test_prism_transformation_degrees(element_type, degree, element_args):
-    e = basix.create_element(element_type, "prism", degree, *element_args)
+    e = basix.create_element(element_type, basix.CellType.prism, degree, *element_args)
     bt = e.base_transformations()
     assert len(bt) == 19
     identity = np.identity(e.dim)
[…]
