+++ python/basix/__init__.py
@@ -8,6 +8,6 @@ functionality can be used via this Python interface.
 from ._basixcpp import __version__
 from . import cell, finite_element, lattice, quadrature, variants
 from ._basixcpp import (CellType, LatticeType, LatticeSimplexMethod, ElementFamily, LagrangeVariant,
-                        QuadratureType)
+                        QuadratureType, PolynomialType)
 from ._basixcpp import (create_lattice, create_element, compute_interpolation_operator, topology,
-                        geometry, make_quadrature, index)
+                        geometry, make_quadrature, index, tabulate_polynomials)
