### 2022-01-07 8357a10
Scipy not used in test. Remove as test dependency.
