+++ .github/workflows/pythonapp.yml
@@ -55,7 +55,7 @@ jobs:
 
       - name: Run units tests
         run: |
-          pip install pytest pytest-xdist numpy sympy numba scipy
+          pip install pytest pytest-xdist numpy sympy numba
           pytest -n auto --durations 20 test/
 
       - name: Run python demos
