+++ pyproject.toml
@@ -7,7 +7,7 @@ build-backend = "setuptools.build_meta"
 build = ["cp{37,38,39}-manylinux_x86_64", "cp{37,38,39}-macosx_x86_64"]
 
 test-command = "cmake -DPython3_EXECUTABLE=$(which python) -G Ninja -B build -S {project}/test/test_cmake && cmake --build build/ && build/a.out && python -m pytest -v -n 2 --durations 20 {project}/test/"
-test-requires = ["pytest", "pytest-xdist", "numpy", "sympy", "numba", "scipy"]
+test-requires = ["pytest", "pytest-xdist", "numpy", "sympy", "numba"]
 
 manylinux-x86_64-image = "manylinux2014"
 
