### 2022-01-09 de34325
Set demo CMake language to CXX
