### 2022-01-11 336d192
Store the Lagrange variant used to initialise an element
