+++ .github/workflows/dolfin-tests.yml
@@ -37,7 +37,7 @@ jobs:
       - name: Install FEniCS Python components
         run: |
           python3 -m pip install git+https://github.com/FEniCS/ufl.git
-          python3 -m pip install git+https://github.com/FEniCS/ffcx.git@mscroggs/function_space
+          python3 -m pip install git+https://github.com/FEniCS/ffcx.git
 
       - name: Run cpp demos
         run: |
@@ -48,7 +48,7 @@ jobs:
         with:
           path: ./dolfinx
           repository: FEniCS/dolfinx
-          ref: mscroggs/function_space_check
+          ref: main
 
       - name: Install DOLFINx
         run: |
