### 2022-01-11 c5f0133
Store the Lagrange variant when creating a Lagrange element (#368)
* Store the Lagrange variant used to initialise an element

* branches

* revert branches

* doc
