+++ cpp/basix/finite-element.h
@@ -224,6 +224,7 @@ public:
   /// element
   /// @param[in] tensor_factors The factors in the tensor product representation
   /// of this element
+  /// @param[in] lvariant The Lagrange variant of the element
   FiniteElement(
       element::family family, cell::type cell_type, int degree,
       const std::vector<std::size_t>& value_shape,
@@ -236,7 +237,8 @@ public:
       int highest_complete_degree,
       std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
           tensor_factors
-      = {});
+      = {},
+      element::lagrange_variant lvariant = element::lagrange_variant::unset);
 
   /// Copy constructor
   FiniteElement(const FiniteElement& element) = default;
@@ -340,6 +342,11 @@ public:
   /// Get the finite element family
   /// @return The family
   element::family family() const;
+
+  /// Get the Lagrange variant of the element, or throw an error if element has
+  /// no Lagrange variant
+  /// @return The Lagrange variant
+  element::lagrange_variant lagrange_variant() const;
 
   /// Get the map type for this element
   /// @return The map type
@@ -788,6 +795,9 @@ private:
 
   // Finite element family
   element::family _family;
+
+  // Lagrange variant
+  element::lagrange_variant _lagrange_variant;
 
   // Degree
   int _degree;
