### 2022-01-13 FEniCS/basix#369 @wence-
[JOSS]: numpy requirement not satisfied after pip install
In a clean virtual environment, installing the python interface following `INSTALL.md` does not provide a usable library because the numpy requirement is not satisfied

Steps to reproduce

```
$ git clone git@github.com/fenics/basix.git
$ python3 -m venv basix-test
$ . basix-test/bin/activate
$ cd basix
$ pip install .
[wait...]
$ cat << EOF | python
import basix
celltype = basix.CellType.triangle
pts = basix.create_lattice(celltype, 1, basix.LatticeType.equispaced, True)
EOF

Traceback (most recent call last):
  File "<stdin>", line 3, in <module>
ModuleNotFoundError: No module named 'numpy'
```
I guess that numpy should be specified as a requirement in `pyproject.toml`?
