### 2022-01-13 FEniCS/basix#370 @wence-
[JOSS]: sympy requirement for some python tests not mentioned in INSTALL.md
The requirement of some tests on sympy is not mentioned in the install instructions, and so collection of those tests fails.

Steps to reproduce

```
$ git clone git@github.com:fenics/basix.git
$ python3 -m venv basix-test
$ . basix-test/bin/activate
$ cd basix
$ pip install .
$ pip install numpy pytest
$ pytest --collect-only test/
[...]
test/test_rt.py:8: in <module>
    import sympy
E   ModuleNotFoundError: No module named 'sympy'
========================================================= short test summary info =========================================================
ERROR test/test_lagrange.py
ERROR test/test_nedelec.py
ERROR test/test_nedelec_second_kind.py
ERROR test/test_orthonormal_basis_symbolic.py
ERROR test/test_quadrature.py
ERROR test/test_rt.py
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 6 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
================================================= 3101 tests collected, 6 errors in 7.14s =================================================
```

After installing sympy the tests pass.
