### 2022-01-13 FEniCS/basix#369 @jhale
scikit-build doesn't even support `setup.cfg` which was a transition step between old-fashioned `setup.py` and modern `pyproject.toml`

https://github.com/scikit-build/scikit-build/issues/591
