### 2022-01-13 e6c1c24
contributing: Fix typo
