+++ .github/workflows/intel.yml
@@ -50,10 +50,10 @@ jobs:
       - name: Install Basix
         run: |
           . /opt/intel/oneapi/setvars.sh
-          pip install .
+          pip install .[test]
 
       - name: Run units tests
         run: |
+          pip install pytest-xdist
           . /opt/intel/oneapi/setvars.sh
-          pip install pytest pytest-xdist numpy sympy numba scipy
           pytest -v -n auto --durations 20 test/
