+++ setup.py
@@ -14,14 +14,13 @@ setup(name="fenics-basix",
       license="MIT",
       packages=["basix"],
       install_requires=["numpy"],
-      extras_requires={
+      extras_require={
           "docs": ["markdown", "pylit3", "pyyaml", "sphinx", "sphinx_rtd_theme"],
           "lint": ["flake8", "pydocstyle"],
           "optional": ["numba"],
           "test": ["pytest", "numpy", "sympy", "numba", "scipy"],
-          "ci": ["pytest-xdist", "fenics-basix[docs]", "fenics-basix[lint]",
-                 "fenics-basix[optional]", "fenics-basix[test]"]
-      },
+          "ci": ["pytest-xdist", "fenics-basix[docs]", "fenics-basix[lint]", "fenics-basix[optional]", "fenics-basix[test]"]
+      }
       cmake_args=[
           '-DPython3_EXECUTABLE=' + sys.executable,
           '-DPython3_LIBRARIES=' + sysconfig.get_config_var("LIBDEST"),
