### 2022-01-13 5cd6e21
Improve install instruction
