### 2022-01-13 9d84350
Python packaging fixes (#372)
* Specify extras, cleanup CI scripts.

* Fix

* Fix syntax

* Typo

* Fix syntax

* Improve install instruction

* Tweak.

* Remove numpy
