+++ pyproject.toml
@@ -6,8 +6,9 @@ build-backend = "setuptools.build_meta"
 [tool.cibuildwheel]
 build = ["cp{37,38,39}-manylinux_x86_64", "cp{37,38,39}-macosx_x86_64"]
 
-test-command = "cmake -DPython3_EXECUTABLE=$(which python) -G Ninja -B build -S {project}/test/test_cmake && cmake --build build/ && build/a.out && python -m pytest -v -n 2 --durations 20 {project}/test/"
-test-requires = ["pytest", "pytest-xdist", "numpy", "sympy", "numba", "scipy"]
+test-command = "cmake -DPython3_EXECUTABLE=$(which python) -G Ninja -B build -S {project}/test/test_cmake && cmake --build build/ && build/a.out && python -m pytest -v -n auto --durations 20 {project}/test/"
+test-requires = ["pytest-xdist"]
+test-extras = ["test"]
 
 manylinux-x86_64-image = "manylinux2014"
 
