### 2022-01-13 #3982 @wence-
Installation issues on first pass review

- [ ] fenics/basix#369
- [ ] fenics/basix#370

(Will update this issue with more comments as I proceed).
