### 2022-01-13 FEniCS/basix#369 @wence-
Thanks, that worked for me.
