### 2022-01-14 c50400a
Add elements_are_equal function
