### 2022-01-14 3ae3845
check map type as well to be safe
