### 2022-01-14 d8c2962
Add __eq__ to pybind wrapper
