### 2022-01-14 7a40bcd
Fix Lagrange value_shape
