### 2022-01-14 eb83f6a
Minor housekeeping (#373)
* Minor housekeeping

* a -> an and regenerate docs

Co-authored-by: Matthew Scroggs <matthew.w.scroggs@gmail.com>
