### 2022-01-14 06d0e19
Check if two elements are equal (#374)
* Add elements_are_equal function

* check map type as well to be safe

* overload ==

* Add __eq__ to pybind wrapper

* newline

* docs

* Doc updates

Co-authored-by: Garth N. Wells <gnw20@cam.ac.uk>
