### 2022-01-14 ce04352
Remove value_size
