+++ test/test_dof_transformations.py
@@ -168,7 +168,7 @@ def test_transformation_of_tabulated_data_triangle(element_type, degree, element
         mapped_values = e.push_forward(reflected_values, J, detJ, K)
 
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose((bt[0].dot(i_slice))[start: start + ndofs],
@@ -197,7 +197,7 @@ def test_transformation_of_tabulated_data_quadrilateral(element_type, degree, el
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(reflected_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose((bt[0].dot(i_slice))[start: start + ndofs],
@@ -227,7 +227,7 @@ def test_transformation_of_tabulated_data_tetrahedron(element_type, degree, elem
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(reflected_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose((bt[0].dot(i_slice))[start: start + ndofs],
@@ -246,7 +246,7 @@ def test_transformation_of_tabulated_data_tetrahedron(element_type, degree, elem
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(rotated_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose(bt[6].dot(i_slice)[start: start + ndofs],
@@ -263,7 +263,7 @@ def test_transformation_of_tabulated_data_tetrahedron(element_type, degree, elem
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(reflected_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose((bt[7].dot(i_slice))[start: start + ndofs],
@@ -297,7 +297,7 @@ def test_transformation_of_tabulated_data_hexahedron(element_type, degree, eleme
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(reflected_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose((bt[0].dot(i_slice))[start: start + ndofs],
@@ -316,7 +316,7 @@ def test_transformation_of_tabulated_data_hexahedron(element_type, degree, eleme
         K = np.array([np.linalg.inv(_J) for p in points])
         mapped_values = e.push_forward(rotated_values, J, detJ, K)
         for i, j in zip(values, mapped_values):
-            for d in range(e.value_size):
+            for d in range(e.value_size()):
                 i_slice = i[:, d]
                 j_slice = j[:, d]
                 assert np.allclose(bt[12].dot(i_slice)[start: start + ndofs],
[…]
