+++ test/test_interpolation_between_elements.py
@@ -36,7 +36,7 @@ def test_different_order_interpolation_lagrange(cell_type, orders):
     lower_element = basix.create_element(basix.ElementFamily.P, cell_type, orders[0], basix.LagrangeVariant.gll_warped)
     higher_element = basix.create_element(basix.ElementFamily.P, cell_type, orders[1], basix.LagrangeVariant.gll_warped)
 
-    run_test(lower_element, higher_element, orders[0], lower_element.value_size)
+    run_test(lower_element, higher_element, orders[0], lower_element.value_size())
 
 
 @pytest.mark.parametrize("variant1", [basix.LagrangeVariant.equispaced, basix.LagrangeVariant.gll_warped,
@@ -50,7 +50,7 @@ def test_different_variant_interpolation(cell_type, order, variant1, variant2):
     lower_element = basix.create_element(basix.ElementFamily.P, cell_type, order, variant1)
     higher_element = basix.create_element(basix.ElementFamily.P, cell_type, order, variant2)
 
-    run_test(lower_element, higher_element, order, lower_element.value_size)
+    run_test(lower_element, higher_element, order, lower_element.value_size())
 
 
 @pytest.mark.parametrize("family, args", [
@@ -66,7 +66,7 @@ def test_different_order_interpolation_vector(family, args, cell_type, orders):
     lower_element = basix.create_element(family, cell_type, orders[0], *args)
     higher_element = basix.create_element(family, cell_type, orders[1], *args)
 
-    run_test(lower_element, higher_element, orders[0] - 1, lower_element.value_size)
+    run_test(lower_element, higher_element, orders[0] - 1, lower_element.value_size())
 
 
 @pytest.mark.parametrize("family, args", [[basix.ElementFamily.Regge, tuple()]])
@@ -76,7 +76,7 @@ def test_different_order_interpolation_matrix(family, args, cell_type, orders):
     lower_element = basix.create_element(family, cell_type, orders[0], *args)
     higher_element = basix.create_element(family, cell_type, orders[1], *args)
 
-    run_test(lower_element, higher_element, orders[0] - 1, lower_element.value_size)
+    run_test(lower_element, higher_element, orders[0] - 1, lower_element.value_size())
 
 
 @pytest.mark.parametrize("family1, args1", [
@@ -98,7 +98,7 @@ def test_different_element_interpolation(family1, args1, family2, args2, cell_ty
     lower_element = basix.create_element(family1, cell_type, order, *args1)
     higher_element = basix.create_element(family2, cell_type, order, *args2)
 
-    run_test(lower_element, higher_element, order - 1, lower_element.value_size)
+    run_test(lower_element, higher_element, order - 1, lower_element.value_size())
 
 
 @pytest.mark.parametrize("cell_type", [basix.CellType.triangle, basix.CellType.tetrahedron,
@@ -110,20 +110,20 @@ def test_blocked_interpolation(cell_type, order):
     lagrange = basix.create_element(basix.ElementFamily.P, cell_type, order, basix.LagrangeVariant.gll_isaac)
 
     n_points = nedelec.points
-    if nedelec.value_size == 2:
+    if nedelec.value_size() == 2:
         n_eval = np.concatenate([n_points[:, 0] ** order, n_points[:, 1] ** order])
     else:
         n_eval = np.concatenate([n_points[:, 0] ** order, 0 * n_points[:, 0], n_points[:, 1] ** order])
     n_coeffs = nedelec.interpolation_matrix @ n_eval
 
     l_points = lagrange.points
-    if nedelec.value_size == 2:
+    if nedelec.value_size() == 2:
         values = [l_points[:, 0] ** order, l_points[:, 1] ** order]
     else:
         values = [l_points[:, 0] ** order, 0 * l_points[:, 0], l_points[:, 1] ** order]
[…]
