### 2022-01-14 a12f7b8
Fix some assertions
