+++ python/wrapper.cpp
@@ -339,13 +339,13 @@ Interface to the Basix C++ library.
                              &FiniteElement::num_entity_closure_dofs)
       .def_property_readonly("entity_closure_dofs",
                              &FiniteElement::entity_closure_dofs)
-      .def("value_size",
-           [](const FiniteElement& e)
-           {
-             return std::accumulate(e.value_shape().begin(),
-                                    e.value_shape().end(), 1,
-                                    std::multiplies<int>());
-           })
+      .def_property_readonly("value_size",
+                             [](const FiniteElement& e)
+                             {
+                               return std::accumulate(e.value_shape().begin(),
+                                                      e.value_shape().end(), 1,
+                                                      std::multiplies<int>());
+                             })
       .def_property_readonly("value_shape", &FiniteElement::value_shape)
       .def_property_readonly("family", &FiniteElement::family)
       .def_property_readonly("lagrange_variant",
