### 2022-01-15 3b9baae
Fix value shape for Lagrange
