### 2022-01-15 19decf8
Add more Python versions to CI, including Python 3.10.
