### 2022-01-15 12f1629
Remove macos testing.
Doesn't provide meaningful extra coverage and CI capacity is limited.
