### 2022-01-15 4aea240
Add more Python versions to CI, including Python 3.10. (#378)
* Add more Python versions to CI, including Python 3.10.

* Remove macos testing.

Doesn't provide meaningful extra coverage and CI capacity is limited.
