### 2022-01-16 bdcf469
Remove value_size().
Was hiding some computation
