+++ cpp/basix/finite-element.h
@@ -332,17 +332,13 @@ public:
   /// @return Polynomial degree
   int degree() const;
 
-  /// Get the element value size
-  /// This is just a convenience function returning product(value_shape)
-  /// @return Value size
-  int value_size() const;
-
-  /// Get the element value tensor shape, e.g. returning [1] for scalars.
+  /// The element value tensor shape, e.g. returning {} for scalars, {3}
+  /// for vectors in 3D, {2, 2} for a rank-2 tensor in 2D.
   /// @return Value shape
   const std::vector<int>& value_shape() const;
 
-  /// Dimension of the finite element space (number of degrees of
-  /// freedom for the element)
+  /// Dimension of the finite element space (number of
+  /// degrees-of-freedom for the element)
   /// @return Number of degrees of freedom
   int dim() const;
 
@@ -350,8 +346,8 @@ public:
   /// @return The family
   element::family family() const;
 
-  /// Get the Lagrange variant of the element, or throw an error if element has
-  /// no Lagrange variant
+  /// Get the Lagrange variant of the element.
+  /// @note Throws an error if element has no Lagrange variant
   /// @return The Lagrange variant
   element::lagrange_variant lagrange_variant() const;
 
@@ -360,8 +356,8 @@ public:
   maps::type map_type() const;
 
   /// Indicates whether this element is the discontinuous variant
-  /// @return True if this element is a discontinuous version
-  /// of the element
+  /// @return True if this element is a discontinuous version of the
+  /// element
   bool discontinuous() const;
 
   /// Indicates whether the dof transformations are all permutations
@@ -377,12 +373,12 @@ public:
   /// points that share a common Jacobian.
   /// @param[in] U The function values on the reference. The indices are
   /// [Jacobian index, point index, components].
-  /// @param[in] J The Jacobian of the mapping. The indices are [Jacobian
-  /// index, J_i, J_j].
-  /// @param[in] detJ The determinant of the Jacobian of the mapping. It has
-  /// length `J.shape(0)`
-  /// @param[in] K The inverse of the Jacobian of the mapping. The indices
-  /// are [Jacobian index, K_i, K_j].
+  /// @param[in] J The Jacobian of the mapping. The indices are
+  /// [Jacobian index, J_i, J_j].
+  /// @param[in] detJ The determinant of the Jacobian of the mapping. It
+  /// has length `J.shape(0)`
+  /// @param[in] K The inverse of the Jacobian of the mapping. The
+  /// indices are [Jacobian index, K_i, K_j].
   /// @return The function values on the cell. The indices are [Jacobian
   /// index, point index, components].
   xt::xtensor<double, 3> push_forward(const xt::xtensor<double, 3>& U,
@@ -726,19 +722,20 @@ public:
   /// coefficients = i_m * values
   /// \endcode
   ///
-  /// To interpolate into a Raviart-Thomas space, the following should be done:
+  /// To interpolate into a Raviart-Thomas space, the following should
+  /// be done:
   /// \code{.pseudo}
   /// i_m = element.interpolation_matrix()
   /// pts = element.points()
-  /// vs = element.value_size()
+  /// vs = prod(element.value_shape())
   /// values = VECTOR(pts.shape(0) * vs)
   /// FOR i, p IN ENUMERATE(pts):
   ///     values[i::pts.shape(0)] = f.evaluate_at(p)
   /// coefficients = i_m * values
   /// \endcode
   ///
-  /// To interpolate into a Lagrange space with a block size, the following
-  /// should be done:
+  /// To interpolate into a Lagrange space with a block size, the
+  /// following should be done:
   /// \code{.pseudo}
   /// i_m = element.interpolation_matrix()
[…]
