+++ python/docs.h
@@ -467,15 +467,16 @@ dict
 )";
 
 const std::string FiniteElement__get_tensor_product_representation = R"(
-Get the tensor product representation of this element, or throw an error
-if no such factorisation exists.
+Get the tensor product representation of this element, or throw an
+error if no such factorisation exists.
 
-The tensor product representation will be a vector of tuples. Each tuple
-contains a vector of finite elements, and a vector on integers. The vector
-of finite elements gives the elements on an interval that appear in the
-tensor product representation. The vector of integers gives the
-permutation between the numbering of the tensor product DOFs and the
-number of the DOFs of this Basix element.
+The tensor product representation will be a vector of tuples. Each
+tuple contains a vector of finite elements, and a vector on
+integers. The vector of finite elements gives the elements on an
+interval that appear in the tensor product representation. The
+vector of integers gives the permutation between the numbering of
+the tensor product DOFs and the number of the DOFs of this Basix
+element.
 
 Returns
 =======
