### 2022-01-16 a075526
Fix value shape for Lagrange elements (#377)
* Fix value shape for Lagrange

* Remove value_size().

Was hiding some computation

* Doc update
