+++ cpp/basix/interpolation.cpp
@@ -28,33 +28,38 @@ basix::compute_interpolation_operator(const FiniteElement& element_from,
   const std::size_t dim_from = element_from.dim();
   const std::size_t npts = tab.shape(1);
 
-  if (element_from.value_size() != element_to.value_size())
+  const std::size_t vs_from = std::accumulate(
+      element_from.value_shape().begin(), element_from.value_shape().end(), 1,
+      std::multiplies<int>());
+  const std::size_t vs_to = std::accumulate(element_to.value_shape().begin(),
+                                            element_to.value_shape().end(), 1,
+                                            std::multiplies<int>());
+
+  if (vs_from != vs_to)
   {
-    if (element_to.value_size() == 1)
+    if (vs_to == 1)
     {
       // Map element_from's components into element_to
-      const std::size_t vs = element_from.value_size();
-      xt::xtensor<double, 2> out({dim_to * vs, dim_from});
+      xt::xtensor<double, 2> out({dim_to * vs_from, dim_from});
       out.fill(0);
-      for (std::size_t i = 0; i < vs; ++i)
+      for (std::size_t i = 0; i < vs_from; ++i)
         for (std::size_t j = 0; j < dim_to; ++j)
           for (std::size_t k = 0; k < dim_from; ++k)
             for (std::size_t l = 0; l < npts; ++l)
-              out(i + j * vs, k) += i_m(j, l) * tab(0, l, k, i);
+              out(i + j * vs_from, k) += i_m(j, l) * tab(0, l, k, i);
 
       return out;
     }
-    else if (element_from.value_size() == 1)
+    else if (vs_from == 1)
     {
       // Map duplicates of element_to to components of element_to
-      const std::size_t vs = element_to.value_size();
-      xt::xtensor<double, 2> out({dim_to, dim_from * vs});
+      xt::xtensor<double, 2> out({dim_to, dim_from * vs_to});
       out.fill(0);
-      for (std::size_t i = 0; i < vs; ++i)
+      for (std::size_t i = 0; i < vs_to; ++i)
         for (std::size_t j = 0; j < dim_from; ++j)
           for (std::size_t k = 0; k < dim_to; ++k)
             for (std::size_t l = 0; l < npts; ++l)
-              out(k, i + j * vs) += i_m(k, i * npts + l) * tab(0, l, j, 0);
+              out(k, i + j * vs_to) += i_m(k, i * npts + l) * tab(0, l, j, 0);
 
       return out;
     }
@@ -66,12 +71,11 @@ basix::compute_interpolation_operator(const FiniteElement& element_from,
   }
   else
   {
-    const std::size_t vs = element_from.value_size();
     xt::xtensor<double, 2> out({dim_to, dim_from});
     out.fill(0);
     for (std::size_t i = 0; i < dim_to; ++i)
       for (std::size_t j = 0; j < dim_from; ++j)
-        for (std::size_t k = 0; k < vs; ++k)
+        for (std::size_t k = 0; k < vs_from; ++k)
           for (std::size_t l = 0; l < npts; ++l)
             out(i, j) += i_m(i, k * npts + l) * tab(0, l, j, k);
 
