### 2022-02-02 e6e418c
functionals must be linear
