### 2022-02-02 #3982 @mscroggs
We fixed a few typos in the paper, so I'm re-generating it
