### 2022-02-14 88ccd68
Add serendipity variants in 1D
