+++ cpp/basix/element-families.h
@@ -21,8 +21,7 @@ enum class lagrange_variant
   gl_warped = 7,
   gl_isaac = 8,
   gl_centroid = 9,
-  integral_legendre = 10,
-  integral_chebyshev = 11,
+  legendre = 10,
   vtk = 20,
 };
 
