+++ cpp/basix/e-serendipity.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "cell.h"
+#include "element-families.h"
 #include "finite-element.h"
 
 namespace basix::element
@@ -16,6 +17,17 @@ namespace basix::element
 /// discontinuous
 /// @return A finite element
 FiniteElement create_serendipity(cell::type celltype, int degree,
+                                 bool discontinuous);
+
+/// Create a serendipity element on cell with given degree
+/// @param[in] celltype The cell type
+/// @param[in] degree The degree of the element
+/// @param[in] variant The variant of the element to be created
+/// @param[in] discontinuous Controls whether the element is continuous or
+/// discontinuous
+/// @return A finite element
+FiniteElement create_serendipity(cell::type celltype, int degree,
+                                 element::lagrange_variant variant,
                                  bool discontinuous);
 
 /// Create a serendipity H(div) element on cell with given degree
