+++ cpp/basix/polynomials.h
@@ -19,6 +19,15 @@ enum class type
   legendre_bubble = 2,
 };
 
+/// Tabulate the bubble polynomial on a given cell
+///
+/// @param[in] celltype Cell type
+/// @param[in] pts Points at which to evaluate the basis. The shape is
+/// (number of points, geometric dimension).
+/// @return The value of the bubble polynomial at each point.
+xt::xtensor<double, 1> tabulate_bubble(cell::type celltype,
+                                       const xt::xarray<double>& pts);
+
 /// Tabulate a set of polynomials.
 ///
 /// @param[in] polytype Polynomial type

[…]
