### 2022-02-14 01d6783
Remove unused code
