+++ cpp/basix/e-serendipity.cpp
@@ -544,13 +544,27 @@ xt::xtensor<double, 2> make_serendipity_curl_space_3d(int degree)
 } // namespace
 
 //----------------------------------------------------------------------------
-FiniteElement basix::element::create_serendipity(cell::type celltype,
-                                                 int degree, bool discontinuous)
+FiniteElement
+basix::element::create_serendipity(cell::type celltype, int degree,
+                                   element::lagrange_variant lvariant,
+                                   bool discontinuous)
 {
   if (celltype != cell::type::interval and celltype != cell::type::quadrilateral
       and celltype != cell::type::hexahedron)
   {
     throw std::runtime_error("Invalid celltype");
+  }
+
+  // If degree < 3, variant makes no difference, so set it to equispaced so
+  // element signatures match
+
+  if (lvariant == element::lagrange_variant::unset)
+  {
+    if (degree < 3)
+      lvariant = element::lagrange_variant::equispaced;
+    else
+      throw std::runtime_error(
+          "serendipity elements of degree > 2 need to be given a variant.");
   }
 
   const std::vector<std::vector<std::vector<int>>> topology
@@ -575,8 +589,8 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
   xt::xtensor<double, 3> edge_transforms, face_transforms;
   if (degree >= 2)
   {
-    FiniteElement moment_space
-        = element::create_dpc(cell::type::interval, degree - 2, true);
+    FiniteElement moment_space = element::create_lagrange(
+        cell::type::interval, degree - 2, lvariant, true);
     std::tie(x[1], M[1]) = moments::make_integral_moments(
         moment_space, celltype, 1, 2 * degree - 2);
     if (tdim > 1)
@@ -644,64 +658,12 @@ FiniteElement basix::element::create_serendipity(cell::type celltype,
         = element::make_discontinuous(x, M, entity_transformations, tdim, 1);
   }
 
-  return FiniteElement(element::family::serendipity, celltype, degree, {1},
-                       wcoeffs, entity_transformations, x, M,
-                       maps::type::identity, discontinuous, degree,
-                       degree < static_cast<int>(tdim) ? 1 : degree / tdim);
+  return FiniteElement(
+      element::family::serendipity, celltype, degree, {1}, wcoeffs,
+      entity_transformations, x, M, maps::type::identity, discontinuous, degree,
+      degree < static_cast<int>(tdim) ? 1 : degree / tdim, {}, lvariant);
 }
 //----------------------------------------------------------------------------
-FiniteElement
-basix::element::create_serendipity(cell::type celltype, int degree,
-                                   element::lagrange_variant variant,
-                                   bool discontinuous)
-{
-  if (celltype != cell::type::interval)
-  {
-    throw std::runtime_error("Invalid celltype");
-  }
-
-  const std::vector<std::vector<std::vector<int>>> topology
-      = cell::topology(celltype);
-
-  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
-  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
-
-  // dim 0 (vertices)
-  const xt::xtensor<double, 2> geometry = cell::geometry(celltype);
-  const std::size_t num_vertices = geometry.shape(0);
-  M[0] = std::vector<xt::xtensor<double, 3>>(num_vertices,
-                                             xt::ones<double>({1, 1, 1}));
-  x[0].resize(geometry.shape(0));
-  for (std::size_t i = 0; i < x[0].size(); ++i)
-  {
-    x[0][i] = xt::reshape_view(
-        xt::row(geometry, i), {static_cast<std::size_t>(1), geometry.shape(1)});
-  }
-
-  if (degree >= 2)
-  {
[…]
