### 2022-02-14 416383a
remove DPC on interval (it's the same as Lagrange)
