### 2022-02-14 a0f1aef
give variant to serendipity in tests
