### 2022-02-15 b06080a
Add variants of serendipity in 1D (#382)
* Add serendipity variants in 1D

* Remove unused code

* Tidying

* remove comment

* remove DPC on interval (it's the same as Lagrange)

* give variant to serendipity in tests

* docs

* ffcx branch
