+++ python/wrapper.cpp
@@ -102,8 +102,7 @@ Interface to the Basix C++ library.
       .value("centroid", lattice::simplex_method::centroid);
 
   py::enum_<polynomials::type>(m, "PolynomialType")
-      .value("legendre", polynomials::type::legendre)
-      .value("chebyshev", polynomials::type::chebyshev);
+      .value("legendre", polynomials::type::legendre);
 
   m.def(
       "tabulate_polynomials",
@@ -402,9 +401,7 @@ Interface to the Basix C++ library.
       .value("gl_warped", element::lagrange_variant::gl_warped)
       .value("gl_isaac", element::lagrange_variant::gl_isaac)
       .value("gl_centroid", element::lagrange_variant::gl_centroid)
-      .value("integral_legendre", element::lagrange_variant::integral_legendre)
-      .value("integral_chebyshev",
-             element::lagrange_variant::integral_chebyshev)
+      .value("legendre", element::lagrange_variant::legendre)
       .value("vtk", element::lagrange_variant::vtk);
 
   // Create FiniteElement
