+++ cpp/basix/e-lagrange.cpp
@@ -876,41 +876,11 @@ FiniteElement create_vtk_element(cell::type celltype, int degree,
                        maps::type::identity, discontinuous, degree, degree);
 }
 //-----------------------------------------------------------------------------
-int get_integral_degree(cell::type celltype, int degree)
+FiniteElement create_legendre(cell::type celltype, int degree,
+                              bool discontinuous)
 {
-  switch (celltype)
-  {
-  case cell::type::interval:
-    return degree - 2;
-  case cell::type::quadrilateral:
-    return degree - 2;
-  case cell::type::hexahedron:
-    return degree - 2;
-  case cell::type::triangle:
-    return degree - 3;
-  case cell::type::tetrahedron:
-    return degree - 4;
-  default:
-    throw std::runtime_error("Unsupported cell type");
-  }
-}
-//-----------------------------------------------------------------------------
-FiniteElement create_integral_lagrange(cell::type celltype, int degree,
-                                       element::lagrange_variant variant,
-                                       bool discontinuous)
-{
-  if (celltype == cell::type::prism or celltype == cell::type::pyramid)
-    throw std::runtime_error(
-        "Integral Lagrange not suppored on pyramids and prisms.");
-
-  polynomials::type polytype;
-  if (variant == element::lagrange_variant::integral_legendre)
-    polytype = polynomials::type::legendre;
-  else if (variant == element::lagrange_variant::integral_chebyshev)
-    polytype = polynomials::type::chebyshev;
-  else
-    throw std::runtime_error(
-        "Unknown integral variant or integral variant not yet implemented. R");
+  if (!discontinuous)
+    throw std::runtime_error("Legendre variant must be discontinuous");
 
   const std::size_t tdim = cell::topological_dimension(celltype);
   const std::size_t ndofs = polyset::dim(celltype, degree);
@@ -921,86 +891,38 @@ FiniteElement create_integral_lagrange(cell::type celltype, int degree,
   std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
   std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
 
-  // Create points at nodes, ordered by topology (vertices first)
-  if (degree == 0)
+  auto [pts, _wts] = quadrature::make_quadrature(quadrature::type::Default,
+                                                 celltype, degree * 2);
+  auto wts = xt::adapt(_wts);
+
+  // Evaluate moment space at quadrature points
+  const xt::xtensor<double, 2> phi = polynomials::tabulate(
+      polynomials::type::legendre, celltype, degree, pts);
+
+  for (std::size_t dim = 0; dim <= tdim; ++dim)
   {
-    if (!discontinuous)
+    M[dim].resize(topology[dim].size());
+    x[dim].resize(topology[dim].size());
+    if (dim < tdim)
     {
-      throw std::runtime_error(
-          "Cannot create a continuous order 0 Lagrange basis function");
-    }
-    x[tdim] = {lattice::create(celltype, 0, lattice::type::equispaced, true)};
-    const std::size_t num_dofs = x[tdim][0].shape(0);
-    std::array<std::size_t, 3> s = {num_dofs, 1, num_dofs};
-    M[tdim].push_back(xt::xtensor<double, 3>(s));
-    xt::view(M[tdim][0], xt::all(), 0, xt::all()) = xt::eye<double>(num_dofs);
-  }
-  else
-  {
-    for (std::size_t dim = 0; dim <= tdim; ++dim)
-    {
-      M[dim].resize(topology[dim].size());
-      x[dim].resize(topology[dim].size());
-
-      if (dim == 0)
[…]
