+++ python/docs.h
@@ -504,7 +504,8 @@ basix.finite_element.FiniteElement
     A finite element
 )";
 
-const std::string create_element__family_cell_degree_variant_discontinuous = R"(
+const std::string create_element__family_cell_degree_lvariant_discontinuous
+    = R"(
 Create an element using a given Lagrange variant
 
 Parameters
@@ -515,7 +516,7 @@ cell : basix.CellType
     The reference cell type that the element is defined on
 degree : int
     The degree of the element
-variant : basix.LagrangeVariant
+lvariant : basix.LagrangeVariant
     The variant of Lagrange to use
 discontinuous : bool
     Indicates whether the element is discontinuous between cells points of the element. The discontinuous element will have the same DOFs, but they will all be associated with the interior of the cell.
@@ -526,7 +527,7 @@ basix.finite_element.FiniteElement
     A finite element
 )";
 
-const std::string create_element__family_cell_degree_variant = R"(
+const std::string create_element__family_cell_degree_lvariant = R"(
 Create a continuous element using a given Lagrange variant
 
 Parameters
@@ -537,8 +538,51 @@ cell : basix.CellType
     The reference cell type that the element is defined on
 degree : int
     The degree of the element
-variant : basix.LagrangeVariant
+lvariant : basix.LagrangeVariant
     The variant of Lagrange to use
+
+Returns
+=======
+basix.finite_element.FiniteElement
+    A finite element
+)";
+
+const std::string create_element__family_cell_degree_dvariant_discontinuous
+    = R"(
+Create an element using a given DPC variant
+
+Parameters
+==========
+family : basix.ElementFamily
+    The element family
+cell : basix.CellType
+    The reference cell type that the element is defined on
+degree : int
+    The degree of the element
+dvariant : basix.DPCVariant
+    The variant of DPC to use
+discontinuous : bool
+    Indicates whether the element is discontinuous between cells points of the element. The discontinuous element will have the same DOFs, but they will all be associated with the interior of the cell.
+
+Returns
+=======
+basix.finite_element.FiniteElement
+    A finite element
+)";
+
+const std::string create_element__family_cell_degree_dvariant = R"(
+Create a continuous element using a given DPC variant
+
+Parameters
+==========
+family : basix.ElementFamily
+    The element family
+cell : basix.CellType
+    The reference cell type that the element is defined on
+degree : int
+    The degree of the element
+dvariant : basix.DPCVariant
+    The variant of DPC to use
 
 Returns
 =======
