+++ cpp/basix/finite-element.cpp
@@ -200,7 +200,8 @@ basix::FiniteElement basix::create_element(element::family family,
     return element::create_serendipity(
         cell, degree, element::lagrange_variant::unset, discontinuous);
   case element::family::DPC:
-    return element::create_dpc(cell, degree, discontinuous);
+    return element::create_dpc(cell, degree, element::dpc_variant::unset,
+                               discontinuous);
   default:
     throw std::runtime_error("Element family not found");
   }
@@ -208,15 +209,29 @@ basix::FiniteElement basix::create_element(element::family family,
 //-----------------------------------------------------------------------------
 basix::FiniteElement basix::create_element(element::family family,
                                            cell::type cell, int degree,
-                                           element::lagrange_variant variant,
+                                           element::dpc_variant dvariant,
+                                           bool discontinuous)
+{
+  switch (family)
+  {
+  case element::family::DPC:
+    return element::create_dpc(cell, degree, dvariant, discontinuous);
+  default:
+    throw std::runtime_error("Cannot pass a DPC variant.");
+  }
+}
+//-----------------------------------------------------------------------------
+basix::FiniteElement basix::create_element(element::family family,
+                                           cell::type cell, int degree,
+                                           element::lagrange_variant lvariant,
                                            bool discontinuous)
 {
   switch (family)
   {
   case element::family::P:
-    return element::create_lagrange(cell, degree, variant, discontinuous);
+    return element::create_lagrange(cell, degree, lvariant, discontinuous);
   case element::family::serendipity:
-    return element::create_serendipity(cell, degree, variant, discontinuous);
+    return element::create_serendipity(cell, degree, lvariant, discontinuous);
   default:
     throw std::runtime_error("Cannot pass a Lagrange variant.");
   }
@@ -224,9 +239,16 @@ basix::FiniteElement basix::create_element(element::family family,
 //-----------------------------------------------------------------------------
 basix::FiniteElement basix::create_element(element::family family,
                                            cell::type cell, int degree,
-                                           element::lagrange_variant variant)
+                                           element::lagrange_variant lvariant)
 {
-  return create_element(family, cell, degree, variant, false);
+  return create_element(family, cell, degree, lvariant, false);
+}
+//-----------------------------------------------------------------------------
+basix::FiniteElement basix::create_element(element::family family,
+                                           cell::type cell, int degree,
+                                           element::dpc_variant dvariant)
+{
+  return create_element(family, cell, degree, dvariant, false);
 }
 //-----------------------------------------------------------------------------
 basix::FiniteElement basix::create_element(element::family family,
@@ -305,12 +327,12 @@ FiniteElement::FiniteElement(
     int highest_complete_degree,
     std::vector<std::tuple<std::vector<FiniteElement>, std::vector<int>>>
         tensor_factors,
-    element::lagrange_variant lvariant)
+    element::lagrange_variant lvariant, element::dpc_variant dvariant)
     : _cell_type(cell_type), _cell_tdim(cell::topological_dimension(cell_type)),
       _cell_subentity_types(cell::subentity_types(cell_type)), _family(family),
-      _lagrange_variant(lvariant), _degree(degree), _map_type(map_type),
-      _entity_transformations(entity_transformations), _x(x),
-      _discontinuous(discontinuous),
+      _lagrange_variant(lvariant), _dpc_variant(dvariant), _degree(degree),
+      _map_type(map_type), _entity_transformations(entity_transformations),
+      _x(x), _discontinuous(discontinuous),
       _degree_bounds({highest_complete_degree, highest_degree}),
       _tensor_factors(tensor_factors)
 {
[…]
