+++ test/test_lagrange.py
@@ -438,3 +438,19 @@ def test_legendre(celltype, degree, variant):
     e = basix.create_element(basix.ElementFamily.P, celltype, degree, variant, True)
     for p in e.points:
         assert in_cell(celltype, p)
+
+
+@pytest.mark.parametrize("variant", [
+    basix.DPCVariant.equispaced_triangle,
+    basix.DPCVariant.stretched_equispaced_triangle,
+    basix.DPCVariant.diagonal_equispaced,
+    basix.DPCVariant.diagonal_gll,
+])
+@pytest.mark.parametrize("celltype", [
+    basix.CellType.quadrilateral, basix.CellType.hexahedron,
+])
+@pytest.mark.parametrize("degree", range(1, 5))
+def test_dpc(celltype, degree, variant):
+    e = basix.create_element(basix.ElementFamily.DPC, celltype, degree, variant, True)
+    for p in e.points:
+        assert in_cell(celltype, p)
