### 2022-02-15 9e2179e
DPC variants on hexes
