+++ cpp/basix/e-lagrange.cpp
@@ -928,16 +928,27 @@ FiniteElement create_legendre(cell::type celltype, int degree,
 xt::xtensor<double, 2> make_dpc_points(cell::type celltype, int degree,
                                        element::dpc_variant variant)
 {
-  if (variant == element::dpc_variant::equispaced_simplex)
+  if (degree == 0)
+    return lattice::create(celltype, 0, lattice::type::equispaced, true);
+
+  if (variant == element::dpc_variant::simplex_equispaced
+      or variant == element::dpc_variant::simplex_gll)
   {
+    lattice::type latticetype;
+    lattice::simplex_method latticesm = lattice::simplex_method::isaac;
+    if (variant == element::dpc_variant::simplex_equispaced)
+      latticetype = lattice::type::equispaced;
+    else if (variant == element::dpc_variant::simplex_gll)
+      latticetype = lattice::type::gll;
+
     switch (celltype)
     {
     case cell::type::quadrilateral:
-      return lattice::create(cell::type::triangle, degree,
-                             lattice::type::equispaced, true);
+      return lattice::create(cell::type::triangle, degree, latticetype, true,
+                             latticesm);
     case cell::type::hexahedron:
-      return lattice::create(cell::type::tetrahedron, degree,
-                             lattice::type::equispaced, true);
+      return lattice::create(cell::type::tetrahedron, degree, latticetype, true,
+                             latticesm);
     default:
       throw std::runtime_error("Invalid cell type");
     }
@@ -1013,6 +1024,7 @@ xt::xtensor<double, 2> make_dpc_points(cell::type celltype, int degree,
            or variant == element::dpc_variant::diagonal_gll)
   {
     lattice::type latticetype;
+    lattice::simplex_method latticesm = lattice::simplex_method::isaac;
     if (variant == element::dpc_variant::diagonal_equispaced)
       latticetype = lattice::type::equispaced;
     else if (variant == element::dpc_variant::diagonal_gll)
@@ -1029,19 +1041,70 @@ xt::xtensor<double, 2> make_dpc_points(cell::type celltype, int degree,
                          / (degree * degree + degree + 1);
 
       std::size_t n = 0;
-      for (int i = 0; i <= degree; ++i)
+      for (int j = 0; j <= degree; ++j)
       {
         const auto interval_pts
-            = lattice::create(cell::type::interval, i, latticetype, true);
-        const double y = gap * (i % 2 == 0 ? i / 2 : degree - (i - 1) / 2);
+            = lattice::create(cell::type::interval, j, latticetype, true);
+        const double y = gap * (j % 2 == 0 ? j / 2 : degree - (j - 1) / 2);
         const double coord0 = y < 1 ? y : y - 1;
         const double coord1 = y < 1 ? 0 : 1;
-        for (int j = 0; j <= i; ++j)
+        for (int i = 0; i <= j; ++i)
         {
-          const double x = interval_pts(j, 0);
+          const double x = interval_pts(i, 0);
           pts(n, 0) = coord0 * (1 - x) + coord1 * x;
           pts(n, 1) = coord1 * (1 - x) + coord0 * x;
           ++n;
+        }
+      }
+      return pts;
+    }
+    case cell::type::hexahedron:
+    {
+      xt::xtensor<double, 2> pts(
+          {static_cast<std::size_t>((degree + 3) * (degree + 2) * (degree + 1)
+                                    / 6),
+           3});
+
+      const double gap
+          = static_cast<double>(3 * degree) / (degree * degree + 1);
+
+      std::size_t n = 0;
+      for (int k = 0; k <= degree; ++k)
+      {
[…]
