+++ cpp/basix/e-serendipity.cpp
@@ -602,7 +602,7 @@ basix::element::create_serendipity(cell::type celltype, int degree,
     // TODO: DPC variant
     FiniteElement moment_space
         = element::create_dpc(cell::type::quadrilateral, degree - 4,
-                              element::dpc_variant::equispaced_simplex, true);
+                              element::dpc_variant::simplex_equispaced, true);
     std::tie(x[2], M[2]) = moments::make_integral_moments(
         moment_space, celltype, 1, 2 * degree - 4);
     if (tdim > 2)
@@ -617,7 +617,7 @@ basix::element::create_serendipity(cell::type celltype, int degree,
     // TODO: DPC variant
     std::tie(x[3], M[3]) = moments::make_integral_moments(
         element::create_dpc(cell::type::hexahedron, degree - 6,
-                            element::dpc_variant::equispaced_simplex, true),
+                            element::dpc_variant::simplex_equispaced, true),
         celltype, 1, 2 * degree - 6);
   }
 
@@ -692,7 +692,7 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
             ? element::create_lagrange(
                 facettype, degree, element::lagrange_variant::equispaced, true)
             : element::create_dpc(facettype, degree,
-                                  element::dpc_variant::equispaced_simplex,
+                                  element::dpc_variant::simplex_equispaced,
                                   true);
   std::tie(x[tdim - 1], M[tdim - 1]) = moments::make_normal_integral_moments(
       facet_moment_space, celltype, tdim, 2 * degree);
@@ -706,7 +706,7 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
   {
     // TODO: DPC variant
     FiniteElement cell_moment_space = element::create_dpc(
-        celltype, degree - 2, element::dpc_variant::equispaced_simplex, true);
+        celltype, degree - 2, element::dpc_variant::simplex_equispaced, true);
     std::tie(x[tdim], M[tdim]) = moments::make_integral_moments(
         cell_moment_space, celltype, tdim, 2 * degree - 2);
   }
@@ -790,7 +790,7 @@ FiniteElement basix::element::create_serendipity_curl(cell::type celltype,
     // TODO: DPC variant
     FiniteElement moment_space
         = element::create_dpc(cell::type::quadrilateral, degree - 2,
-                              element::dpc_variant::equispaced_simplex, true);
+                              element::dpc_variant::simplex_equispaced, true);
     std::tie(x[2], M[2]) = moments::make_integral_moments(
         moment_space, celltype, tdim, 2 * degree - 2);
     if (tdim == 3)
@@ -803,7 +803,7 @@ FiniteElement basix::element::create_serendipity_curl(cell::type celltype,
         // TODO: DPC variant
         std::tie(x[3], M[3]) = moments::make_integral_moments(
             element::create_dpc(cell::type::hexahedron, degree - 4,
-                                element::dpc_variant::equispaced_simplex, true),
+                                element::dpc_variant::simplex_equispaced, true),
             celltype, tdim, 2 * degree - 4);
       }
     }
