+++ .github/workflows/dolfin-tests.yml
@@ -37,7 +37,7 @@ jobs:
       - name: Install FEniCS Python components
         run: |
           python3 -m pip install git+https://github.com/FEniCS/ufl.git
-          python3 -m pip install git+https://github.com/FEniCS/ffcx.git@mscroggs/serendipity-variants-new
+          python3 -m pip install git+https://github.com/FEniCS/ffcx.git
 
       - name: Run cpp demos
         run: |
