### 2022-02-15 3cda4c8
Add variants of DPC spaces (#384)
* ffcx branch

* Add DPC variant option to constuctors

* working on DPC variants

* DPC variants on hexes

* ffc branch

* ]

* flake

* docs clang format

* dolfinx branch

* DPC not on interval
