### 2022-02-15 fca0584
Pass both variants when creating serendipity
