+++ cpp/basix/finite-element.h
@@ -940,6 +940,20 @@ FiniteElement create_element(element::family family, cell::type cell,
                              int degree, element::lagrange_variant lvariant,
                              bool discontinuous);
 
+/// Create an element using a given Lagrange variant and a given DPC variant
+/// @param[in] family The element family
+/// @param[in] cell The reference cell type that the element is defined on
+/// @param[in] degree The degree of the element
+/// @param[in] lvariant The variant of Lagrange to use
+/// @param[in] dvariant The variant of DPC to use
+/// @param[in] discontinuous Indicates whether the element is discontinuous
+/// between cells points of the element. The discontinuous element will have the
+/// same DOFs, but they will all be associated with the interior of the cell.
+/// @return A finite element
+FiniteElement create_element(element::family family, cell::type cell,
+                             int degree, element::lagrange_variant lvariant,
+                             element::dpc_variant dvariant, bool discontinuous);
+
 /// Create an element using a given DPC variant
 /// @param[in] family The element family
 /// @param[in] cell The reference cell type that the element is defined on
@@ -975,6 +989,19 @@ FiniteElement create_element(element::family family, cell::type cell,
 /// @return A finite element
 FiniteElement create_element(element::family family, cell::type cell,
                              int degree, element::lagrange_variant lvariant);
+
+/// Create a continuous element using a given Lagrange variant and a given DPC
+/// variant
+/// @param[in] family The element family
+/// @param[in] cell The reference cell type that the element is defined
+/// on
+/// @param[in] degree The degree of the element
+/// @param[in] lvariant The variant of Lagrange to use
+/// @param[in] dvariant The variant of DPC to use
+/// @return A finite element
+FiniteElement create_element(element::family family, cell::type cell,
+                             int degree, element::lagrange_variant lvariant,
+                             element::dpc_variant dvariant);
 
 /// Create a continuous element using a given DPC variant
 /// @param[in] family The element family
