### 2022-02-21 67ea55c
Remove -DDEBUG definition that doesn't do anything.
The standard is to use -DNDEBUG for non-debug builds
