+++ .github/workflows/build-wheels.yml
@@ -45,7 +45,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          ref: ${{ inputs.basix_ref }}
+          ref: ${{ github.event.inputs.basix_ref }}
 
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.1.2
@@ -61,7 +61,7 @@ jobs:
     steps:
       - uses: actions/checkout@v2
         with:
-          ref: ${{ inputs.basix_ref }}
+          ref: ${{ github.event.inputs.basix_ref }}
 
       - name: Install skbuild
         run: python -m pip install scikit-build
@@ -84,14 +84,14 @@ jobs:
           path: dist
 
       - uses: pypa/gh-action-pypi-publish@v1.4.2
-        if: ${{ inputs.test_pypi_publish }}
+        if: ${{ github.event.inputs.test_pypi_publish }}
         with:
           user: __token__
           password: ${{ secrets.PYPI_TOKEN }}
           repository_url: https://pypi.org/legacy/
 
       - uses: pypa/gh-action-pypi-publish@v1.4.2
-        if: ${{ inputs.pypi_publish }}
+        if: ${{ github.event.inputs.pypi_publish }}
         with:
           user: __token__
           password: ${{ secrets.PYPI_TEST_TOKEN }}
