+++ pyproject.toml
@@ -1,14 +1,15 @@
 [build-system]
-requires = ["setuptools>=42", "wheel", "pybind11>=2.6.4", "cmake>=3.16", "scikit-build>=0.12"]
+requires = ["setuptools>=42", "wheel", "pybind11>=2.9.1", "cmake>=3.16", "scikit-build>=0.12"]
 
 build-backend = "setuptools.build_meta"
 
 [tool.cibuildwheel]
-build = ["cp{37,38,39}-manylinux_x86_64", "cp{37,38,39}-macosx_x86_64"]
+build = ["cp{37,38,39}-manylinux_x86_64", "cp{37,38,39}-macosx_x86_64", "cp{38,39}-macosx_arm64"]
 
 test-command = "cmake -DPython3_EXECUTABLE=$(which python) -G Ninja -B build -S {project}/test/test_cmake && cmake --build build/ && build/a.out && python -m pytest -v -n auto --durations 20 {project}/test/"
 test-requires = ["pytest-xdist"]
 test-extras = ["test"]
+test-skip = "*-macosx_arm64"
 
 manylinux-x86_64-image = "manylinux2014"
 
