### 2022-02-28 29a6c9d
Store variant for discontinuous Lagrange (#391)
