### 2022-03-02 375e868
Convert variant names to lowercase to allow eg "GLL" to be used
