### 2022-03-03 8984211
Remove memory assignment in polyset on triangle
