### 2022-03-03 38f82ea
Remove memory from tetrahedron
