### 2022-03-03 51c6eff
remove memory assignment for pyramid
