### 2022-03-04 9156021
Remove memory assignment for quads
