### 2022-03-04 0647001
Remove memory assigment for hexes
