### 2022-03-07 6441032
Start removing memory from prism polyset
