+++ cpp/basix/polyset.cpp
@@ -924,6 +924,12 @@ void tabulate_polyset_hex_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   }
 }
 //-----------------------------------------------------------------------------
+std::size_t prism_idp(std::size_t px, std::size_t py, std::size_t pz,
+                      std::size_t n)
+{
+  return (n + 1) * (py * n + 1 - (py - 2) * (py - 1) / 2) + (n + 1) * px + pz;
+}
+//-----------------------------------------------------------------------------
 void tabulate_polyset_prism_derivs(xt::xtensor<double, 3>& P, std::size_t n,
                                    std::size_t nderiv,
                                    const xt::xtensor<double, 2>& x)
@@ -939,31 +945,165 @@ void tabulate_polyset_prism_derivs(xt::xtensor<double, 3>& P, std::size_t n,
       {static_cast<std::size_t>(polyset::nderivs(cell::type::triangle, nderiv)),
        x01.shape(0),
        static_cast<std::size_t>(polyset::dim(cell::type::triangle, n))});
-  tabulate_polyset_triangle_derivs(pxy, n, nderiv, x01);
-  // FIXME: remove this memory assignment
-  xt::xtensor<double, 3> pz(
-      {static_cast<std::size_t>(nderiv + 1), x2.shape(0), n + 1});
-  tabulate_polyset_line_derivs(pz, n, nderiv, x2);
+  // tabulate_polyset_triangle_derivs(pxy, n, nderiv, x01);
+
+  assert(x01.shape(1) == 2);
+
+  const auto X = x01 * 2.0 - 1.0;
+  auto X0 = xt::col(X, 0);
+  auto X1 = xt::col(X, 1);
+
+  // f3 = ((1 - y) / 2)^2
+  const auto f3 = xt::square(1.0 - X1) * 0.25;
+
+  // Iterate over derivatives in increasing order, since higher derivatives
+  for (std::size_t kk = 0; kk <= nderiv; ++kk)
+  {
+    for (std::size_t kx = 0; kx <= kk; ++kx)
+    {
+      const std::size_t ky = kk - kx;
+      if (kx == 0 and ky == 0)
+        xt::view(pxy, idx(kx, ky), xt::all(), 0) = 1.0;
+      else
+        xt::view(pxy, idx(kx, ky), xt::all(), 0) = 0.0;
+
+      for (std::size_t p = 1; p < n + 1; ++p)
+      {
+        auto p0 = xt::view(pxy, idx(kx, ky), xt::all(), idx(p, 0));
+        const double a
+            = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+        p0 = (X0 + 0.5 * X1 + 0.5)
+             * xt::view(pxy, idx(kx, ky), xt::all(), idx(p - 1, 0)) * a;
+        if (kx > 0)
+        {
+          auto result0
+              = xt::view(pxy, idx(kx - 1, ky), xt::all(), idx(p - 1, 0));
+          p0 += 2 * kx * a * result0;
+        }
+
+        if (ky > 0)
+        {
+          auto result0
+              = xt::view(pxy, idx(kx, ky - 1), xt::all(), idx(p - 1, 0));
+          p0 += ky * a * result0;
+        }
+
+        if (p > 1)
+        {
+          // y^2 terms
+          p0 -= f3 * xt::view(pxy, idx(kx, ky), xt::all(), idx(p - 2, 0))
+                * (a - 1.0);
+          if (ky > 0)
+          {
+            auto result0
+                = xt::view(pxy, idx(kx, ky - 1), xt::all(), idx(p - 2, 0));
[…]
