### 2022-03-08 eec4dcc
start removing the final memory assignment
