+++ cpp/basix/polyset.cpp
@@ -938,112 +938,21 @@ void tabulate_polyset_prism_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   const std::size_t m = (n + 1) * (n + 1) * (n + 2) / 2;
   const std::size_t md = (nderiv + 1) * (nderiv + 2) * (nderiv + 3) / 6;
 
-  const xt::xtensor<double, 2> x01 = xt::view(x, xt::all(), xt::range(0, 2));
-  const xt::xtensor<double, 1> x2 = xt::col(x, 2);
-  // FIXME: remove this memory assignment
-  xt::xtensor<double, 3> pxy(
-      {static_cast<std::size_t>(polyset::nderivs(cell::type::triangle, nderiv)),
-       x01.shape(0),
-       static_cast<std::size_t>(polyset::dim(cell::type::triangle, n))});
-  // tabulate_polyset_triangle_derivs(pxy, n, nderiv, x01);
-
-  assert(x01.shape(1) == 2);
-
-  const auto X = x01 * 2.0 - 1.0;
-  auto X0 = xt::col(X, 0);
-  auto X1 = xt::col(X, 1);
-
-  // f3 = ((1 - y) / 2)^2
-  const auto f3 = xt::square(1.0 - X1) * 0.25;
-
-  // Iterate over derivatives in increasing order, since higher derivatives
-  for (std::size_t kk = 0; kk <= nderiv; ++kk)
-  {
-    for (std::size_t kx = 0; kx <= kk; ++kx)
-    {
-      const std::size_t ky = kk - kx;
-      if (kx == 0 and ky == 0)
-        xt::view(pxy, idx(kx, ky), xt::all(), 0) = 1.0;
-      else
-        xt::view(pxy, idx(kx, ky), xt::all(), 0) = 0.0;
-
-      for (std::size_t p = 1; p < n + 1; ++p)
-      {
-        auto p0 = xt::view(pxy, idx(kx, ky), xt::all(), idx(p, 0));
-        const double a
-            = static_cast<double>(2 * p - 1) / static_cast<double>(p);
-        p0 = (X0 + 0.5 * X1 + 0.5)
-             * xt::view(pxy, idx(kx, ky), xt::all(), idx(p - 1, 0)) * a;
-        if (kx > 0)
-        {
-          auto result0
-              = xt::view(pxy, idx(kx - 1, ky), xt::all(), idx(p - 1, 0));
-          p0 += 2 * kx * a * result0;
-        }
-
-        if (ky > 0)
-        {
-          auto result0
-              = xt::view(pxy, idx(kx, ky - 1), xt::all(), idx(p - 1, 0));
-          p0 += ky * a * result0;
-        }
-
-        if (p > 1)
-        {
-          // y^2 terms
-          p0 -= f3 * xt::view(pxy, idx(kx, ky), xt::all(), idx(p - 2, 0))
-                * (a - 1.0);
-          if (ky > 0)
-          {
-            auto result0
-                = xt::view(pxy, idx(kx, ky - 1), xt::all(), idx(p - 2, 0));
-            p0 -= ky * (X1 - 1.0) * result0 * (a - 1.0);
-          }
-
-          if (ky > 1)
-          {
-            auto result0
-                = xt::view(pxy, idx(kx, ky - 2), xt::all(), idx(p - 2, 0));
-            p0 -= ky * (ky - 1) * result0 * (a - 1.0);
-          }
-        }
-      }
-
-      for (std::size_t p = 0; p < n; ++p)
-      {
-        auto p0 = xt::view(pxy, idx(kx, ky), xt::all(), idx(p, 0));
[…]
