### 2022-03-08 465ce33
Finish removing memory assignments
