+++ cpp/basix/polyset.cpp
@@ -901,27 +901,14 @@ void tabulate_polyset_hex_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   }
 
   // Normalise
-  for (std::size_t kx = 0; kx <= nderiv; ++kx)
-  {
-    for (std::size_t ky = 0; ky <= nderiv - kx; ++ky)
-    {
-      for (std::size_t kz = 0; kz <= nderiv - kx - ky; ++kz)
+  for (std::size_t px = 0; px <= n; ++px)
+    for (std::size_t py = 0; py <= n; ++py)
+      for (std::size_t pz = 0; pz <= n; ++pz)
       {
-        for (std::size_t px = 0; px <= n; ++px)
-        {
-          for (std::size_t py = 0; py <= n; ++py)
-          {
-            for (std::size_t pz = 0; pz <= n; ++pz)
-            {
-              xt::view(P, idx(kx, ky, kz), xt::all(), hex_idp(px, py, pz, n))
-                  *= std::sqrt(2 * px + 1) * std::sqrt(2 * py + 1)
-                     * std::sqrt(2 * pz + 1);
-            }
-          }
-        }
+        xt::view(P, xt::all(), xt::all(), hex_idp(px, py, pz, n))
+            *= std::sqrt(2 * px + 1) * std::sqrt(2 * py + 1)
+               * std::sqrt(2 * pz + 1);
       }
-    }
-  }
 }
 //-----------------------------------------------------------------------------
 std::size_t prism_idp(std::size_t px, std::size_t py, std::size_t pz,
@@ -958,119 +945,105 @@ void tabulate_polyset_prism_derivs(xt::xtensor<double, 3>& P, std::size_t n,
     // Get reference to this derivative
     if (kz == 0)
     {
+      // Iterate over derivatives in increasing order, since higher
+      // derivatives
       for (std::size_t kx = 0; kx <= nderiv; ++kx)
+      {
         for (std::size_t ky = 0; ky <= nderiv - kx; ++ky)
-          for (std::size_t p = 0; p <= n; ++p)
-            for (std::size_t q = 0; q <= n - p; ++q)
+        {
+          if (kx == 0 and ky == 0)
+            xt::view(P, idx(kx, ky, 0), xt::all(), prism_idp(0, 0, 0, n)) = 1.0;
+          else
+            xt::view(P, idx(kx, ky, 0), xt::all(), prism_idp(0, 0, 0, n)) = 0.0;
+
+          for (std::size_t p = 1; p < n + 1; ++p)
+          {
+            auto p0
+                = xt::view(P, idx(kx, ky, 0), xt::all(), prism_idp(p, 0, 0, n));
+
+            const double a
+                = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+            p0 = (X0 + 0.5 * X1 + 0.5)
+                 * xt::view(P, idx(kx, ky, 0), xt::all(),
+                            prism_idp(p - 1, 0, 0, n))
+                 * a;
+            if (kx > 0)
             {
-              // Iterate over derivatives in increasing order, since higher
-              // derivatives
-              for (std::size_t kk = 0; kk <= nderiv; ++kk)
+              auto result0 = xt::view(P, idx(kx - 1, ky, 0), xt::all(),
+                                      prism_idp(p - 1, 0, 0, n));
+              p0 += 2 * kx * a * result0;
+            }
+
+            if (ky > 0)
+            {
+              auto result0 = xt::view(P, idx(kx, ky - 1, 0), xt::all(),
+                                      prism_idp(p - 1, 0, 0, n));
+              p0 += ky * a * result0;
+            }
+
+            if (p > 1)
+            {
+              // y^2 terms
[…]
