+++ cpp/basix/polyset.cpp
@@ -37,14 +37,9 @@ void tabulate_polyset_point_derivs(xt::xtensor<double, 3>& P, std::size_t,
   assert(P.shape(0) == nderiv + 1);
   assert(P.shape(1) == x.shape(0));
   assert(P.shape(2) == 1);
-  for (std::size_t k = 0; k <= nderiv; ++k)
-  {
-    auto result = xt::view(P, k, xt::all(), xt::all());
-    if (k == 0)
-      xt::col(result, 0) = 1.0;
-    else
-      xt::col(result, 0) = 0.0;
-  }
+
+  xt::view(P, xt::all(), xt::all(), 0) = 0.0;
+  xt::view(P, 0, xt::all(), 0) = 1.0;
 }
 //-----------------------------------------------------------------------------
 // Compute the complete set of derivatives from 0 to nderiv, for all the
@@ -62,15 +57,14 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
   assert(P.shape(0) == nderiv + 1);
   assert(P.shape(1) == x.shape(0));
   assert(P.shape(2) == m);
+
+  xt::view(P, xt::all(), xt::all(), 0) = 0.0;
+  xt::view(P, 0, xt::all(), 0) = 1.0;
+
   for (std::size_t k = 0; k <= nderiv; ++k)
   {
     // Get reference to this derivative
     auto result = xt::view(P, k, xt::all(), xt::all());
-    if (k == 0)
-      xt::col(result, 0) = 1.0;
-    else
-      xt::col(result, 0) = 0.0;
-
     auto result0 = xt::view(P, k - 1, xt::all(), xt::all());
     for (std::size_t p = 1; p <= degree; ++p)
     {
@@ -84,9 +78,8 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
   }
 
   // Normalise
-  for (std::size_t k = 0; k < nderiv + 1; ++k)
-    for (std::size_t p = 0; p <= degree; ++p)
-      xt::view(P, k, xt::all(), p) *= std::sqrt(2 * p + 1);
+  for (std::size_t p = 0; p <= degree; ++p)
+    xt::view(P, xt::all(), xt::all(), p) *= std::sqrt(2 * p + 1);
 }
 //-----------------------------------------------------------------------------
 // Compute the complete set of derivatives from 0 to nderiv, for all the
@@ -96,8 +89,8 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
 // above, but with a change of variables. The polynomials are then
 // extended in the q direction, using the relation given in Sherwin and
 // Karniadakis 1995 (https://doi.org/10.1016/0045-7825(94)00745-9).
-void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, int n,
-                                      int nderiv,
+void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
+                                      std::size_t nderiv,
                                       const xt::xtensor<double, 2>& pts)
 {
   assert(pts.shape(1) == 2);
@@ -115,18 +108,15 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, int n,
   // f3 = ((1 - y) / 2)^2
   const auto f3 = xt::square(1.0 - x1) * 0.25;
 
-  // Iterate over derivatives in increasing order, since higher derivatives
-  for (int k = 0; k <= nderiv; ++k)
-  {
[…]
