+++ cpp/basix/polyset.cpp
@@ -116,7 +116,7 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   {
     for (std::size_t ky = 0; ky <= nderiv - kx; ++ky)
     {
-      for (std::size_t p = 1; p < n + 1; ++p)
+      for (std::size_t p = 1; p <= n; ++p)
       {
         auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 0));
         const double a
@@ -397,11 +397,11 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
   }
 
   // Normalise
-  for (std::size_t p = 0; p < n + 1; ++p)
+  for (std::size_t p = 0; p <= n; ++p)
   {
-    for (std::size_t q = 0; q < n + 1 - p; ++q)
+    for (std::size_t q = 0; q <= n - p; ++q)
     {
-      for (std::size_t r = 0; r < n + 1 - p - q; ++r)
+      for (std::size_t r = 0; r <= n - p - q; ++r)
       {
         xt::view(P, xt::all(), xt::all(), idx(p, q, r))
             *= std::sqrt(2 * (p + 0.5) * (p + q + 1.0) * (p + q + r + 1.5)) * 2;
@@ -442,17 +442,17 @@ void tabulate_polyset_pyramid_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   xt::view(P, xt::all(), xt::all(), pyr_idx(0, 0, 0)) = 0.0;
   xt::view(P, idx(0, 0, 0), xt::all(), pyr_idx(0, 0, 0)) = 1.0;
 
-  for (std::size_t k = 0; k < nderiv + 1; ++k)
+  for (std::size_t k = 0; k <= nderiv; ++k)
   {
-    for (std::size_t j = 0; j < k + 1; ++j)
+    for (std::size_t j = 0; j <= k; ++j)
     {
-      for (std::size_t kx = 0; kx < j + 1; ++kx)
+      for (std::size_t kx = 0; kx <= j; ++kx)
       {
         const std::size_t ky = j - kx;
         const std::size_t kz = k - j;
 
         // r = 0
-        for (std::size_t p = 0; p < n + 1; ++p)
+        for (std::size_t p = 0; p <= n; ++p)
         {
           if (p > 0)
           {
@@ -578,7 +578,7 @@ void tabulate_polyset_pyramid_derivs(xt::xtensor<double, 3>& P, std::size_t n,
           }
         }
 
-        for (std::size_t r = 1; r < n + 1; ++r)
+        for (std::size_t r = 1; r <= n; ++r)
         {
           for (std::size_t p = 0; p < n - r; ++p)
           {
@@ -605,11 +605,11 @@ void tabulate_polyset_pyramid_derivs(xt::xtensor<double, 3>& P, std::size_t n,
     }
   }
 
-  for (std::size_t r = 0; r < n + 1; ++r)
+  for (std::size_t r = 0; r <= n; ++r)
   {
-    for (std::size_t p = 0; p < n - r + 1; ++p)
+    for (std::size_t p = 0; p <= n - r; ++p)
     {
-      for (std::size_t q = 0; q < n - r + 1; ++q)
+      for (std::size_t q = 0; q <= n - r; ++q)
       {
         xt::view(P, xt::all(), xt::all(), pyr_idx(p, q, r))
             *= std::sqrt(2 * (q + 0.5) * (p + 0.5) * (p + q + r + 1.5)) * 2;
[…]
