### 2022-03-08 0669c34
Simplify sqrts
