### 2022-03-11 66afd52
fewer memory assignment
