### 2022-03-11 0a58f1d
move first terms out of loops
