### 2022-03-11 64e61ca
no ifs in quads
