### 2022-03-11 038f833
removing ifs from loops
