+++ cpp/basix/polyset.cpp
@@ -38,7 +38,7 @@ void tabulate_polyset_point_derivs(xt::xtensor<double, 3>& P, std::size_t,
   assert(P.shape(1) == x.shape(0));
   assert(P.shape(2) == 1);
 
-  xt::view(P, xt::all(), xt::all(), 0) = 0.0;
+  std::fill(P.begin(), P.end(), 0.0);
   xt::view(P, 0, xt::all(), 0) = 1.0;
 }
 //-----------------------------------------------------------------------------
@@ -52,16 +52,28 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
                                   const xt::xtensor<double, 1>& x)
 {
   assert(x.shape(0) > 0);
-  const auto X = x * 2.0 - 1.0;
   const std::size_t m = (degree + 1);
   assert(P.shape(0) == nderiv + 1);
   assert(P.shape(1) == x.shape(0));
   assert(P.shape(2) == m);
 
-  xt::view(P, xt::all(), xt::all(), 0) = 0.0;
+  std::fill(P.begin(), P.end(), 0.0);
   xt::view(P, 0, xt::all(), 0) = 1.0;
 
-  for (std::size_t k = 0; k <= nderiv; ++k)
+  if (degree == 0)
+    return;
+
+  { // scope
+    auto result = xt::view(P, 0, xt::all(), xt::all());
+    xt::col(result, 1) = (x * 2.0 - 1.0) * xt::col(result, 0);
+    for (std::size_t p = 2; p <= degree; ++p)
+    {
+      const double a = 1.0 - 1.0 / static_cast<double>(p);
+      xt::col(result, p) = (x * 2.0 - 1.0) * xt::col(result, p - 1) * (a + 1.0)
+                           - xt::col(result, p - 2) * a;
+    }
+  }
+  for (std::size_t k = 1; k <= nderiv; ++k)
   {
     // Get reference to this derivative
     auto result = xt::view(P, k, xt::all(), xt::all());
@@ -69,11 +81,9 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t degree,
     for (std::size_t p = 1; p <= degree; ++p)
     {
       const double a = 1.0 - 1.0 / static_cast<double>(p);
-      xt::col(result, p) = X * xt::col(result, p - 1) * (a + 1.0);
-      if (k > 0)
-        xt::col(result, p) += 2 * k * xt::col(result0, p - 1) * (a + 1.0);
-      if (p > 1)
-        xt::col(result, p) -= xt::col(result, p - 2) * a;
+      xt::col(result, p) = (x * 2.0 - 1.0) * xt::col(result, p - 1) * (a + 1.0)
+                           + 2 * k * xt::col(result0, p - 1) * (a + 1.0)
+                           - xt::col(result, p - 2) * a;
     }
   }
 
@@ -760,16 +770,13 @@ void tabulate_polyset_hex_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   { return (n + 1) * (n + 1) * px + (n + 1) * py + pz; };
 
   // Compute 1D basis
-  const xt::xtensor<double, 1> x0 = xt::col(x, 0);
-  const xt::xtensor<double, 1> x1 = xt::col(x, 1);
-  const xt::xtensor<double, 1> x2 = xt::col(x, 2);
+  const auto x0 = xt::col(x, 0);
+  const auto x1 = xt::col(x, 1);
+  const auto x2 = xt::col(x, 2);
 
   assert(x0.shape(0) > 0);
   assert(x1.shape(0) > 0);
[…]
