### 2022-03-11 2be2ec6
Remove ifs for triangle polyset
