### 2022-03-11 af1982d
Revert "Remove ifs for triangle polyset"
This reverts commit 2be2ec6eeeb13ad8b2ad792c64f502459ea7976a.
