### 2022-03-11 dafeb44
simplifications
