### 2022-03-11 972a666
Revert "move ifs out of loops for hex"
This reverts commit 1718e1d80ff68bcccb3b2ee6224fef954257470e.
