### 2022-03-11 5017dea
Revert "Revert "move ifs out of loops for hex""
This reverts commit 972a6660d1ef2bb540b849e8636d58bd3634dd6b.
