+++ cpp/basix/polyset.cpp
@@ -101,11 +101,10 @@ void tabulate_polyset_line_derivs(xt::xtensor<double, 3>& P, std::size_t n,
 // Karniadakis 1995 (https://doi.org/10.1016/0045-7825(94)00745-9).
 void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
                                       std::size_t nderiv,
-                                      const xt::xtensor<double, 2>& pts)
+                                      const xt::xtensor<double, 2>& x)
 {
-  assert(pts.shape(1) == 2);
+  assert(x.shape(1) == 2);
 
-  const auto x = pts * 2.0 - 1.0;
   auto x0 = xt::col(x, 0);
   auto x1 = xt::col(x, 1);
 
@@ -116,7 +115,7 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   assert(P.shape(2) == m);
 
   // f3 = ((1 - y) / 2)^2
-  const auto f3 = xt::square(1.0 - x1) * 0.25;
+  const auto f3 = xt::square(1.0 - (x1 * 2.0 - 1.0)) * 0.25;
 
   std::fill(P.begin(), P.end(), 0.0);
   xt::view(P, idx(0, 0), xt::all(), 0) = 1.0;
@@ -137,7 +136,7 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
         auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 0));
         const double a
             = static_cast<double>(2 * p - 1) / static_cast<double>(p);
-        p0 = (x0 + 0.5 * x1 + 0.5)
+        p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
              * xt::view(P, idx(kx, ky), xt::all(), idx(p - 1, 0)) * a;
         if (kx > 0)
         {
@@ -157,7 +156,7 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
                 * (a - 1.0);
           if (ky > 0)
           {
-            p0 -= ky * (x1 - 1.0)
+            p0 -= ky * ((x1 * 2.0 - 1.0) - 1.0)
                   * xt::view(P, idx(kx, ky - 1), xt::all(), idx(p - 2, 0))
                   * (a - 1.0);
           }
@@ -175,7 +174,7 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
       {
         auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 0));
         auto p1 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 1));
-        p1 = p0 * (x1 * (1.5 + p) + 0.5 + p);
+        p1 = p0 * ((x1 * 2.0 - 1.0) * (1.5 + p) + 0.5 + p);
         if (ky > 0)
         {
           p1 += 2 * ky * (1.5 + p)
@@ -186,7 +185,8 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
         {
           const auto [a1, a2, a3] = jrc(2 * p + 1, q);
           xt::view(P, idx(kx, ky), xt::all(), idx(p, q + 1))
-              = xt::view(P, idx(kx, ky), xt::all(), idx(p, q)) * (x1 * a1 + a2)
+              = xt::view(P, idx(kx, ky), xt::all(), idx(p, q))
+                    * ((x1 * 2.0 - 1.0) * a1 + a2)
[…]
