### 2022-03-18 16dee24
Test if interpolation matrix is an identity matrix
