### 2022-03-18 defa288
Add Legendre variant of DPC
