### 2022-03-22 de5504f
Use better polysets for all serendipity type spaces
