### 2022-03-22 195ae65
Use legendre for moments of Sdiv
