+++ test/test_dof_transformations.py
@@ -114,12 +114,6 @@ def test_hexahedron_transformation_degrees(element_type, degree, element_args):
     identity = np.identity(e.dim)
     for i, degree in enumerate([2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
                                4, 2, 4, 2, 4, 2, 4, 2, 4, 2, 4, 2]):
-        if not np.allclose(
-            np.linalg.matrix_power(bt[i], degree),
-            identity
-        ):
-            for i in np.linalg.matrix_power(bt[i], degree):
-                print(list(i))
         assert np.allclose(
             np.linalg.matrix_power(bt[i], degree),
             identity)
