### 2022-03-22 bea65d7
Revert "Use legendre for moments of Sdiv"
This reverts commit 195ae65e410bf081b5ebdb0daff0d5a27c0aaf8e.
