### 2022-03-22 4734f65
legendre by default if variants not implemented yet
