### 2022-03-22 287b28e
Make serendipity family polysets more stable (#397)
* Add Legendre variant of DPC

* merge

* Use better polynet for DPC

* Use better polysets for all serendipity type spaces

* add print

* Use legendre for moments of Sdiv

* Revert "Use legendre for moments of Sdiv"

This reverts commit 195ae65e410bf081b5ebdb0daff0d5a27c0aaf8e.

* legendre by default if variants not implemented yet
