### 2022-03-30 9b970f5
Avoid C++ errors (#399)
* Add test that either elements are created or RuntimeError is hit

* Correct order 0 element creation
