+++ test/test_create.py
@@ -1,55 +1,32 @@
 import basix
-import numpy
 import pytest
 
 
-def xtest_create_simple():
-    # Creates Lagrange P1 element on triangle
-
-    # Point evaluation of polynomial set
-    degree = 1
-    points = numpy.array([[0, 0], [1, 0], [0, 1]], dtype=numpy.float64)
-    matrix = numpy.identity(points.shape[0])
-
-    # Create element from space and dual
-    coeff_space = numpy.identity(points.shape[0])
-    fe = basix.create_new_element("Custom element", "triangle", degree, [1], points, matrix, coeff_space,
-                                  [[1, 1, 1], [0, 0, 0], [0]], [numpy.identity(3) for i in range(3)],
-                                  basix.MappingType.identity)
-    numpy.set_printoptions(suppress=True, precision=2)
-    points = numpy.array([[.5, 0], [0, .5], [.5, .5]], dtype=numpy.float64)
-    print(fe.tabulate(0, points))
-
-
-def xtest_create_custom():
-    # Creates second order element on triangle
-
-    # Point evaluation of polynomial set
-    degree = 2
-    points = numpy.array([[0, .5], [0.5, 0], [0.5, 0.5], [0.25, 0.25], [0.25, 0.5], [0.5, 0.25]], dtype=numpy.float64)
-    matrix = numpy.identity(points.shape[0])
-
-    # Create element from space and dual
-    coeff_space = numpy.identity(points.shape[0])
-    fe = basix.create_new_element("Custom element", "triangle", degree, [1], points, matrix, coeff_space,
-                                  [[0, 0, 0], [1, 1, 1], [3]],
-                                  [numpy.identity(5) for i in range(3)],
-                                  basix.MappingType.identity)
-    numpy.set_printoptions(suppress=True, precision=2)
-    points = numpy.array([[.25, 0], [0, .25], [.25, .25]], dtype=numpy.float64)
-    print(fe.tabulate(0, points))
-
-
-def xtest_create_invalid():
-    degree = 2
-    # Try to create an invalid element of order 2
-    points = numpy.array([[0, 0.25], [0, 0.75], [0.25, 0.75], [0.75, 0.25],
-                          [0.25, 0.0], [0.75, 0.0]], dtype=numpy.float64)
-    matrix = numpy.identity(points.shape[0])
-
-    # Create element from space and dual
-    coeff_space = numpy.identity(points.shape[0])
-    with pytest.raises(RuntimeError):
-        basix.create_new_element("Custom element", "triangle", degree, [1], points, matrix, coeff_space,
-                                 [[0, 0, 0], [2, 2, 2], [0]],
-                                 [numpy.identity(6) for i in range(3)], basix.MappingType.identity)
+@pytest.mark.parametrize("cell", [
+    basix.CellType.interval,
+    basix.CellType.triangle,
+    basix.CellType.quadrilateral,
+    basix.CellType.tetrahedron,
+    basix.CellType.hexahedron,
+    basix.CellType.prism,
+    basix.CellType.pyramid,
+])
+@pytest.mark.parametrize("degree", range(5))
+@pytest.mark.parametrize("element, variant", [
+    (basix.ElementFamily.P, [basix.LagrangeVariant.gll_isaac]),
+    (basix.ElementFamily.RT, []),
+    (basix.ElementFamily.BDM, []),
+    (basix.ElementFamily.N1E, []),
+    (basix.ElementFamily.N2E, []),
+    (basix.ElementFamily.Regge, []),
+    (basix.ElementFamily.bubble, []),
+    (basix.ElementFamily.serendipity, [basix.LagrangeVariant.legendre, basix.DPCVariant.legendre]),
[…]
