### 2022-03-30 926e0b5
Throw RuntimeError if user tries to make element with negative degree
