### 2022-03-30 e45b4a9
Throw RuntimeError if user tries to make element with negative degree (#400)
* Throw RuntimeError if user tries to make element with negative degree

* Test that all cells and elements are included in new test

* Test creating discontinuous version too
