### 2022-03-30 46bdfed
Disable sympy tests
