+++ test/test_continuity.py
@@ -3,7 +3,8 @@ import pytest
 import numpy as np
 
 elements = [
-    (basix.ElementFamily.P, [basix.LagrangeVariant.gll_isaac]),
+    (basix.ElementFamily.P, [basix.LagrangeVariant.equispaced]),
+    (basix.ElementFamily.P, [basix.LagrangeVariant.gll_warped]),
     (basix.ElementFamily.RT, []),
     (basix.ElementFamily.BDM, []),
     (basix.ElementFamily.N1E, []),
@@ -37,9 +38,16 @@ def create_continuity_map_interval(map_type, start, end):
     raise NotImplementedError
 
 
+def create_continuity_map_triangle(map_type, v0, v1, v2):
+    if map_type == MapType.identity:
+        return lambda x: x
+
+    raise NotImplementedError
+
+
 @pytest.mark.parametrize("degree", range(1, 5))
 @pytest.mark.parametrize("element, variant", elements)
-def test_continuity_2d(degree, element, variant):
+def test_continuity_interval_facet(degree, element, variant):
     """Test that basis functions between neighbouring cells of different types will be continuous."""
     elements = {}
     for cell in [basix.CellType.triangle, basix.CellType.quadrilateral]:
@@ -48,21 +56,60 @@ def test_continuity_2d(degree, element, variant):
         except RuntimeError:
             pass
 
-    if len(elements) > 1:
-        facets = [
-            [np.array([0, 0]), np.array([1, 0]), {basix.CellType.triangle: 2, basix.CellType.quadrilateral: 0}],
-            [np.array([0, 0]), np.array([0, 1]), {basix.CellType.triangle: 1, basix.CellType.quadrilateral: 1}],
-        ]
+    if len(elements) <= 1:
+        pytest.skip()
 
-        for start, end, cellmap in facets:
-            points = np.array([start + i/10 * end for i in range(11)])
+    facets = [
+        [np.array([0, 0]), np.array([1, 0]), {basix.CellType.triangle: 2, basix.CellType.quadrilateral: 0}],
+        [np.array([0, 0]), np.array([0, 1]), {basix.CellType.triangle: 1, basix.CellType.quadrilateral: 1}],
+    ]
 
-            data = None
+    for start, end, cellmap in facets:
+        points = np.array([start + i/10 * (end - start) for i in range(11)])
 
-            for c, e in elements.items():
+        data = None
+
+        for c, e in elements.items():
+            tab = e.tabulate(0, points)[0]
+            continuity_map = create_continuity_map_interval(e.map_type, start, end)
+            entity_tab = [continuity_map(tab[:, i, :]) for i in e.entity_dofs[1][cellmap[c]]]
+            if data is None:
+                data = entity_tab
+            else:
+                assert np.allclose(data, entity_tab)
+
+
+@pytest.mark.parametrize("degree", range(1, 5))
+@pytest.mark.parametrize("element, variant", elements)
+def test_continuity_triangle_facet(degree, element, variant):
+    """Test that basis functions between neighbouring cells of different types will be continuous."""
+    elements = {}
+    for cell in [basix.CellType.tetrahedron, basix.CellType.prism, basix.CellType.pyramid]:
+        try:
+            elements[cell] = basix.create_element(element, cell, degree, *variant)
+        except RuntimeError:
+            pass
+
+    if len(elements) <= 1:
+        pytest.skip()
+
+    facets = [
+        [np.array([0, 0, 0]), np.array([1, 0, 0]), np.array([0, 1, 0]), {basix.CellType.tetrahedron: 3, basix.CellType.prism: 0}],
[…]
