+++ python/docs.h
@@ -484,6 +484,40 @@ List[Tuple[List[basix.FiniteElement], List[int]]]
     The tensor product representation
 )";
 
+const std::string create_custom_element = R"(
+Create a custom finite element
+
+Parameters
+==========
+cell_type : basix.CellType
+    The cell type
+degree : int
+    The degree of the element
+value_shape : TODO
+    The value shape of the element
+wcoeffs : TODO
+    Matrices for the kth value index containing the expansion coefficients defining a polynomial basis spanning the polynomial space for this element
+entity_transformations : TODO
+    Entity transformations representing the effect rotating and reflecting subentities of the cell has on the DOFs.
+x : TODO
+    Interpolation points. Shape is (tdim, entity index, point index, dim)
+M : TODO
+    The interpolation matrices. Indices are (tdim, entity index, dof, vs, point_index)
+map_type : basix.MapType
+    The type of map to be used to map values from the reference to a cell
+discontinuous : bool
+    Indicates whether or not this is the discontinuous version of the element
+highest_degree : int
+    The lowest degree n such that the highest degree polynomial in this element is contained in a Lagrange (or vector Lagrange) element of degree n
+highest_complete_degree : int
+    The highest degree n such that a Lagrange (or vector Lagrange) element of degree n is a subspace of this element
+
+Returns
+=======
+basix.finite_element.FiniteElement
+    A finite element
+)";
+
 const std::string create_element__family_cell_degree_discontinuous = R"(
 Create an element
 
