+++ test/test_version.py
@@ -1,3 +1,7 @@
+# Copyright (c) 2021 Chris Ricahardson and Garth Wells
+# FEniCS Project
+# SPDX-License-Identifier: MIT
+
 import pkg_resources
 import basix
 
