+++ python/wrapper.cpp
@@ -151,7 +151,7 @@ Interface to the Basix C++ library.
       basix::docstring::create_lattice__celltype_n_type_exterior_method
           .c_str());
 
-  py::enum_<maps::type>(m, "MappingType")
+  py::enum_<maps::type>(m, "MapType")
       .value("identity", maps::type::identity)
       .value("covariantPiola", maps::type::covariantPiola)
       .value("contravariantPiola", maps::type::contravariantPiola)
@@ -407,6 +407,57 @@ Interface to the Basix C++ library.
       .value("legendre", element::dpc_variant::legendre);
 
   // Create FiniteElement
+  m.def(
+      "create_custom_element",
+      [](cell::type cell_type, int degree, const std::vector<int>& value_shape,
+         const py::array_t<double, py::array::c_style>& wcoeffs,
+         const py::dict entity_transformations,
+         const std::vector<
+             std::vector<py::array_t<double, py::array::c_style>>>& x,
+         const std::vector<
+             std::vector<py::array_t<double, py::array::c_style>>>& M,
+         maps::type map_type, bool discontinuous, int highest_degree,
+         int highest_complete_degree) -> FiniteElement {
+        if (x.size() != 4)
+          throw std::runtime_error("x has the wrong size");
+        if (M.size() != 4)
+          throw std::runtime_error("M has the wrong size");
+
+        std::map<cell::type, xt::xtensor<double, 3>> _et;
+        for (auto item : entity_transformations)
+        {
+          cell::type ct = item.first.cast<cell::type>();
+          py::array_t<double, py::array::c_style> mat
+              = item.second.cast<py::array>();
+          _et[ct] = adapt_x(mat);
+        }
+
+        xt::xtensor<double, 2> _wco = adapt_x(wcoeffs);
+
+        std::array<std::vector<xt::xtensor<double, 2>>, 4> _x;
+        for (int i = 0; i < 4; ++i)
+        {
+          for (std::size_t j = 0; j < x[i].size(); ++j)
+            _x[i].push_back(adapt_x(x[i][j]));
+        }
+
+        std::array<std::vector<xt::xtensor<double, 3>>, 4> _M;
+        for (int i = 0; i < 4; ++i)
+        {
+          for (std::size_t j = 0; j < M[i].size(); ++j)
+            _M[i].push_back(adapt_x(M[i][j]));
+        }
+
+        std::vector<std::size_t> _vs(value_shape.size());
+        for (std::size_t i = 0; i < value_shape.size(); ++i)
+          _vs[i] = static_cast<std::size_t>(value_shape[i]);
+
+        return basix::create_custom_element(
+            cell_type, degree, _vs, _wco, _et, _x, _M, map_type, discontinuous,
+            highest_degree, highest_complete_degree);
+      },
+      basix::docstring::create_custom_element.c_str());
+
   m.def(
       "create_element",
       [](element::family family_name, cell::type cell_name, int degree,
