### 2022-03-31 309dcc1
Add UFL element wrapper, store data needed to recreate custom element
