+++ cpp/basix/finite-element.cpp
@@ -141,6 +141,9 @@ basix::FiniteElement basix::create_element(element::family family,
                                            cell::type cell, int degree,
                                            bool discontinuous)
 {
+  if (family == element::family::custom)
+    throw std::runtime_error("Cannot create a custom element directly. Try "
+                             "using `create_custom_element` instead");
   if (degree < 0)
     throw std::runtime_error("Cannot create an element with a negative degree");
   switch (family)
@@ -215,6 +218,9 @@ basix::FiniteElement basix::create_element(element::family family,
                                            element::dpc_variant dvariant,
                                            bool discontinuous)
 {
+  if (family == element::family::custom)
+    throw std::runtime_error("Cannot create a custom element directly. Try "
+                             "using `create_custom_element` instead");
   if (degree < 0)
     throw std::runtime_error("Cannot create an element with a negative degree");
   switch (family)
@@ -235,6 +241,9 @@ basix::FiniteElement basix::create_element(element::family family,
                                            element::lagrange_variant lvariant,
                                            bool discontinuous)
 {
+  if (family == element::family::custom)
+    throw std::runtime_error("Cannot create a custom element directly. Try "
+                             "using `create_custom_element` instead");
   if (degree < 0)
     throw std::runtime_error("Cannot create an element with a negative degree");
   switch (family)
@@ -255,6 +264,9 @@ basix::FiniteElement basix::create_element(element::family family,
                                            element::dpc_variant dvariant,
                                            bool discontinuous)
 {
+  if (family == element::family::custom)
+    throw std::runtime_error("Cannot create a custom element directly. Try "
+                             "using `create_custom_element` instead");
   if (degree < 0)
     throw std::runtime_error("Cannot create an element with a negative degree");
   switch (family)
@@ -393,10 +405,15 @@ FiniteElement::FiniteElement(
       _tensor_factors(tensor_factors)
 {
   _dual_matrix = compute_dual_matrix(cell_type, wcoeffs, M, x, degree);
-  xt::xtensor<double, 2> B_cmajor({wcoeffs.shape(0), wcoeffs.shape(1)});
-  B_cmajor.assign(wcoeffs);
+
+  if (family == element::family::custom)
+  {
+    _wcoeffs = xt::xtensor<double, 2>({wcoeffs.shape(0), wcoeffs.shape(1)});
+    _wcoeffs.assign(wcoeffs);
+    _M = M;
+  }
   // Compute C = (BD^T)^{-1} B
-  auto result = math::solve(_dual_matrix, B_cmajor);
+  auto result = math::solve(_dual_matrix, wcoeffs);
 
   _coeffs = xt::xtensor<double, 2>({result.shape(0), result.shape(1)});
   _coeffs.assign(result);
@@ -652,13 +669,13 @@ FiniteElement::tabulate_shape(std::size_t nd, std::size_t num_points) const
 xt::xtensor<double, 4>
 FiniteElement::tabulate(int nd, const xt::xarray<double>& x) const
 {
-  xt::xarray<double> _x = x;
-  if (_x.dimension() == 1)
-    _x.reshape({_x.shape(0), 1});
+  xt::xarray<double> __x = x;
+  if (__x.dimension() == 1)
+    __x.reshape({__x.shape(0), 1});
 
   auto shape = tabulate_shape(nd, x.shape(0));
   xt::xtensor<double, 4> data(shape);
-  tabulate(nd, _x, data);
+  tabulate(nd, __x, data);
 
   return data;
 }
[…]
