### 2022-04-01 b204720
Allow user to create a custom element (#402)
* Allow user to create custom element

* Add test of custom element on quadrilateral

* Update docs

* typo

* Update docs
