### 2022-04-04 48f44f4
entity transformation computation for quads
