+++ cpp/basix/dof-transformations.cpp
@@ -126,6 +126,93 @@ compute_entity_transformations_triangle(
   return out;
 }
 //-----------------------------------------------------------------------------
+std::map<cell::type, xt::xtensor<double, 3>>
+compute_entity_transformations_quadrilateral(
+    const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
+    const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
+    const xt::xtensor<double, 2>& coeffs, const int degree, const int vs,
+    maps::type map_type)
+{
+  std::map<cell::type, xt::xtensor<double, 3>> out;
+  if (x[1].size() == 0 or x[1][0].shape(0) == 0)
+  {
+    out[cell::type::interval] = xt::xtensor<double, 3>({1, 0, 0});
+    return out;
+  }
+
+  const xt::xtensor<double, 2>& pts = x[1][0];
+  const xt::xtensor<double, 3>& imat = M[1][0];
+  const std::size_t ndofs = imat.shape(0);
+  const std::size_t npts = pts.shape(0);
+  const int psize = polyset::dim(cell::type::quadrilateral, degree);
+
+  std::size_t ptstart = 0;
+  for (std::size_t i = 0; i < x[0].size(); ++i)
+    ptstart += x[0][i].shape(0);
+
+  std::size_t dofstart = 0;
+  for (std::size_t i = 0; i < M[0].size(); ++i)
+    dofstart += M[0][i].shape(0);
+
+  std::size_t total_ndofs = 0;
+  for (int d = 0; d <= 3; ++d)
+    for (std::size_t i = 0; i < M[d].size(); ++i)
+      total_ndofs += M[d][i].shape(0);
+
+  // Map the points to reverse the edge, then tabulate at those points
+  xt::xtensor<double, 2> mapped_pts(pts.shape());
+  xt::col(mapped_pts, 0) = 1. - xt::col(pts, 0);
+  xt::col(mapped_pts, 1) = xt::col(pts, 1);
+
+  xt::xtensor<double, 2> polyset_vals = xt::view(
+      polyset::tabulate(cell::type::quadrilateral, degree, 0, mapped_pts), 0,
+      xt::all(), xt::all());
+  xt::xtensor<double, 3> tabulated_data(
+      {npts, total_ndofs, static_cast<std::size_t>(vs)});
+
+  for (int j = 0; j < vs; ++j)
+  {
+    auto data_view = xt::view(tabulated_data, xt::all(), xt::all(), j);
+    xt::xtensor<double, 2> C
+        = xt::view(coeffs, xt::all(), xt::range(psize * j, psize * j + psize));
+    auto result = math::dot(polyset_vals, xt::transpose(C));
+    data_view.assign(result);
+  }
+
+  // Pull back
+  const xt::xtensor<double, 2> J = {{-1., 0.}, {0., 1.}};
+  const double detJ = -1.;
+  const xt::xtensor<double, 2> K = {{-1., 0.}, {0., 1.}};
+  xt::xtensor<double, 3> pulled_data(tabulated_data.shape());
+
+  xt::xtensor<double, 2> temp_data(
+      {pulled_data.shape(1), pulled_data.shape(2)});
+  for (std::size_t i = 0; i < npts; ++i)
+  {
+    pull_back(map_type, temp_data,
+              xt::view(tabulated_data, i, xt::all(), xt::all()), J, detJ, K);
+    auto pview = xt::view(pulled_data, i, xt::all(), xt::all());
+    pview.assign(temp_data);
+  }
+
+  // Interpolate to calculate coefficients
+  xt::xtensor<double, 3> dof_data = xt::view(
[…]
