+++ cpp/basix/dof-transformations.cpp
@@ -9,6 +9,12 @@
 #include <xtensor/xview.hpp>
 
 using namespace basix;
+typedef std::map<
+    cell::type,
+    std::vector<std::tuple<
+        std::function<xt::xtensor<double, 1>(const xt::xtensor<double, 1>&)>,
+        xt::xtensor<double, 2>, double, xt::xtensor<double, 2>>>>
+    mapinfo_t;
 
 namespace
 {
@@ -39,6 +45,45 @@ void pull_back(maps::type map_type, xt::xtensor<double, 2>& u,
   }
 }
 //-----------------------------------------------------------------------------
+mapinfo_t get_mapinfo(cell::type cell_type)
+{
+  switch (cell_type)
+  {
+  case cell::type::interval:
+    return mapinfo_t();
+  case cell::type::triangle:
+  {
+    mapinfo_t mapinfo;
+    { // scope
+      const xt::xtensor<double, 2> J = {{0., 1.}, {1., 0.}};
+      const double detJ = -1.;
+      const xt::xtensor<double, 2> K = {{0., 1.}, {1., 0.}};
+      auto map = [](const xt::xtensor<double, 1>& pt) {
+        return xt::xtensor<double, 1>({pt[1], pt[0]});
+      };
+      mapinfo[cell::type::interval].push_back(std::make_tuple(map, J, detJ, K));
+    }
+    return mapinfo;
+  }
+  case cell::type::quadrilateral:
+  {
+    mapinfo_t mapinfo;
+    { // scope
+      const xt::xtensor<double, 2> J = {{-1., 0.}, {0., 1.}};
+      const double detJ = -1.;
+      const xt::xtensor<double, 2> K = {{-1., 0.}, {0., 1.}};
+      auto map = [](const xt::xtensor<double, 1>& pt) {
+        return xt::xtensor<double, 1>({1 - pt[0], pt[1]});
+      };
+      mapinfo[cell::type::interval].push_back(std::make_tuple(map, J, detJ, K));
+    }
+    return mapinfo;
+  }
+  default:
+    throw std::runtime_error("Unsupported cell type");
+  }
+}
+//-----------------------------------------------------------------------------
 xt::xtensor<double, 2> compute_transformation(
     cell::type cell_type,
     const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
@@ -52,8 +97,8 @@ xt::xtensor<double, 2> compute_transformation(
   if (x[tdim].size() == 0 or x[tdim][0].shape(0) == 0)
     return xt::xtensor<double, 2>({0, 0});
 
-  const xt::xtensor<double, 2>& pts = x[1][0];
-  const xt::xtensor<double, 3>& imat = M[1][0];
+  const xt::xtensor<double, 2>& pts = x[tdim][0];
+  const xt::xtensor<double, 3>& imat = M[tdim][0];
 
   const std::size_t ndofs = imat.shape(0);
   const std::size_t npts = pts.shape(0);
@@ -116,122 +161,6 @@ xt::xtensor<double, 2> compute_transformation(
   return transform;
 }
 //-----------------------------------------------------------------------------
-std::map<cell::type, xt::xtensor<double, 3>>
-compute_entity_transformations_triangle(
-    const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
-    const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
-    const xt::xtensor<double, 2>& coeffs, const int degree, const int vs,
-    maps::type map_type)
-{
-  std::map<cell::type, xt::xtensor<double, 3>> out;
-  if (x[1].size() == 0 or x[1][0].shape(0) == 0)
-  {
[…]
