### 2022-04-04 aa79f45
Compute transformations for prism
