+++ cpp/basix/dof-transformations.cpp
@@ -19,6 +19,18 @@ typedef std::map<
 namespace
 {
 //-----------------------------------------------------------------------------
+int find_first_subentity(cell::type cell_type, cell::type entity_type)
+{
+  const int edim = cell::topological_dimension(entity_type);
+  std::vector<cell::type> entities = cell::subentity_types(cell_type)[edim];
+  for (std::size_t i = 0; i < entities.size(); ++i)
+  {
+    if (entities[i] == entity_type)
+      return i;
+  }
+  throw std::runtime_error("Entity not found");
+}
+//-----------------------------------------------------------------------------
 void pull_back(maps::type map_type, xt::xtensor<double, 2>& u,
                const xt::xtensor<double, 2>& U, const xt::xtensor<double, 2>& J,
                const double detJ, const xt::xtensor<double, 2>& K)
@@ -128,7 +140,7 @@ mapinfo_t get_mapinfo(cell::type cell_type)
     mapinfo[cell::type::quadrilateral] = {};
     { // scope
       auto map = [](const xt::xtensor<double, 1>& pt) {
-        return xt::xtensor<double, 1>({1 - pt[0], pt[1], pt[1]});
+        return xt::xtensor<double, 1>({1 - pt[0], pt[1], pt[2]});
       };
       const xt::xtensor<double, 2> J
           = {{-1., 0., 0.}, {0., 1., 0.}, {0., 0., 1.}};
@@ -163,6 +175,71 @@ mapinfo_t get_mapinfo(cell::type cell_type)
     }
     return mapinfo;
   }
+  case cell::type::prism:
+  {
+    mapinfo_t mapinfo;
+    mapinfo[cell::type::interval] = {};
+    mapinfo[cell::type::triangle] = {};
+    mapinfo[cell::type::quadrilateral] = {};
+    { // scope
+      auto map = [](const xt::xtensor<double, 1>& pt) {
+        return xt::xtensor<double, 1>({1 - pt[0], pt[1], pt[2]});
+      };
+      const xt::xtensor<double, 2> J
+          = {{-1., 0., 0.}, {0., 1., 0.}, {0., 0., 1.}};
+      const double detJ = -1.;
+      const xt::xtensor<double, 2> K
+          = {{-1., 0., 0.}, {0., 1., 0.}, {0., 0., 1.}};
+      mapinfo[cell::type::interval].push_back(std::make_tuple(map, J, detJ, K));
+    }
+    { // scope
+      auto map = [](const xt::xtensor<double, 1>& pt) {
+        return xt::xtensor<double, 1>({1 - pt[1] - pt[0], pt[0], pt[2]});
+      };
+      const xt::xtensor<double, 2> J
+          = {{-1., -1., 0.}, {1., 0., 0.}, {0., 0., 1.}};
+      const double detJ = 1.;
+      const xt::xtensor<double, 2> K
+          = {{0., 1., 0.}, {-1., -1., 0.}, {0., 0., 1.}};
+      mapinfo[cell::type::triangle].push_back(std::make_tuple(map, J, detJ, K));
+    }
+    { // scope
+      auto map = [](const xt::xtensor<double, 1>& pt) {
+        return xt::xtensor<double, 1>({pt[1], pt[0], pt[2]});
+      };
+      const xt::xtensor<double, 2> J
+          = {{0., 1., 0.}, {1., 0., 0.}, {0., 0., 1.}};
+      const double detJ = -1.;
+      const xt::xtensor<double, 2> K
+          = {{0., 1., 0.}, {1., 0., 0.}, {0., 0., 1.}};
+      mapinfo[cell::type::triangle].push_back(std::make_tuple(map, J, detJ, K));
+    }
+    { // scope
[…]
