### 2022-04-04 3b1ae3c
remove entity transformation computations for each element
