+++ cpp/basix/e-raviart-thomas.cpp
@@ -85,8 +85,6 @@ FiniteElement basix::element::create_rt(cell::type celltype, int degree,
       facettype, degree - 1, element::lagrange_variant::equispaced, true);
   std::tie(x[tdim - 1], M[tdim - 1]) = moments::make_normal_integral_moments(
       facet_moment_space, celltype, tdim, 2 * degree - 1);
-  xt::xtensor<double, 3> facet_transforms
-      = moments::create_normal_moment_dof_transformations(facet_moment_space);
 
   // Add integral moments on interior
   if (degree > 1)
@@ -101,27 +99,13 @@ FiniteElement basix::element::create_rt(cell::type celltype, int degree,
   const std::vector<std::vector<std::vector<int>>> topology
       = cell::topology(celltype);
 
-  std::map<cell::type, xt::xtensor<double, 3>> entity_transformations;
-
-  if (tdim == 2)
-  {
-    entity_transformations[cell::type::interval] = facet_transforms;
-  }
-  else if (tdim == 3)
-  {
-    entity_transformations[cell::type::interval]
-        = xt::xtensor<double, 3>({1, 0, 0});
-    entity_transformations[cell::type::triangle] = facet_transforms;
-  }
-
   if (discontinuous)
   {
-    std::tie(x, M, entity_transformations)
-        = element::make_discontinuous(x, M, entity_transformations, tdim, tdim);
+    std::tie(x, M) = element::make_discontinuous(x, M, tdim, tdim);
   }
 
-  return FiniteElement(
-      element::family::RT, celltype, degree, {tdim}, B, entity_transformations,
-      x, M, maps::type::contravariantPiola, discontinuous, degree, degree - 1);
+  return FiniteElement(element::family::RT, celltype, degree, {tdim}, B, x, M,
+                       maps::type::contravariantPiola, discontinuous, degree,
+                       degree - 1);
 }
 //-----------------------------------------------------------------------------

[…]
