+++ .github/workflows/intel.yml
@@ -52,8 +52,15 @@ jobs:
           . /opt/intel/oneapi/setvars.sh
           pip install .[test]
 
-      - name: Run units tests
+      - name: Run units tests ?
         run: |
           pip install pytest-xdist
+
+      - name: Run units tests ??
+        run: |
+          . /opt/intel/oneapi/setvars.sh
+
+      - name: Run units tests
+        run: |
           . /opt/intel/oneapi/setvars.sh
           pytest -v -n auto --durations 20 test/
