### 2022-04-05 f76ec11
try different initialisatrion
