### 2022-04-05 d69467a
simplify intel workflow
