### 2022-04-05 6a6072a
Revert "try different initialisatrion"
This reverts commit f76ec11afafe39cb482132490aaea26cabed34a8.
