+++ .github/workflows/intel.yml
@@ -44,5 +44,4 @@ jobs:
         run: |
           pip install pytest-xdist
           . /opt/intel/oneapi/setvars.sh
-          pytest -v test/
-          # pytest -v -n auto --durations 20 test/
+          pytest -v -n auto --durations 20 test/
