### 2022-04-05 abcb56e
Restore intel branch options
