### 2022-04-05 45b640b
Throw error instead of hanging
