+++ .github/workflows/intel.yml
@@ -11,7 +11,7 @@ on:
 
 jobs:
   build:
-    name: Build and test (
+    name: Build and test
     runs-on: ubuntu-latest
 
     defaults:
