### 2022-04-05 171ac8c
Compute entity transformations algorithmically for all elements rather than implementing them for each element (#406)
* Generate dof transformations for triangles

* entity transformation computation for quads

* refactor

* refactor

* compute transformations for tets and hexes

* Compute transformations for prism

* pyramids

* docs

* remove entity transformation computations for each element

* docs

* test ci

* revert ci test changes

* try different initialisatrion

* xdist

* serial tests

* Remove prints from tests

* simplify intel workflow

* Revert "try different initialisatrion"


[…]
