+++ cpp/basix/moments.cpp
@@ -85,198 +85,6 @@ map_points(const cell::type celltype0, const cell::type celltype1, const P& x)
 } // namespace
 
 //-----------------------------------------------------------------------------
-xt::xtensor<double, 3> moments::create_dot_moment_dof_transformations(
-    const FiniteElement& moment_space)
-{
-  // This function can be dramatically simplified and made
-  // understandable by using tensors to give more logic to the objects
-
-  cell::type celltype = moment_space.cell_type();
-  if (celltype == cell::type::point)
-    return {};
-
-  xt::xarray<double> pts = moment_space.points();
-  if (pts.shape(1) == 1)
-    pts.reshape({pts.shape(0)});
-
-  const xt::xtensor<double, 2>& P = moment_space.interpolation_matrix();
-  xt::xtensor<double, 3> tpts;
-  xt::xtensor<double, 3> J, K;
-  switch (celltype)
-  {
-  case cell::type::interval:
-  {
-    tpts = xt::atleast_3d(1.0 - pts);
-    J = {{{-1.0}}};
-    K = {{{-1.0}}};
-    break;
-  }
-  case cell::type::triangle:
-  {
-    std::array<std::size_t, 3> shape = {2, pts.shape(0), pts.shape(1)};
-    tpts = xt::zeros<double>(shape);
-    for (std::size_t i = 0; i < pts.shape(0); ++i)
-    {
-      tpts(0, i, 0) = pts(i, 1);
-      tpts(0, i, 1) = 1 - pts(i, 0) - pts(i, 1);
-      tpts(1, i, 0) = pts(i, 1);
-      tpts(1, i, 1) = pts(i, 0);
-    }
-    J = {{{0.0, 1.0}, {-1.0, -1.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
-    K = {{{-1.0, -1.0}, {1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
-    break;
-  }
-  case cell::type::quadrilateral:
-  {
-    std::array<std::size_t, 3> shape0 = {2, pts.shape(0), pts.shape(1)};
-    tpts = xt::zeros<double>(shape0);
-    for (std::size_t i = 0; i < pts.shape(0); ++i)
-    {
-      tpts(0, i, 0) = pts(i, 1);
-      tpts(0, i, 1) = 1.0 - pts(i, 0);
-      tpts(1, i, 0) = pts(i, 1);
-      tpts(1, i, 1) = pts(i, 0);
-    }
-    J = {{{0.0, 1.0}, {-1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
-    K = {{{0.0, -1.0}, {1.0, 0.0}}, {{0.0, 1.0}, {1.0, 0.0}}};
-    break;
-  }
-  default:
-    throw std::runtime_error(
-        "DOF transformations only implemented for tdim <= 2.");
-  }
-
-  std::array<std::size_t, 3> shape
-      = {tpts.shape(0), (std::size_t)moment_space.dim(),
-         (std::size_t)moment_space.dim()};
-  xt::xtensor<double, 3> out = xt::zeros<double>(shape);
-  for (std::size_t i = 0; i < tpts.shape(0); ++i)
-  {
-    auto _tpoint = xt::view(tpts, i, xt::all(), xt::all());
-    xt::xtensor<double, 3> moment_space_pts = xt::view(
-        moment_space.tabulate(0, _tpoint), 0, xt::all(), xt::all(), xt::all());
-
-    // Tile the J and J^-1 for passing into the mapping function. This
-    // could be avoided with some changes to calls to map functions
-    // taking just one J and J^1
[…]
