+++ demo/python/demo_custom_element.py
@@ -1,0 +1,217 @@
+# =========================
+# Creating a custom element
+# =========================
+#
+# In Basix, it is possible to create custom finite elements. This demo describes
+# how to do this and what data you need to provide when creating a custom element.
+#
+# First, we import Basix and Numpy.
+
+import basix
+import numpy as np
+from basix import CellType, MapType, PolynomialType, LatticeType
+
+# Lagrange element with bubble
+# ============================
+#
+# As a first example, we create a degree 1 Lagrange element with a quadratic bubble
+# on a quadrilateral cell. This element will span the following set of polynomials:
+#
+# .. math::
+#    \left\{1, x, y, xy, x(1-x)y(1-y)\right\}.
+#
+# We will define the degrees of freedom (DOFs) of this element by placing a point
+# evaluation at each vertex, plus one at the midpoint of the cell.
+#
+# Polynomial coefficients
+# -----------------------
+#
+# When creating a custom element, we must input the coefficients that define
+# a basis of the set of polynomials that out element spans. In this example,
+# we will represent the 5 functions above in terms of the 9 orthogonal polynomials
+# of degree <=2 on a quadrilateral, so we create a 5 by 9 matrix.
+
+wcoeffs = np.zeros((5, 9))
+
+# The degree 3 orthonormal polynomials for a quadrilateral will have their highest
+# degree terms in the following order:
+#
+# ..math::
+#   1, y, y^2, x, xy, xy^2, x^2, x^2y, x^2y^2
+#
+# As our polynomial space contains 1, y, x and xy. The first four rows of the matrix
+# contain a single 1 for the four orthogonal polynomials with these are their highest
+# degree terms.
+
+
+wcoeffs[0, 0] = 1
+wcoeffs[1, 1] = 1
+wcoeffs[2, 3] = 1
+wcoeffs[3, 4] = 1
+
+# The final row of the matrix defines the polynomials x(1-x)y(1-y). As the polynomials
+# are orthonormal, we can represent this as
+#
+# .. math::
+#    \sum_{i=0}^9\int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y p_i(x, y),
+#
+# and so the coefficients we want to put in the final row of our matrix are:
+#
+# .. math::
+#    \int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y.
+#
+# We compute these integrals using a degree 4 quadrature rule (this is the largest degree
+# that the integrand will be, so these integrals will be exact).
+
+pts, wts = basix.make_quadrature(CellType.quadrilateral, 4)
+poly = basix.tabulate_polynomials(PolynomialType.legendre, CellType.quadrilateral, 2, pts)
+f = pts[:, 0] * (1 - pts[:, 0]) * pts[:, 1] * (1 - pts[:, 1])
+for i in range(9):
+    wcoeffs[4, i] = sum(f * poly[:, i] * wts)
+
+# Interpolation
+# -------------
+#
+# Next, we compute the points and matrices that define how functions can be interpolated
+# into this space. These are representations of the functionals that are used in the
+# Ciarlet definition of the finite element -- in this example, these are evaluations
+# at the points described above.
+#
+# First, we define the points. We create an array of points for each entity of each
+# dimension. For each vertex of the cell, we include the coordinates of that vertex.
+# For the interior of the cell, we include a point at (0.5,0.5).
+#
+# The shape of each of the point lists is (number of points, dimension).
+
+x = [[], [], [], []]
+x[0].append(np.array([[0.0, 0.0]]))
+x[0].append(np.array([[1.0, 0.0]]))
[…]
