### 2022-04-06 5f2fc8d
Update demo/python/demo_custom_element.py
Co-authored-by: Igor Baratta <igorbaratta@gmail.com>
