### 2022-04-06 4d3ecd7
Remove highest_degree as it's always equal to degree
