+++ test/test_custom_element.py
@@ -22,7 +22,7 @@ def test_lagrange_custom_triangle_degree1():
 
     element = basix.create_custom_element(
         basix.CellType.triangle, 1, [], wcoeffs,
-        x, M, basix.MapType.identity, False, 1, 1)
+        x, M, basix.MapType.identity, False, 1)
 
     points = basix.create_lattice(basix.CellType.triangle, 5, basix.LatticeType.equispaced, True)
     assert np.allclose(lagrange.tabulate(1, points), element.tabulate(1, points))
