### 2022-04-06 bb3acce
Remove `highest_degree` (#408)
* Add runtime error

* degrees

* Remove highest_degree as it's always equal to degree

* update test
