### 2022-04-06 20f16d1
remove highest_degree
