+++ test/test_custom_element.py
@@ -3,6 +3,7 @@
 # SPDX-License-Identifier: MIT
 
 import basix
+import pytest
 import numpy as np
 
 
@@ -74,3 +75,200 @@ def test_lagrange_custom_quadrilateral_degree1():
     points = basix.create_lattice(basix.CellType.quadrilateral, 5, basix.LatticeType.equispaced, True)
     assert np.allclose(lagrange.tabulate(1, points), element.tabulate(1, points))
     assert np.allclose(lagrange.base_transformations(), element.base_transformations())
+
+
+def test_raviart_thomas_triangle_degree1():
+    """Test that Raviart-Thomas element created as a custom element agrees with built-in Raviart-Thomas."""
+
+    wcoeffs = np.zeros((3, 6))
+    wcoeffs[0, 0] = 1
+    wcoeffs[1, 3] = 1
+
+    pts, wts = basix.make_quadrature(basix.CellType.triangle, 2)
+    poly = basix.tabulate_polynomials(basix.PolynomialType.legendre, basix.CellType.triangle, 1, pts)
+    for i in range(3):
+        wcoeffs[2, i] = sum(pts[:, 0] * poly[:, i] * wts)
+        wcoeffs[2, 3 + i] = sum(pts[:, 1] * poly[:, i] * wts)
+
+    pts, wts = basix.make_quadrature(basix.CellType.interval, 1)
+
+    x = [[], [], [], []]
+    for _ in range(3):
+        x[0].append(np.zeros((0, 2)))
+    x[1].append(np.array([[1 - p[0], p[0]] for p in pts]))
+    x[1].append(np.array([[0, p[0]] for p in pts]))
+    x[1].append(np.array([[p[0], 0] for p in pts]))
+    x[2].append(np.zeros((0, 2)))
+
+    M = [[], [], [], []]
+    for _ in range(3):
+        M[0].append(np.zeros((0, 2, 0)))
+    for normal in [[-1, -1], [-1, 0], [0, 1]]:
+        M[1].append(np.array([[normal[0] * wts, normal[1] * wts]]))
+    M[2].append(np.zeros((0, 2, 0)))
+
+    element = basix.create_custom_element(
+        basix.CellType.triangle, 1, [2], wcoeffs, x, M, basix.MapType.contravariantPiola, False, 0)
+
+    rt = basix.create_element(
+        basix.ElementFamily.RT, basix.CellType.triangle, 1)
+
+    points = basix.create_lattice(basix.CellType.triangle, 5, basix.LatticeType.equispaced, True)
+    assert np.allclose(rt.tabulate(1, points), element.tabulate(1, points))
+    assert np.allclose(rt.base_transformations(), element.base_transformations())
+
+
+def create_lagrange1_quad(cell_type=basix.CellType.quadrilateral, degree=1, wcoeffs=None, x=None, M=None,
+                          value_shape=None, discontinuous=False):
+    """Attempt to create a Lagrange 1 element on a quad."""
+    if wcoeffs is None:
+        wcoeffs = np.eye(4)
+    if x is None:
+        z = np.zeros((0, 2))
+        x = [[np.array([[0., 0.]]), np.array([[1., 0.]]), np.array([[0., 1.]]), np.array([[1., 1.]])],
+             [z, z, z, z], [z], []]
+    if M is None:
+        z = np.zeros((0, 1, 0))
+        M = [[np.array([[[1.]]]), np.array([[[1.]]]), np.array([[[1.]]]), np.array([[[1.]]])],
+             [z, z, z, z], [z], []]
+    if value_shape is None:
+        value_shape = []
+    basix.create_custom_element(
[…]
