### 2022-04-06 6f4f3e9
Add checks to inputs of custom element
