### 2022-04-06 e32b16c
check number of wcoeffs rows
