### 2022-04-06 7bf1d9c
Validate inputs to `create_custom_element` (#410)
* Add tests with incorrect inputs

* remove print

* Add checks to inputs of custom element

* check number of wcoeffs rows
