### 2022-04-06 55a7fc5
Add custom element demo (#407)
* Add custom element demo

* Update demo/python/demo_custom_element.py

Co-authored-by: Igor Baratta <igorbaratta@gmail.com>

* remove highest_degree

* add more explanation

Co-authored-by: Igor Baratta <igorbaratta@gmail.com>
