### 2022-04-06 fd879fa
spaces before math (#411)
