### 2022-04-06 082b501
Add inline math to demo (#412)
* spaces before math

* Add inline math to demo
