+++ cpp/basix/e-hhj.cpp
@@ -1,0 +1,166 @@
+// Copyright (c) 2022 Matthew Scroggs
+// FEniCS Project
+// SPDX-License-Identifier:    MIT
+
+#include "e-hhj.h"
+#include "e-lagrange.h"
+#include "element-families.h"
+#include "maps.h"
+#include "math.h"
+#include "polyset.h"
+#include "quadrature.h"
+#include <xtensor/xbuilder.hpp>
+#include <xtensor/xview.hpp>
+
+using namespace basix;
+
+//-----------------------------------------------------------------------------
+FiniteElement basix::element::create_hhj(cell::type celltype, int degree,
+                                         bool discontinuous)
+{
+  if (celltype != cell::type::triangle and celltype != cell::type::tetrahedron)
+    throw std::runtime_error("Unsupported celltype");
+
+  const std::size_t tdim = cell::topological_dimension(celltype);
+
+  const int nc = tdim * (tdim + 1) / 2;
+  const int basis_size = polyset::dim(celltype, degree);
+  const std::size_t ndofs = basis_size * nc;
+  const std::size_t psize = basis_size * tdim * tdim;
+
+  xt::xtensor<double, 2> wcoeffs = xt::zeros<double>({ndofs, psize});
+  int s = basis_size;
+  for (std::size_t i = 0; i < tdim; ++i)
+  {
+    for (std::size_t j = 0; j < tdim; ++j)
+    {
+      int xoff = i + tdim * j;
+      int yoff = i + j;
+      if (tdim == 3 and i > 0 and j > 0)
+        ++yoff;
+
+      xt::view(wcoeffs, xt::range(yoff * s, yoff * s + s),
+               xt::range(xoff * s, xoff * s + s))
+          = xt::eye<double>(s);
+    }
+  }
+
+  const std::vector<std::vector<std::vector<int>>> topology
+      = cell::topology(celltype);
+  const xt::xtensor<double, 2> geometry = cell::geometry(celltype);
+
+  std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
+  std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
+
+  x[0].resize(topology[0].size());
+  M[0].resize(topology[0].size());
+  for (std::size_t e = 0; e < topology[0].size(); ++e)
+  {
+    x[0][e] = xt::xtensor<double, 2>({0, tdim});
+    M[0][e] = xt::xtensor<double, 3>({0, tdim * tdim, 0});
+  }
+  // Loop over edge and higher dimension entities
+  for (std::size_t d = 1; d < topology.size(); ++d)
+  {
+    x[d].resize(topology[d].size());
+    M[d].resize(topology[d].size());
+
+    if (static_cast<std::size_t>(degree) + 1 < d)
+    {
+      for (std::size_t e = 0; e < topology[d].size(); ++e)
+      {
+        x[d][e] = xt::xtensor<double, 2>({0, tdim});
+        M[d][e] = xt::xtensor<double, 3>({0, tdim * tdim, 0});
+      }
+    }
+    else
+    {
+
+      // Loop over entities of dimension dim
+      for (std::size_t e = 0; e < topology[d].size(); ++e)
+      {
+        // Entity coordinates
+        const xt::xtensor<double, 2> entity_x
+            = cell::sub_entity_geometry(celltype, d, e);
+
+        // Tabulate points in lattice
+        cell::type ct = cell::sub_entity_type(celltype, d, e);
+
+        const std::size_t ndofs = polyset::dim(ct, degree + 1 - d);
+        const auto [_pts, wts]
+            = quadrature::make_quadrature(ct, degree + (degree + 1 - d));
+
+        xt::xarray<double> pts = _pts;
+        if (d == 1)
+          pts.reshape({pts.shape(0), 1});
[…]
