+++ cpp/basix/finite-element.cpp
@@ -73,9 +73,9 @@ constexpr int num_transformations(cell::type cell_type)
 //-----------------------------------------------------------------------------
 xt::xtensor<double, 2>
 compute_dual_matrix(cell::type cell_type, const xt::xtensor<double, 2>& B,
-                    const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
+                    const std::array<std::vector<xt::xtensor<double, 4>>, 4>& M,
                     const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
-                    int degree)
+                    int degree, int nderivs)
 {
   std::size_t num_dofs(0), vs(0);
   for (auto& Md : M)
@@ -102,28 +102,27 @@ compute_dual_matrix(cell::type cell_type, const xt::xtensor<double, 2>& B,
     {
       // Evaluate polynomial basis at x[d]
       const xt::xtensor<double, 2>& x_e = x[d][e];
-      xt::xtensor<double, 2> P;
+      xt::xtensor<double, 3> P;
       if (x_e.shape(1) == 1 and x_e.shape(0) != 0)
       {
         auto pts = xt::view(x_e, xt::all(), 0);
-        P = xt::view(polyset::tabulate(cell_type, degree, 0, pts), 0, xt::all(),
-                     xt::all());
+        P = polyset::tabulate(cell_type, degree, nderivs, pts);
       }
       else if (x_e.shape(0) != 0)
       {
-        P = xt::view(polyset::tabulate(cell_type, degree, 0, x_e), 0, xt::all(),
-                     xt::all());
+        P = polyset::tabulate(cell_type, degree, nderivs, x_e);
       }
 
       // Me: [dof, vs, point]
       const xt::xtensor<double, 3>& Me = M[d][e];
 
       // Compute dual matrix contribution
-      for (std::size_t i = 0; i < Me.shape(0); ++i)      // Dof index
-        for (std::size_t j = 0; j < Me.shape(1); ++j)    // Value index
-          for (std::size_t k = 0; k < Me.shape(2); ++k)  // Point
-            for (std::size_t l = 0; l < P.shape(1); ++l) // Polynomial term
-              D(dof_index + i, j, l) += Me(i, j, k) * P(k, l);
+      for (std::size_t i = 0; i < Me.shape(1); ++i)        // Dof index
+        for (std::size_t j = 0; j < Me.shape(2); ++j)      // Value index
+          for (std::size_t k = 0; k < Me.shape(3); ++k)    // Point
+            for (std::size_t l = 0; l < Me.shape(0); ++l)  // Derivative
+              for (std::size_t m = 0; m < P.shape(1); ++m) // Polynomial term
+                D(dof_index + i, j, m) += Me(i, j, k, l) * P(l, k, m);
 
       dof_index += M[d][e].shape(0);
     }
@@ -298,10 +297,10 @@ basix::FiniteElement basix::create_element(element::family family,
 }
 //-----------------------------------------------------------------------------
 std::tuple<std::array<std::vector<xt::xtensor<double, 2>>, 4>,
-           std::array<std::vector<xt::xtensor<double, 3>>, 4>>
+           std::array<std::vector<xt::xtensor<double, 4>>, 4>>
 basix::element::make_discontinuous(
     const std::array<std::vector<xt::xtensor<double, 2>>, 4>& x,
-    const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M, int tdim,
+    const std::array<std::vector<xt::xtensor<double, 4>>, 4>& M, int tdim,
     int value_size)
 {
   std::size_t npoints = 0;
@@ -315,7 +314,7 @@ basix::element::make_discontinuous(
     }
   }
 
[…]
