### 2022-04-08 3ac564f
avoid segfaults
