+++ test/test_create.py
@@ -62,10 +62,12 @@ def test_create_element(cell, degree, element, variant):
     print(element.name, cell.name, degree)
     try:
         basix.create_element(element, cell, degree, *variant)
-    except RuntimeError:
-        pass
+    except RuntimeError as e:
+        if len(e.args) == 0 or "dgesv" in e.args[0]:
+            raise e
 
     try:
         basix.create_element(element, cell, degree, *variant, True)
-    except RuntimeError:
-        pass
+    except RuntimeError as e:
+        if len(e.args) == 0 or "dgesv" in e.args[0]:
+            raise e
