### 2022-04-08 d814281
Correctly initialise M and x
