+++ cpp/basix/e-serendipity.cpp
@@ -534,6 +534,9 @@ FiniteElement basix::element::create_serendipity(
     cell::type celltype, int degree, element::lagrange_variant lvariant,
     element::dpc_variant dvariant, bool discontinuous)
 {
+  if (degree == 0)
+    throw std::runtime_error("Cannot create degree 0 serendipity");
+
   if (celltype != cell::type::interval and celltype != cell::type::quadrilateral
       and celltype != cell::type::hexahedron)
   {
@@ -585,20 +588,48 @@ FiniteElement basix::element::create_serendipity(
     std::tie(x[1], M[1]) = moments::make_integral_moments(
         moment_space, celltype, 1, 2 * degree - 2);
   }
-
-  if (tdim >= 2 and degree >= 4)
+  else
   {
-    FiniteElement moment_space = element::create_dpc(
-        cell::type::quadrilateral, degree - 4, dvariant, true);
-    std::tie(x[2], M[2]) = moments::make_integral_moments(
-        moment_space, celltype, 1, 2 * degree - 4);
+    x[1] = std::vector<xt::xtensor<double, 2>>(
+        topology[1].size(), xt::xtensor<double, 2>({0, tdim}));
+    M[1] = std::vector<xt::xtensor<double, 3>>(
+        topology[1].size(), xt::xtensor<double, 3>({0, 1, 0}));
   }
 
-  if (tdim == 3 and degree >= 6)
+  if (tdim >= 2)
   {
-    std::tie(x[3], M[3]) = moments::make_integral_moments(
-        element::create_dpc(cell::type::hexahedron, degree - 6, dvariant, true),
-        celltype, 1, 2 * degree - 6);
+    if (degree >= 4)
+    {
+      FiniteElement moment_space = element::create_dpc(
+          cell::type::quadrilateral, degree - 4, dvariant, true);
+      std::tie(x[2], M[2]) = moments::make_integral_moments(
+          moment_space, celltype, 1, 2 * degree - 4);
+    }
+    else
+    {
+      x[2] = std::vector<xt::xtensor<double, 2>>(
+          topology[2].size(), xt::xtensor<double, 2>({0, tdim}));
+      M[2] = std::vector<xt::xtensor<double, 3>>(
+          topology[2].size(), xt::xtensor<double, 3>({0, 1, 0}));
+    }
+  }
+
+  if (tdim == 3)
+  {
+    if (degree >= 6)
+    {
+      std::tie(x[3], M[3]) = moments::make_integral_moments(
+          element::create_dpc(cell::type::hexahedron, degree - 6, dvariant,
+                              true),
+          celltype, 1, 2 * degree - 6);
+    }
+    else
+    {
+      x[3] = std::vector<xt::xtensor<double, 2>>(
+          topology[3].size(), xt::xtensor<double, 2>({0, tdim}));
+      M[3] = std::vector<xt::xtensor<double, 3>>(
+          topology[3].size(), xt::xtensor<double, 3>({0, 1, 0}));
+    }
   }
 
   xt::xtensor<double, 2> wcoeffs;
@@ -639,6 +670,15 @@ FiniteElement basix::element::create_serendipity_div(cell::type celltype,
   std::array<std::vector<xt::xtensor<double, 3>>, 4> M;
   std::array<std::vector<xt::xtensor<double, 2>>, 4> x;
 
+  for (std::size_t i = 0; i < tdim - 1; ++i)
+  {
+    x[i] = std::vector<xt::xtensor<double, 2>>(
+        cell::num_sub_entities(celltype, i), xt::xtensor<double, 2>({0, tdim}));
+    M[i] = std::vector<xt::xtensor<double, 3>>(
[…]
