### 2022-04-08 2ad64b7
correct regge value size
