+++ cpp/basix/e-regge.cpp
@@ -62,7 +62,7 @@ create_regge_interpolation(cell::type celltype, int degree)
   x[0] = std::vector<xt::xtensor<double, 2>>(
       cell::num_sub_entities(celltype, 0), xt::xtensor<double, 2>({0, tdim}));
   M[0] = std::vector<xt::xtensor<double, 3>>(
-      cell::num_sub_entities(celltype, 0), xt::xtensor<double, 3>({0, 1, 0}));
+      cell::num_sub_entities(celltype, 0), xt::xtensor<double, 3>({0, tdim * tdim, 0}));
 
   // Loop over edge and higher dimension entities
   for (std::size_t d = 1; d < topology.size(); ++d)
