### 2022-04-08 0d556f9
Raise more useful RuntimeErrors, include empty arrays in x and M (#415)
* Don't allow cryptic runtime errors when creating an element

* Correctly initialise M and x

* can't make degree 0 serendipity elements

* correct regge value size
