+++ cpp/basix/e-hermite.cpp
@@ -21,21 +21,18 @@ using namespace basix;
 FiniteElement basix::element::create_hermite(cell::type celltype, int degree,
                                              bool discontinuous)
 {
-  switch (celltype)
-  {
-  case cell::type::triangle:
-    if (degree < 3)
-      throw std::runtime_error(
-          "Hermite element on a triangle must have degree 3");
-    break;
-  default:
+  if (celltype != cell::type::interval and celltype != cell::type::triangle
+      and celltype != cell::type::tetrahedron)
     throw std::runtime_error("Unsupported cell type");
-  }
+
+  if (degree != 3)
+    throw std::runtime_error("Hermite element must have degree 3");
 
   const std::size_t tdim = cell::topological_dimension(celltype);
   const std::size_t ndofs = polyset::dim(celltype, degree);
   const std::vector<std::vector<std::vector<int>>> topology
       = cell::topology(celltype);
+  const xt::xtensor<double, 2> geometry = cell::geometry(celltype);
 
   const std::size_t deriv_count = polyset::nderivs(celltype, 1);
 
@@ -54,24 +51,44 @@ FiniteElement basix::element::create_hermite(cell::type celltype, int degree,
     const xt::xtensor<double, 2> entity_x
         = cell::sub_entity_geometry(celltype, 0, e);
     x[0][e] = entity_x;
-    const std::array<std::size_t, 4> sh = {3, 1, 1, deriv_count};
+    const std::array<std::size_t, 4> sh = {1 + tdim, 1, 1, deriv_count};
     M[0][e] = xt::zeros<double>(sh);
     M[0][e](0, 0, 0, 0) = 1;
-    M[0][e](1, 0, 0, 1) = 1;
-    M[0][e](2, 0, 0, 2) = 1;
+    for (std::size_t d = 0; d < tdim; ++d)
+      M[0][e](d + 1, 0, 0, d + 1) = 1;
   }
 
   for (std::size_t e = 0; e < topology[1].size(); ++e)
   {
     x[1][e] = xt::xtensor<double, 2>({0, tdim});
-    const std::array<std::size_t, 4> sh = {0, 1, 0, deriv_count};
-    M[1][e] = xt::xtensor<double, 4>(sh);
+    M[1][e] = xt::xtensor<double, 4>({0, 1, 0, deriv_count});
   }
 
-  x[2][0] = {{1. / 3, 1. / 3}};
-  const std::array<std::size_t, 4> sh = {1, 1, 1, deriv_count};
-  M[2][0] = xt::zeros<double>(sh);
-  M[2][0](0, 0, 0, 0) = 1;
+  if (tdim >= 2)
+  {
+    for (std::size_t e = 0; e < topology[2].size(); ++e)
+    {
+      xt::xtensor<double, 1> midpoint = xt::zeros<double>({tdim});
+      for (auto p : topology[2][e])
+        midpoint += xt::row(geometry, p);
+      midpoint /= topology[2][e].size();
+      x[2][e] = xt::xtensor<double, 2>({1, tdim});
+      for (std::size_t i = 0; i < tdim; ++i)
+        x[2][e](0, i) = midpoint(i);
+      const std::array<std::size_t, 4> sh = {1, 1, 1, deriv_count};
+      M[2][e] = xt::zeros<double>(sh);
+      M[2][e](0, 0, 0, 0) = 1;
+    }
+  }
+
+  if (tdim == 3)
+  {
+    for (std::size_t e = 0; e < topology[2].size(); ++e)
+    {
+      x[3][0] = xt::xtensor<double, 2>({0, tdim});
[…]
