+++ test/test_hermite.py
@@ -75,7 +75,10 @@ def test_hermite_tetrahedron():
 
     # Hermite basis functions taken from Symfem
     sym_basis = [
-        2*x**3 + 13*x**2*y + 13*x**2*z - 3*x**2 + 13*x*y**2 + 33*x*y*z - 13*x*y + 13*x*z**2 - 13*x*z + 2*y**3 + 13*y**2*z - 3*y**2 + 13*y*z**2 - 13*y*z + 2*z**3 - 3*z**2 + 1,
+        (
+            2*x**3 + 13*x**2*y + 13*x**2*z - 3*x**2 + 13*x*y**2 + 33*x*y*z - 13*x*y + 13*x*z**2
+            - 13*x*z + 2*y**3 + 13*y**2*z - 3*y**2 + 13*y*z**2 - 13*y*z + 2*z**3 - 3*z**2 + 1
+        ),
         x**3 + 3*x**2*y + 3*x**2*z - 2*x**2 + 2*x*y**2 + 4*x*y*z - 3*x*y + 2*x*z**2 - 3*x*z + x,
         2*x**2*y + 3*x*y**2 + 4*x*y*z - 3*x*y + y**3 + 3*y**2*z - 2*y**2 + 2*y*z**2 - 3*y*z + y,
         2*x**2*z + 4*x*y*z + 3*x*z**2 - 3*x*z + 2*y**2*z + 3*y*z**2 - 3*y*z + z**3 - 2*z**2 + z,
