### 2022-04-11 a2e67c3
Simplify calculation of e
