### 2022-04-11 af31047
Add scaled identity map
