+++ cpp/basix/maps.h
@@ -18,11 +18,12 @@ namespace basix::maps
 /// Map type
 enum class type
 {
-  identity,
-  covariantPiola,
-  contravariantPiola,
-  doubleCovariantPiola,
-  doubleContravariantPiola,
+  identity = 0,
+  scaledIdentity = 1,
+  covariantPiola = 2,
+  contravariantPiola = 3,
+  doubleCovariantPiola = 4,
+  doubleContravariantPiola = 5,
 };
 
 namespace impl
@@ -56,6 +57,15 @@ void dot21(Vec&& r, const Mat0& A, const Mat1& B)
   }
 }
 } // namespace impl
+
+/// Scaled identity map
+template <typename O, typename P, typename Q, typename R>
+void scaled_identity(O&& r, const P& U, const Q& /*J*/, double detJ,
+                     const R& /*K*/)
+{
+  r.assign(U);
+  std::for_each(r.begin(), r.end(), [detJ](auto& ri) { ri *= detJ; });
+}
 
 /// Covariant Piola map
 template <typename O, typename P, typename Q, typename R>
