+++ cpp/basix/finite-element.h
@@ -429,20 +429,24 @@ public:
       return [](O& u, const P& U, const Q&, double, const R&) { u.assign(U); };
     case maps::type::scaledIdentity:
       return [](O& u, const P& U, const Q&, double, const R&) {
-        maps::scaledIdentity(u, U, J, detJ, J)
+        maps::scaled_identity(u, U, J, detJ, J)
       };
     case maps::type::covariantPiola:
-      return [](O& u, const P& U, const Q& J, double detJ, const R& K)
-      { maps::covariant_piola(u, U, J, detJ, K); };
+      return [](O& u, const P& U, const Q& J, double detJ, const R& K) {
+        maps::covariant_piola(u, U, J, detJ, K);
+      };
     case maps::type::contravariantPiola:
-      return [](O& u, const P& U, const Q& J, double detJ, const R& K)
-      { maps::contravariant_piola(u, U, J, detJ, K); };
+      return [](O& u, const P& U, const Q& J, double detJ, const R& K) {
+        maps::contravariant_piola(u, U, J, detJ, K);
+      };
     case maps::type::doubleCovariantPiola:
-      return [](O& u, const P& U, const Q& J, double detJ, const R& K)
-      { maps::double_covariant_piola(u, U, J, detJ, K); };
+      return [](O& u, const P& U, const Q& J, double detJ, const R& K) {
+        maps::double_covariant_piola(u, U, J, detJ, K);
+      };
     case maps::type::doubleContravariantPiola:
-      return [](O& u, const P& U, const Q& J, double detJ, const R& K)
-      { maps::double_contravariant_piola(u, U, J, detJ, K); };
+      return [](O& u, const P& U, const Q& J, double detJ, const R& K) {
+        maps::double_contravariant_piola(u, U, J, detJ, K);
+      };
     default:
       throw std::runtime_error("Map not implemented");
     }
