### 2022-04-19 11e5144
moves ifs out of loops in triangle polyset
