+++ cpp/basix/polyset.cpp
@@ -127,59 +127,250 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
   }
 
   // Iterate over derivatives in increasing order, since higher derivatives
-  for (std::size_t kx = 0; kx <= nderiv; ++kx)
+  // kx = 0 and ky = 0 and p = 1
+  { // scope
+    auto p0 = xt::view(P, idx(0, 0), xt::all(), idx(1, 0));
+    p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
+         * xt::view(P, idx(0, 0), xt::all(), idx(0, 0));
+  }
+  // kx = 0 and ky = 0 and p > 1
+  for (std::size_t p = 2; p <= n; ++p)
   {
-    for (std::size_t ky = 0; ky <= nderiv - kx; ++ky)
+    auto p0 = xt::view(P, idx(0, 0), xt::all(), idx(p, 0));
+    const double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+    p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
+             * xt::view(P, idx(0, 0), xt::all(), idx(p - 1, 0)) * a
+         - f3 * xt::view(P, idx(0, 0), xt::all(), idx(p - 2, 0)) * (a - 1.0);
+  }
+
+  for (std::size_t p = 0; p < n; ++p)
+  {
+    auto p0 = xt::view(P, idx(0, 0), xt::all(), idx(p, 0));
+    auto p1 = xt::view(P, idx(0, 0), xt::all(), idx(p, 1));
+    p1 = p0 * ((x1 * 2.0 - 1.0) * (1.5 + p) + 0.5 + p);
+    for (std::size_t q = 1; q < n - p; ++q)
     {
-      for (std::size_t p = 1; p <= n; ++p)
+      const auto [a1, a2, a3] = jrc(2 * p + 1, q);
+      xt::view(P, idx(0, 0), xt::all(), idx(p, q + 1))
+          = xt::view(P, idx(0, 0), xt::all(), idx(p, q))
+                * ((x1 * 2.0 - 1.0) * a1 + a2)
+            - xt::view(P, idx(0, 0), xt::all(), idx(p, q - 1)) * a3;
+    }
+  }
+  // kx = 0 and ky = 1
+  if (nderiv >= 1)
+  {
+    // p = 1
+    { // scope
+      auto p0 = xt::view(P, idx(0, 1), xt::all(), idx(1, 0));
+      p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
+               * xt::view(P, idx(0, 1), xt::all(), idx(0, 0))
+           + xt::view(P, idx(0, 0), xt::all(), idx(0, 0));
+    }
+    // p > 1
+    for (std::size_t p = 2; p <= n; ++p)
+    {
+      auto p0 = xt::view(P, idx(0, 1), xt::all(), idx(p, 0));
+      const double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+      p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
+               * xt::view(P, idx(0, 1), xt::all(), idx(p - 1, 0)) * a
+           + a * xt::view(P, idx(0, 0), xt::all(), idx(p - 1, 0))
+           - f3 * xt::view(P, idx(0, 1), xt::all(), idx(p - 2, 0)) * (a - 1.0)
[…]
