### 2022-04-19 3f98c09
start working on tetrahedron polyset
