### 2022-04-19 0d2a3bd
more tet polyset progress
