+++ cpp/basix/polyset.cpp
@@ -297,10 +297,96 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
   }
-  // kx = 0 and ky = 0 and kz > 0
-  for (std::size_t kz = 1; kz <= nderiv; ++kz)
+  // kx = 0 and ky = 0 and kz = 1
+  if (nderiv >= 1)
   {
-    for (std::size_t p = 1; p <= n; ++p)
+    // p = 1
+    { // scope
+      auto p00 = xt::view(P, idx(0, 0, 1), xt::all(), idx(1, 0, 0));
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(0, 0, 1), xt::all(), idx(0, 0, 0))
+            + xt::view(P, idx(0, 0, 0), xt::all(), idx(0, 0, 0));
+    }
+    // p > 1
+    for (std::size_t p = 2; p <= n; ++p)
+    {
+      auto p00 = xt::view(P, idx(0, 0, 1), xt::all(), idx(p, 0, 0));
+      double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(0, 0, 1), xt::all(), idx(p - 1, 0, 0)) * a
+            + a * xt::view(P, idx(0, 0, 0), xt::all(), idx(p - 1, 0, 0))
+            + -f2 * xt::view(P, idx(0, 0, 1), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0)
+            - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+                  * xt::view(P, idx(0, 0, 0), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0);
+    }
+
+    for (std::size_t p = 0; p < n; ++p)
+    {
+      auto p10 = xt::view(P, idx(0, 0, 1), xt::all(), idx(p, 1, 0));
+      p10 = xt::view(P, idx(0, 0, 1), xt::all(), idx(p, 0, 0))
+                * ((1.0 + (x1 * 2.0 - 1.0)) * p
+                   + (2.0 + (x1 * 2.0 - 1.0) * 3.0 + (x2 * 2.0 - 1.0)) * 0.5)
+            + xt::view(P, idx(0, 0, 0), xt::all(), idx(p, 0, 0));
+
+      for (std::size_t q = 1; q < n - p; ++q)
+      {
+        auto [aq, bq, cq] = jrc(2 * p + 1, q);
+        auto pq1 = xt::view(P, idx(0, 0, 1), xt::all(), idx(p, q + 1, 0));
+        pq1 = xt::view(P, idx(0, 0, 1), xt::all(), idx(p, q, 0))
+                  * (f3 * aq + f4 * bq)
+              - xt::view(P, idx(0, 0, 1), xt::all(), idx(p, q - 1, 0)) * f5 * cq
+              + xt::view(P, idx(0, 0, 0), xt::all(), idx(p, q, 0)) * (aq - bq)
+              + (1.0 - (x2 * 2.0 - 1.0))
+                    * xt::view(P, idx(0, 0, 0), xt::all(), idx(p, q - 1, 0))
+                    * cq;
+      }
+    }
+
+    for (std::size_t p = 0; p < n; ++p)
+    {
[…]
