+++ cpp/basix/polyset.cpp
@@ -466,7 +466,88 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
   // kx = 0 and ky > 0
   for (std::size_t ky = 1; ky <= nderiv; ++ky)
   {
-    for (std::size_t kz = 0; kz <= nderiv - ky; ++kz)
+    // kz = 0
+    { // scope
+      for (std::size_t p = 1; p <= n; ++p)
+      {
+        auto p00 = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, 0, 0));
+        double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+        p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+               + 1.0)
+                  * xt::view(P, idx(0, ky, 0), xt::all(), idx(p - 1, 0, 0)) * a
+              + ky * a
+                    * xt::view(P, idx(0, ky - 1, 0), xt::all(),
+                               idx(p - 1, 0, 0));
+
+        if (p > 1)
+        {
+          p00 += -f2 * xt::view(P, idx(0, ky, 0), xt::all(), idx(p - 2, 0, 0))
+                     * (a - 1.0)
+                 - ky * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+                       * xt::view(P, idx(0, ky - 1, 0), xt::all(),
+                                  idx(p - 2, 0, 0))
+                       * (a - 1.0);
+
+          if (ky > 1)
+          {
+            p00 -= ky * (ky - 1)
+                   * xt::view(P, idx(0, ky - 2, 0), xt::all(), idx(p - 2, 0, 0))
+                   * (a - 1.0);
+          }
+        }
+      }
+
+      for (std::size_t p = 0; p < n; ++p)
+      {
+        auto p10 = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, 1, 0));
+        p10 = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, 0, 0))
+                  * ((1.0 + (x1 * 2.0 - 1.0)) * p
+                     + (2.0 + (x1 * 2.0 - 1.0) * 3.0 + (x2 * 2.0 - 1.0)) * 0.5)
+              + 2 * ky * xt::view(P, idx(0, ky - 1, 0), xt::all(), idx(p, 0, 0))
+                    * (1.5 + p);
+
+        for (std::size_t q = 1; q < n - p; ++q)
+        {
+          auto [aq, bq, cq] = jrc(2 * p + 1, q);
+          auto pq1 = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q + 1, 0));
+          pq1 = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q, 0))
+                    * (f3 * aq + f4 * bq)
+                - xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q - 1, 0)) * f5
+                      * cq
+                + 2 * ky
+                      * xt::view(P, idx(0, ky - 1, 0), xt::all(), idx(p, q, 0))
+                      * aq;
+        }
+      }
+
+      for (std::size_t p = 0; p < n; ++p)
+      {
+        for (std::size_t q = 0; q < n - p; ++q)
+        {
+          auto pq = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q, 1));
+          pq = xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q, 0))
[…]
