+++ cpp/basix/polyset.cpp
@@ -317,7 +317,7 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
              + 1.0)
                 * xt::view(P, idx(0, 0, 1), xt::all(), idx(p - 1, 0, 0)) * a
             + a * xt::view(P, idx(0, 0, 0), xt::all(), idx(p - 1, 0, 0))
-            + -f2 * xt::view(P, idx(0, 0, 1), xt::all(), idx(p - 2, 0, 0))
+            - f2 * xt::view(P, idx(0, 0, 1), xt::all(), idx(p - 2, 0, 0))
                   * (a - 1.0)
             - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
                   * xt::view(P, idx(0, 0, 0), xt::all(), idx(p - 2, 0, 0))
@@ -395,7 +395,7 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
                 * xt::view(P, idx(0, 0, kz), xt::all(), idx(p - 1, 0, 0)) * a
             + kz * a
                   * xt::view(P, idx(0, 0, kz - 1), xt::all(), idx(p - 1, 0, 0))
-            + -f2 * xt::view(P, idx(0, 0, kz), xt::all(), idx(p - 2, 0, 0))
+            - f2 * xt::view(P, idx(0, 0, kz), xt::all(), idx(p - 2, 0, 0))
                   * (a - 1.0)
             - kz * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
                   * xt::view(P, idx(0, 0, kz - 1), xt::all(), idx(p - 2, 0, 0))
@@ -463,12 +463,296 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
   }
-  // kx = 0 and ky > 0
-  for (std::size_t ky = 1; ky <= nderiv; ++ky)
+  // kx = 0 and ky = 1
+  if (nderiv >= 1)
   {
     // kz = 0
     { // scope
-      for (std::size_t p = 1; p <= n; ++p)
+      // p = 1
+      { // scope
+        auto p00 = xt::view(P, idx(0, 1, 0), xt::all(), idx(1, 0, 0));
+        p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+               + 1.0)
+                  * xt::view(P, idx(0, 1, 0), xt::all(), idx(0, 0, 0))
+              + xt::view(P, idx(0, 0, 0), xt::all(), idx(0, 0, 0));
+      }
+      // p > 1
+      for (std::size_t p = 2; p <= n; ++p)
+      {
+        auto p00 = xt::view(P, idx(0, 1, 0), xt::all(), idx(p, 0, 0));
+        double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+        p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+               + 1.0)
+                  * xt::view(P, idx(0, 1, 0), xt::all(), idx(p - 1, 0, 0)) * a
+              + a * xt::view(P, idx(0, 0, 0), xt::all(), idx(p - 1, 0, 0))
+              - f2 * xt::view(P, idx(0, 1, 0), xt::all(), idx(p - 2, 0, 0))
+                    * (a - 1.0)
+              - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
[…]
