+++ cpp/basix/polyset.cpp
@@ -657,18 +657,15 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
               - kz * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
                     * xt::view(P, idx(0, 1, kz - 1), xt::all(),
                                idx(p - 2, 0, 0))
+                    * (a - 1.0)
+              - kz * (kz - 1)
+                    * xt::view(P, idx(0, 1, kz - 2), xt::all(),
+                               idx(p - 2, 0, 0))
+                    * (a - 1.0)
+              - 2.0 * kz
+                    * xt::view(P, idx(0, 0, kz - 1), xt::all(),
+                               idx(p - 2, 0, 0))
                     * (a - 1.0);
-
-        if (kz > 1)
-        {
-          p00 -= kz * (kz - 1)
-                 * xt::view(P, idx(0, 1, kz - 2), xt::all(), idx(p - 2, 0, 0))
-                 * (a - 1.0);
-        }
-
-        p00 -= 2.0 * kz
-               * xt::view(P, idx(0, 0, kz - 1), xt::all(), idx(p - 2, 0, 0))
-               * (a - 1.0);
       }
 
       for (std::size_t p = 0; p < n; ++p)
@@ -695,15 +692,11 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
                 + kz * (1.0 - (x2 * 2.0 - 1.0))
                       * xt::view(P, idx(0, 1, kz - 1), xt::all(),
                                  idx(p, q - 1, 0))
+                      * cq
+                - kz * (kz - 1)
+                      * xt::view(P, idx(0, 1, kz - 2), xt::all(),
+                                 idx(p, q - 1, 0))
                       * cq;
-
-          if (kz > 1)
-          {
-            // Quadratic term in z
-            pq1 -= kz * (kz - 1)
-                   * xt::view(P, idx(0, 1, kz - 2), xt::all(), idx(p, q - 1, 0))
-                   * cq;
-          }
         }
       }
 
@@ -767,14 +760,11 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
               - ky * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
                     * xt::view(P, idx(0, ky - 1, 0), xt::all(),
                                idx(p - 2, 0, 0))
+                    * (a - 1.0)
+              - ky * (ky - 1)
+                    * xt::view(P, idx(0, ky - 2, 0), xt::all(),
+                               idx(p - 2, 0, 0))
                     * (a - 1.0);
-
-        if (ky > 1)
-        {
-          p00 -= ky * (ky - 1)
-                 * xt::view(P, idx(0, ky - 2, 0), xt::all(), idx(p - 2, 0, 0))
-                 * (a - 1.0);
-        }
       }
 
       for (std::size_t p = 0; p < n; ++p)
@@ -826,8 +816,109 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
         }
       }
     }
-    // kz > 0
-    for (std::size_t kz = 1; kz <= nderiv - ky; ++kz)
+    // kz = 1
+    if (ky < nderiv)
+    {
+      // p = 1
+      { // scope
+        auto p00 = xt::view(P, idx(0, ky, 1), xt::all(), idx(1, 0, 0));
+        p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
[…]
