+++ cpp/basix/polyset.cpp
@@ -731,7 +731,7 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
   }
-  // kx = 0 and ky > 1
+  // kx = 0 and 1 < ky < nderiv
   for (std::size_t ky = 2; ky <= nderiv; ++ky)
   {
     // kz = 0
@@ -817,17 +817,15 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
     // kz = 1
-    if (ky < nderiv)
-    {
-      // p = 1
-      { // scope
-        auto p00 = xt::view(P, idx(0, ky, 1), xt::all(), idx(1, 0, 0));
-        p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
-               + 1.0)
-                  * xt::view(P, idx(0, ky, 1), xt::all(), idx(0, 0, 0))
-              + ky * xt::view(P, idx(0, ky - 1, 1), xt::all(), idx(0, 0, 0))
-              + xt::view(P, idx(0, ky, 0), xt::all(), idx(0, 0, 0));
-      }
+    // p = 1
+    { // scope
+      auto p00 = xt::view(P, idx(0, ky, 1), xt::all(), idx(1, 0, 0));
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(0, ky, 1), xt::all(), idx(0, 0, 0))
+            + ky * xt::view(P, idx(0, ky - 1, 1), xt::all(), idx(0, 0, 0))
+            + xt::view(P, idx(0, ky, 0), xt::all(), idx(0, 0, 0));
+    }
       // p > 1
       for (std::size_t p = 2; p <= n; ++p)
       {
@@ -858,7 +856,6 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
                                idx(p - 2, 0, 0))
                     * (a - 1.0);
       }
-
       for (std::size_t p = 0; p < n; ++p)
       {
         auto p10 = xt::view(P, idx(0, ky, 1), xt::all(), idx(p, 1, 0));
@@ -899,7 +896,6 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
                      * xt::view(P, idx(0, ky, 0), xt::all(), idx(p, q, 0));
         }
       }
-
       for (std::size_t p = 0; p + 1 < n; ++p)
       {
         for (std::size_t q = 0; q + 1 < n - p; ++q)
@@ -916,7 +912,6 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
           }
         }
       }
-    }
     // kz > 1
     for (std::size_t kz = 2; kz <= nderiv - ky; ++kz)
     {
@@ -1029,6 +1024,99 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
                   + 2 * kz * ar
                         * xt::view(P, idx(0, ky, kz - 1), xt::all(),
                                    idx(p, q, r));
+          }
+        }
+      }
+    }
+  }
+  // kx = 0 and ky = nderiv
+  if (nderiv >= 1)
+  {
+    // kz = 0
+    { // scope
+      // p = 1
+      { // scope
[…]
