+++ cpp/basix/polyset.cpp
@@ -1122,10 +1122,10 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
   }
-  // kx > 0
-  for (std::size_t kx = 1; kx <= nderiv; ++kx)
+  // 0 < kx < nderiv
+  for (std::size_t kx = 1; kx < nderiv; ++kx)
   {
-    // ky = 0 amd kz = 0
+    // ky = 0 and kz = 0
     // p = 1
     {
       auto p00 = xt::view(P, idx(kx, 0, 0), xt::all(), idx(1, 0, 0));
@@ -1190,9 +1190,8 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
         }
       }
     }
-    // ky = 0 amd kz = 1
-    if (kx < nderiv)
-    {
+    // ky = 0 and kz = 1
+    { // scope
       // p = 1
       for (std::size_t p = 1; p <= n; ++p)
       {
@@ -1375,12 +1374,304 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
         }
       }
     }
-    // ky > 0
-    for (std::size_t ky = 1; ky <= nderiv - kx; ++ky)
+    // ky = 1 and kz = 0
+    // p = 1
+    { // scope
+      auto p00 = xt::view(P, idx(kx, 1, 0), xt::all(), idx(1, 0, 0));
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(kx, 1, 0), xt::all(), idx(0, 0, 0))
+            + 2 * kx * xt::view(P, idx(kx - 1, 1, 0), xt::all(), idx(0, 0, 0))
+            + xt::view(P, idx(kx, 0, 0), xt::all(), idx(0, 0, 0));
+    }
+    // p > 1
+    for (std::size_t p = 2; p <= n; ++p)
     {
-      for (std::size_t kz = 0; kz <= nderiv - kx - ky; ++kz)
+      auto p00 = xt::view(P, idx(kx, 1, 0), xt::all(), idx(p, 0, 0));
+      double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(kx, 1, 0), xt::all(), idx(p - 1, 0, 0)) * a
+            + 2 * kx * a
+                  * xt::view(P, idx(kx - 1, 1, 0), xt::all(), idx(p - 1, 0, 0))
+            + a * xt::view(P, idx(kx, 0, 0), xt::all(), idx(p - 1, 0, 0))
+            - f2 * xt::view(P, idx(kx, 1, 0), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0)
+            - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+                  * xt::view(P, idx(kx, 0, 0), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0);
+    }
+
+    for (std::size_t p = 0; p < n; ++p)
+    {
+      auto p10 = xt::view(P, idx(kx, 1, 0), xt::all(), idx(p, 1, 0));
+      p10 = xt::view(P, idx(kx, 1, 0), xt::all(), idx(p, 0, 0))
[…]
