### 2022-04-19 a017c18
finish removing ifs for tets
