+++ cpp/basix/polyset.cpp
@@ -1122,8 +1122,8 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
       }
     }
   }
-  // 0 < kx < nderiv
-  for (std::size_t kx = 1; kx < nderiv; ++kx)
+  // 0 < kx < nderiv - 1
+  for (std::size_t kx = 1; kx < nderiv - 1; ++kx)
   {
     // ky = 0 and kz = 0
     // p = 1
@@ -1446,8 +1446,95 @@ void tabulate_polyset_tetrahedron_derivs(xt::xtensor<double, 3>& P,
         }
       }
     }
-    // kz > 0
-    for (std::size_t kz = 1; kz <= nderiv - kx - 1; ++kz)
+    // kz = 1
+    // p = 1
+    { // scope
+      auto p00 = xt::view(P, idx(kx, 1, 1), xt::all(), idx(1, 0, 0));
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(kx, 1, 1), xt::all(), idx(0, 0, 0))
+            + 2 * kx * xt::view(P, idx(kx - 1, 1, 1), xt::all(), idx(0, 0, 0))
+            + xt::view(P, idx(kx, 0, 1), xt::all(), idx(0, 0, 0))
+            + xt::view(P, idx(kx, 1, 0), xt::all(), idx(0, 0, 0));
+    }
+    // p > 1
+    for (std::size_t p = 2; p <= n; ++p)
+    {
+      auto p00 = xt::view(P, idx(kx, 1, 1), xt::all(), idx(p, 0, 0));
+      double a = static_cast<double>(2 * p - 1) / static_cast<double>(p);
+      p00 = ((x0 * 2.0 - 1.0) + 0.5 * ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+             + 1.0)
+                * xt::view(P, idx(kx, 1, 1), xt::all(), idx(p - 1, 0, 0)) * a
+            + 2 * kx * a
+                  * xt::view(P, idx(kx - 1, 1, 1), xt::all(), idx(p - 1, 0, 0))
+            + a * xt::view(P, idx(kx, 0, 1), xt::all(), idx(p - 1, 0, 0))
+            + a * xt::view(P, idx(kx, 1, 0), xt::all(), idx(p - 1, 0, 0))
+            - f2 * xt::view(P, idx(kx, 1, 1), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0)
+            - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+                  * xt::view(P, idx(kx, 0, 1), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0)
+            - ((x1 * 2.0 - 1.0) + (x2 * 2.0 - 1.0))
+                  * xt::view(P, idx(kx, 1, 0), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0)
+            - 2.0 * xt::view(P, idx(kx, 0, 0), xt::all(), idx(p - 2, 0, 0))
+                  * (a - 1.0);
+    }
+
+    for (std::size_t p = 0; p < n; ++p)
+    {
+      auto p10 = xt::view(P, idx(kx, 1, 1), xt::all(), idx(p, 1, 0));
+      p10 = xt::view(P, idx(kx, 1, 1), xt::all(), idx(p, 0, 0))
[…]
