### 2022-04-19 76148c3
avoid underflow
