### 2022-04-20 eeb0b67
Wrap elements to be used directly in UFL (#405)
* Allow user to create custom element

* Add test of custom element on quadrilateral

* Update docs

* typo

* Update docs

* Add UFL wrapper for element

* Add UFL element wrapper, store data needed to recreate custom element

* pydocstyle

* install ufl

* git+

* Include more in signature

* fix signature, flake

* Better signature for custom element

* ()

* curly braces

* rename variable

* docs

* more explanation
