### 2022-04-20 a13b335
Add docs to polyset
