+++ cpp/basix/polyset.h
@@ -9,9 +9,117 @@
 #include <xtensor/xtensor.hpp>
 
 /// Polynomial expansion sets
+
+/// In Basix, the bases of the polynomial sets spanned by finite elements are
+/// represented by a set of coefficients of orthonormal polynomials on the cell.
+/// By orthogonality, we mean that
 ///
-/// \section Interval
+/// \f[\int_R p_ip_j=\begin{cases}1&i=j\\0&i\not=j\end{cases}\f]
 ///
+/// ### Interval
+/// Legendre polynomials form a set of orthogonal polynomials on an interval.
+/// Legrendre polynomials are a special case of the <a
+/// href="https://en.wikipedia.org/wiki/Jacobi_polynomials">Jacobi
+/// polynomials</a>, \f$P_n^{\alpha, \beta}\f$, with the weights \f$(\alpha,
+/// \beta) = (0, 0)\f$. Using the Jacobi recurrence relation, we have
+///
+/// \f[ n P^{0,0}_n(x) = (2n - 1) x P^{0,0}_{n - 1}(x) - (n-1)
+/// P^{0,0}_{n-2}(x),\f]
+///
+/// starting with \f$P^{0,0}_0 = 1\f$. Legendre polynomials are orthogonal when
+/// integrated over the interval [-1, 1].
+///
+/// In Basix, the interval [0, 1] is used as the reference, so we apply the
+/// transformation \f$x\mapsto 2x-1\f$ to the standard Legendre polynomials. The
+/// orthogonal polynomials we use are therefore given by the recurrence relation
+///
+/// \f[n p_n(x)=(2n-1)(2x-1)p_{n-1}(x)-(n-1)p_{n-2}(x).\f]
+///
+/// For a given set of points, the recurrence relation can be used to compute
+/// the polynomial set up to arbitrary order.
+///
+/// ### Triangle (Dubiner's basis)
+/// See <a href="https://doi.org/10.1016/0045-7825(94)00745-9">Sherwin &
+/// Karniadakis (1995)</a>.
+///
+/// An orthogonal set over a triangle can be obtained with a change of variable
+/// to \f$\zeta = 2\frac{1+x}{1 - y} - 1\f$, and a modified polynomial. The
+/// polynomial basis becomes:
+///
+/// \f[ Q_{p,q} = P^{0,0}_p(\zeta) \left(\frac{1-y}{2}\right)^p\ P_q^{2p+1,
+/// 0}(y), \f]
+///
+/// with a Legendre Polynomial of \f$\zeta\f$ and a weighted Jacobi Polynomial
+/// of \f$y\f$. In order to calculate these without actually multiplying
+/// together the two polynomials directly, we can first compute \f$Q_{p, 0}\f$
+/// using a recurrence relation (since \f$P_0^{2p + 1, 0} = 1\f$):
+///
+/// \f[ p Q_{p,0} = (2p-1)(2x + y+1)Q_{p-1,0} - (p-1)(1-y)^2 Q_{p-2, 0}.\f]
+///
+/// Subsequently, we can calculate \f$Q_{p,q}\f$ by building up from
+/// \f$Q_{p,0}\f$ with another recurrence relation:
+///
+/// \f[ Q_{p,q} = (a_0 y + a_1) Q_{p, q-1} + a_2 Q_{p, q-2}, \f]
+///
+/// where \f$a_0, a_1, a_2\f$ are the coefficients in the Jacobi recurrence
+/// relation with \f$(\alpha,\beta) = (2p+1, 0)\f$, see <a
+/// href="https://en.wikipedia.org/wiki/Jacobi_polynomials#Recurrence_relations">Wikipedia</a>.
+/// Note that \f$a_2 = 0\f$ when \f$q < 2\f$.
+///
+/// ### Tetrahedron
[…]
