### 2022-04-20 96ed4b0
Add docs to polyset.h (#427)
* start on docs

* Add docs to polyset

* don't include namespace docs in docstring of first function
