### 2022-04-20 bacc1d3
Update shape of M
