+++ python/wrapper.cpp
@@ -13,15 +13,19 @@
 #include <basix/polynomials.h>
 #include <basix/polyset.h>
 #include <basix/quadrature.h>
-#include <pybind11/numpy.h>
-#include <pybind11/operators.h>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
+#include <nanobind/nanobind.h>
+#include <nanobind/tensor.h>
+// #include <pybind11/numpy.h>
+// #include <pybind11/operators.h>
+// #include <pybind11/pybind11.h>
+// #include <pybind11/stl.h>
 #include <string>
+#include <vector>
 #include <xtensor/xadapt.hpp>
+#include <xtensor/xtensor.hpp>
 #include <xtl/xspan.hpp>
 
-namespace py = pybind11;
+namespace nb = nanobind;
 using namespace basix;
 
 namespace
@@ -45,20 +49,34 @@ const std::string& cell_type_to_str(cell::type type)
   return it->second;
 }
 
-auto adapt_x(const py::array_t<double, py::array::c_style>& x)
+auto adapt_x(const nb::tensor<double>& x)
 {
+  std::size_t xsize = 1;
   std::vector<std::size_t> shape;
-  for (pybind11::ssize_t i = 0; i < x.ndim(); ++i)
+  for (std::size_t i = 0; i < x.ndim(); ++i)
+  {
     shape.push_back(x.shape(i));
-  return xt::adapt(x.data(), x.size(), xt::no_ownership(), shape);
+    xsize *= shape.back();
+  }
+  return xt::adapt(static_cast<const double*>(x.data()), xsize,
+                   xt::no_ownership(), shape);
 }
+
+// auto adapt_x(const py::array_t<double, py::array::c_style>& x)
+// {
+//   std::vector<std::size_t> shape;
+//   for (pybind11::ssize_t i = 0; i < x.ndim(); ++i)
+//     shape.push_back(x.shape(i));
+//   return xt::adapt(x.data(), x.size(), xt::no_ownership(), shape);
+// }
+
 } // namespace
 
-PYBIND11_MODULE(_basixcpp, m)
+NB_MODULE(_basixcpp, m)
 {
-  m.doc() = R"(
-Interface to the Basix C++ library.
-)";
+  //   m.doc() = R"(
+  // Interface to the Basix C++ library.
+  // )";
 
   m.attr("__version__") = basix::version();
 
@@ -68,10 +86,9 @@ Interface to the Basix C++ library.
       [](cell::type celltype)
       {
         xt::xtensor<double, 2> g = cell::geometry(celltype);
-        auto strides = g.strides();
-        for (auto& s : strides)
-          s *= sizeof(double);
-        return py::array_t<double>(g.shape(), strides, g.data());
+        std::array<std::size_t, 2> shape = {g.shape(0), g.shape(1)};
+        return nb::tensor<nb::numpy, double, nb::shape<nb::any, nb::any>>(
+            g.data(), shape.size(), shape.data());
       },
       basix::docstring::geometry.c_str());
   m.def("sub_entity_connectivity", &cell::sub_entity_connectivity,
@@ -82,45 +99,54 @@ Interface to the Basix C++ library.
       {
         xt::xtensor<double, 2> g
             = cell::sub_entity_geometry(celltype, dim, index);
-        auto strides = g.strides();
-        for (auto& s : strides)
-          s *= sizeof(double);
-        return py::array_t<double>(g.shape(), strides, g.data());
+        std::array<std::size_t, 2> shape = {g.shape(0), g.shape(1)};
+        return nb::tensor<nb::numpy, double, nb::shape<nb::any, nb::any>>(
+            g.data(), shape.size(), shape.data());
       },
[…]
