+++ python/setup.py
@@ -60,8 +60,13 @@ class CMakeBuild(build_ext):
         build_args += ['--', '-j3']
 
         env = os.environ.copy()
-        import pybind11
-        env['pybind11_DIR'] = pybind11.get_cmake_dir()
+        # import pybind11
+        # env['pybind11_DIR'] = pybind11.get_cmake_dir()
+        # env['CXXFLAGS'] = '{} -DVERSION_INFO=\\"{}\\"'.format(env.get('CXXFLAGS', ''),
+        #                                                       self.distribution.get_version())
+        import nanobind
+        env['nanobind_DIR'] = nanobind.cmake_dir()
+        print("*******",  nanobind.cmake_dir())
         env['CXXFLAGS'] = '{} -DVERSION_INFO=\\"{}\\"'.format(env.get('CXXFLAGS', ''),
                                                               self.distribution.get_version())
 
@@ -82,7 +87,8 @@ setup(name='fenics-basix',
       classifiers=[_f for _f in CLASSIFIERS.split('\n') if _f],
       platforms=["Linux", "Mac OS-X", "Unix"],
       packages=["basix"],
-      setup_requires=["pybind11"],
+      #   setup_requires=["pybind11"],
+      setup_requires=["nanobind"],
       ext_modules=[CMakeExtension('basix._basixcpp')],
       cmdclass=dict(build_ext=CMakeBuild),
       zip_safe=False)
