### 2022-04-21 322502b
no HHJ on tetrahedra
