+++ cpp/basix/finite-element.cpp
@@ -410,9 +410,9 @@ basix::FiniteElement basix::create_custom_element(
 
   // Check that wcoeffs have the correct shape
   if (wcoeffs.shape(1) != psize * value_size)
-    throw std::runtime_error("wcoeffs has the wrong shape");
+    throw std::runtime_error("wcoeffs has the wrong number of rows");
   if (wcoeffs.shape(0) != ndofs)
-    throw std::runtime_error("wcoeffs has the wrong shape");
+    throw std::runtime_error("wcoeffs has the wrong number of columns");
 
   // Check that x has the right shape
   for (std::size_t i = 0; i <= 3; ++i)
@@ -421,10 +421,14 @@ basix::FiniteElement basix::create_custom_element(
         != (i > tdim ? 0
                      : static_cast<std::size_t>(
                          cell::num_sub_entities(cell_type, i))))
-      throw std::runtime_error("x has the wrong shape");
+      throw std::runtime_error("x has the wrong number of entities");
     for (std::size_t j = 0; j < x[i].size(); ++j)
+    {
+      if (x[i][j].dimension() != 2)
+        throw std::runtime_error("x has the wrong dimension");
       if (x[i][j].shape(1) != tdim)
-        throw std::runtime_error("x has the wrong shape");
+        throw std::runtime_error("x has a point with the wrong tdim");
+    }
   }
 
   // Check that M has the right shape
@@ -434,13 +438,17 @@ basix::FiniteElement basix::create_custom_element(
         != (i > tdim ? 0
                      : static_cast<std::size_t>(
                          cell::num_sub_entities(cell_type, i))))
-      throw std::runtime_error("M has the wrong shape");
+      throw std::runtime_error("M has the wrong number of entities");
     for (std::size_t j = 0; j < M[i].size(); ++j)
     {
+      if (M[i][j].dimension() != 3)
+        throw std::runtime_error("M has the wrong dimension");
       if (M[i][j].shape(2) != x[i][j].shape(0))
-        throw std::runtime_error("M has the wrong shape");
+        throw std::runtime_error(
+            "M has the wrong shape (dimension 2 is wrong)");
       if (M[i][j].shape(1) != value_size)
-        throw std::runtime_error("M has the wrong shape");
+        throw std::runtime_error(
+            "M has the wrong shape (dimension 1 is wrong)");
     }
   }
 
