### 2022-04-21 fdd3086
Add checks on dimension of entries in x and M when creating custom element (#432)
* Add new tests

* more and more helpful error messages

* Errors in wrapper.cpp

* flake

* rows/cols other way round
