### 2022-04-21 3a565fb
plot degree 4 too
