+++ python/wrapper.cpp
@@ -153,7 +153,7 @@ Interface to the Basix C++ library.
 
   py::enum_<maps::type>(m, "MapType")
       .value("identity", maps::type::identity)
-      .value("scaledIdentity", maps::type::scaledIdentity)
+      .value("L2Piola", maps::type::L2Piola)
       .value("covariantPiola", maps::type::covariantPiola)
       .value("contravariantPiola", maps::type::contravariantPiola)
       .value("doubleCovariantPiola", maps::type::doubleCovariantPiola)
