+++ cpp/basix/maps.h
@@ -19,7 +19,7 @@ namespace basix::maps
 enum class type
 {
   identity = 0,
-  scaledIdentity = 1,
+  L2Piola = 1,
   covariantPiola = 2,
   contravariantPiola = 3,
   doubleCovariantPiola = 4,
@@ -58,13 +58,12 @@ void dot21(Vec&& r, const Mat0& A, const Mat1& B)
 }
 } // namespace impl
 
-/// Scaled identity map
+/// L2 Piola map
 template <typename O, typename P, typename Q, typename R>
-void scaled_identity(O&& r, const P& U, const Q& /*J*/, double detJ,
-                     const R& /*K*/)
+void l2_piola(O&& r, const P& U, const Q& /*J*/, double detJ, const R& /*K*/)
 {
   r.assign(U);
-  std::for_each(r.begin(), r.end(), [detJ](auto& ri) { ri *= detJ; });
+  std::for_each(r.begin(), r.end(), [detJ](auto& ri) { ri /= detJ; });
 }
 
 /// Covariant Piola map
