### 2022-04-22 da6a4fc
Add scaled identity map (#417)
* Add scaled identity map

* scaledIdentity -> scaled_identity

* corrections

* typo

* L2 Piola map
