+++ test/test_custom_element.py
@@ -30,6 +30,29 @@ def test_lagrange_custom_triangle_degree1():
     assert np.allclose(lagrange.base_transformations(), element.base_transformations())
 
 
+def test_lagrange_custom_triangle_degree1_l2piola():
+    """Test a custom element with a L2 Piola map."""
+
+    wcoeffs = np.eye(3)
+    z = np.zeros((0, 2))
+    x = [[np.array([[0., 0.]]), np.array([[1., 0.]]), np.array([[0., 1.]])],
+         [z, z, z], [z], []]
+    z = np.zeros((0, 1, 0))
+    M = [[np.array([[[1.]]]), np.array([[[1.]]]), np.array([[[1.]]])],
+         [z, z, z], [z], []]
+
+    lagrange = basix.create_element(
+        basix.ElementFamily.P, basix.CellType.triangle, 1)
+
+    element = basix.create_custom_element(
+        basix.CellType.triangle, 1, [], wcoeffs,
+        x, M, basix.MapType.L2Piola, False, 1)
+
+    points = basix.create_lattice(basix.CellType.triangle, 5, basix.LatticeType.equispaced, True)
+    assert np.allclose(lagrange.tabulate(1, points), element.tabulate(1, points))
+    assert np.allclose(lagrange.base_transformations(), element.base_transformations())
+
+
 def test_lagrange_custom_triangle_degree4():
     """Test that Lagrange element created as a custom element agrees with built-in Lagrange."""
 
