### 2022-04-22 5215717
Add demo for conforming CR custom element via custom element interface (#433)
* Add demo for conforming CR custom element

* plot degree 4 too

* copyedit

* matplotlib needed to test demos
