### 2022-04-22 c8738ca
Add test that elements of the same family are continuous between different cell types (#434)
* Use Legendre moments by default, add test that same family on traingle and quad is continuous

* Disable sympy tests

* Test quadrilateral facets

* MapType

* Add HHJ to test
