### 2022-04-22 a5c2037
Allow variants to be passed into all elements in P and S families
