+++ cpp/basix/e-nce-rtc.cpp
@@ -160,7 +160,7 @@ FiniteElement basix::element::create_rtc(cell::type celltype, int degree,
 
   return FiniteElement(element::family::RT, celltype, degree, {tdim}, wcoeffs,
                        x, M, maps::type::contravariantPiola, discontinuous,
-                       degree - 1);
+                       degree - 1, {}, lvariant);
 }
 //-----------------------------------------------------------------------------
 FiniteElement basix::element::create_nce(cell::type celltype, int degree,
@@ -351,6 +351,6 @@ FiniteElement basix::element::create_nce(cell::type celltype, int degree,
 
   return FiniteElement(element::family::N1E, celltype, degree, {tdim}, wcoeffs,
                        x, M, maps::type::covariantPiola, discontinuous,
-                       degree - 1);
+                       degree - 1, {}, lvariant);
 }
 //-----------------------------------------------------------------------------
