### 2022-04-22 0645c58
simplify create_finite_element
