### 2022-04-22 0ddb168
use default value for discontinuous rather than overloading
