+++ python/generate_docs.py
@@ -25,7 +25,7 @@ def unreplace(txt):
 
 def remove_types(matches):
     """Remove the types from a function declaration."""
-    vars = [i.strip().split(" ")[-1].split("=")[0] for i in matches[1].split(",")]
+    vars = [i.strip().split(" = ")[0].split(" ")[-1] for i in matches[1].split(",")]
     return "(" + ", ".join(vars) + ")"
 
 
