+++ python/wrapper.cpp
@@ -504,8 +504,7 @@ Interface to the Basix C++ library.
   m.def(
       "create_element",
       [](element::family family_name, cell::type cell_name, int degree,
-         bool discontinuous) -> FiniteElement
-      {
+         bool discontinuous = false) -> FiniteElement {
         return basix::create_element(family_name, cell_name, degree,
                                      discontinuous);
       },
@@ -516,7 +515,7 @@ Interface to the Basix C++ library.
       "create_element",
       [](element::family family_name, cell::type cell_name, int degree,
          element::lagrange_variant lvariant,
-         bool discontinuous) -> FiniteElement {
+         bool discontinuous = false) -> FiniteElement {
         return basix::create_element(family_name, cell_name, degree, lvariant,
                                      discontinuous);
       },
@@ -526,7 +525,8 @@ Interface to the Basix C++ library.
   m.def(
       "create_element",
       [](element::family family_name, cell::type cell_name, int degree,
-         element::dpc_variant dvariant, bool discontinuous) -> FiniteElement {
+         element::dpc_variant dvariant,
+         bool discontinuous = false) -> FiniteElement {
         return basix::create_element(family_name, cell_name, degree, dvariant,
                                      discontinuous);
       },
@@ -537,47 +537,13 @@ Interface to the Basix C++ library.
       "create_element",
       [](element::family family_name, cell::type cell_name, int degree,
          element::lagrange_variant lvariant, element::dpc_variant dvariant,
-         bool discontinuous) -> FiniteElement {
+         bool discontinuous = false) -> FiniteElement {
         return basix::create_element(family_name, cell_name, degree, lvariant,
                                      dvariant, discontinuous);
       },
       basix::docstring::
           create_element__family_cell_degree_lvariant_dvariant_discontinuous
               .c_str());
-
-  m.def(
-      "create_element",
-      [](element::family family_name, cell::type cell_name,
-         int degree) -> FiniteElement
-      { return basix::create_element(family_name, cell_name, degree); },
-      basix::docstring::create_element__family_cell_degree.c_str());
-
-  m.def(
-      "create_element",
-      [](element::family family_name, cell::type cell_name, int degree,
-         element::lagrange_variant lvariant) -> FiniteElement {
-        return basix::create_element(family_name, cell_name, degree, lvariant);
-      },
-      basix::docstring::create_element__family_cell_degree_lvariant.c_str());
-
-  m.def(
-      "create_element",
-      [](element::family family_name, cell::type cell_name, int degree,
-         element::dpc_variant dvariant) -> FiniteElement {
-        return basix::create_element(family_name, cell_name, degree, dvariant);
-      },
-      basix::docstring::create_element__family_cell_degree_dvariant.c_str());
-
-  m.def(
-      "create_element",
-      [](element::family family_name, cell::type cell_name, int degree,
-         element::lagrange_variant lvariant,
-         element::dpc_variant dvariant) -> FiniteElement {
-        return basix::create_element(family_name, cell_name, degree, lvariant,
-                                     dvariant);
-      },
-      basix::docstring::create_element__family_cell_degree_lvariant_dvariant
-          .c_str());
 
   // Interpolate between elements
   m.def(
