+++ cpp/basix/finite-element.h
@@ -1035,19 +1035,6 @@ FiniteElement create_custom_element(
     const std::array<std::vector<xt::xtensor<double, 3>>, 4>& M,
     maps::type map_type, bool discontinuous, int highest_complete_degree);
 
-/// Create an element using a given Lagrange variant
-/// @param[in] family The element family
-/// @param[in] cell The reference cell type that the element is defined on
-/// @param[in] degree The degree of the element
-/// @param[in] lvariant The variant of Lagrange to use
-/// @param[in] discontinuous Indicates whether the element is discontinuous
-/// between cells points of the element. The discontinuous element will have the
-/// same DOFs, but they will all be associated with the interior of the cell.
-/// @return A finite element
-FiniteElement create_element(element::family family, cell::type cell,
-                             int degree, element::lagrange_variant lvariant,
-                             bool discontinuous);
-
 /// Create an element using a given Lagrange variant and a given DPC variant
 /// @param[in] family The element family
 /// @param[in] cell The reference cell type that the element is defined on
@@ -1060,7 +1047,21 @@ FiniteElement create_element(element::family family, cell::type cell,
 /// @return A finite element
 FiniteElement create_element(element::family family, cell::type cell,
                              int degree, element::lagrange_variant lvariant,
-                             element::dpc_variant dvariant, bool discontinuous);
+                             element::dpc_variant dvariant,
+                             bool discontinuous = false);
+
+/// Create an element using a given Lagrange variant
+/// @param[in] family The element family
+/// @param[in] cell The reference cell type that the element is defined on
+/// @param[in] degree The degree of the element
+/// @param[in] lvariant The variant of Lagrange to use
+/// @param[in] discontinuous Indicates whether the element is discontinuous
+/// between cells points of the element. The discontinuous element will have the
+/// same DOFs, but they will all be associated with the interior of the cell.
+/// @return A finite element
+FiniteElement create_element(element::family family, cell::type cell,
+                             int degree, element::lagrange_variant lvariant,
+                             bool discontinuous = false);
 
 /// Create an element using a given DPC variant
 /// @param[in] family The element family
@@ -1073,7 +1074,7 @@ FiniteElement create_element(element::family family, cell::type cell,
 /// @return A finite element
 FiniteElement create_element(element::family family, cell::type cell,
                              int degree, element::dpc_variant dvariant,
-                             bool discontinuous);
+                             bool discontinuous = false);
 
 /// Create an element
 /// @param[in] family The element family
@@ -1086,49 +1087,7 @@ FiniteElement create_element(element::family family, cell::type cell,
 /// with the interior of the cell.
 /// @return A finite element
 FiniteElement create_element(element::family family, cell::type cell,
-                             int degree, bool discontinuous);
-
-/// Create a continuous element using a given Lagrange variant
-/// @param[in] family The element family
-/// @param[in] cell The reference cell type that the element is defined
-/// on
-/// @param[in] degree The degree of the element
-/// @param[in] lvariant The variant of Lagrange to use
-/// @return A finite element
-FiniteElement create_element(element::family family, cell::type cell,
-                             int degree, element::lagrange_variant lvariant);
-
-/// Create a continuous element using a given Lagrange variant and a given DPC
-/// variant
-/// @param[in] family The element family
-/// @param[in] cell The reference cell type that the element is defined
-/// on
-/// @param[in] degree The degree of the element
-/// @param[in] lvariant The variant of Lagrange to use
-/// @param[in] dvariant The variant of DPC to use
-/// @return A finite element
-FiniteElement create_element(element::family family, cell::type cell,
-                             int degree, element::lagrange_variant lvariant,
-                             element::dpc_variant dvariant);
-
[…]

[…]
