### 2022-04-23 f0ec48a
Allow variants to be passed into more elements (#436)
* expand scope of test_create

* Allow variants to be passed into all elements in P and S families

* one more variant in test

* store variants

* ffc branch

* simplify create_finite_element

* ffc branch

* ignore default arguments in docs
