+++ .github/workflows/ffcx-tests.yml
@@ -33,7 +33,7 @@ jobs:
         with:
           path: ./ffcx
           repository: FEniCS/ffcx
-          ref: mscroggs/no-variants
+          ref: main
 
       - name: Install FFCx
         run: |
