+++ demo/python/demo_custom_element.py
@@ -52,9 +52,10 @@ wcoeffs[3, 4] = 1
 # are orthonormal, we can represent this as
 #
 # .. math::
-#    x(1-x)y(1-y) = \sum_{i=0}^9\int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y\; p_i(x, y),
+#    x(1-x)y(1-y) = \sum_{i=0}^8\int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y\; p_i(x, y),
 #
-# and so the coefficients we want to put in the final row of our matrix are:
+# where $p_0$ to $p_8$ are the orthonormal polynomials. Therefore the coefficients we want
+# to put in the final row of our matrix are:
 #
 # .. math::
 #    \int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y.
