### 2022-04-24 83aed93
Make custom element demo clearer (#440)
* order to match implementation

* more demo updates
