+++ demo/python/demo_custom_element.py
@@ -52,9 +52,10 @@ wcoeffs[3, 4] = 1
 # are orthonormal, we can represent this as
 #
 # .. math::
-#    \sum_{i=0}^9\int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y\; p_i(x, y),
+#    x(1-x)y(1-y) = \sum_{i=0}^8\int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y\; p_i(x, y),
 #
-# and so the coefficients we want to put in the final row of our matrix are:
+# where $p_0$ to $p_8$ are the orthonormal polynomials. Therefore the coefficients we want
+# to put in the final row of our matrix are:
 #
 # .. math::
 #    \int_0^1\int_0^1p_i(x, y)x(1-x)y(1-y)\,\mathrm{d}x\,\mathrm{d}y.
@@ -64,7 +65,9 @@ wcoeffs[3, 4] = 1
 
 pts, wts = basix.make_quadrature(CellType.quadrilateral, 4)
 poly = basix.tabulate_polynomials(PolynomialType.legendre, CellType.quadrilateral, 2, pts)
-f = pts[:, 0] * (1 - pts[:, 0]) * pts[:, 1] * (1 - pts[:, 1])
+x = pts[:, 0]
+y = pts[:, 1]
+f = x * (1 - x) * y * (1 - y)
 for i in range(9):
     wcoeffs[4, i] = sum(f * poly[:, i] * wts)
 
@@ -142,7 +145,9 @@ print(element.tabulate(0, points))
 # Degree 1 Ravairt--Thomas element
 # ================================
 #
-# As a second example, we create a degree 1 Raviart--Thomas element on a triangle. This element
+# As a second example, we create a degree 1 Raviart--Thomas element on a triangle. Details of
+# the definition of this element can be found at
+# https://defelement.com/elements/raviart-thomas.html. This element
 # spans:
 #
 # .. math::
@@ -175,15 +180,18 @@ wcoeffs[1, 3] = 1
 
 pts, wts = basix.make_quadrature(CellType.triangle, 2)
 poly = basix.tabulate_polynomials(PolynomialType.legendre, CellType.triangle, 1, pts)
+x = pts[:, 0]
+y = pts[:, 1]
 for i in range(3):
-    wcoeffs[2, i] = sum(pts[:, 0] * poly[:, i] * wts)
-    wcoeffs[2, 3 + i] = sum(pts[:, 1] * poly[:, i] * wts)
+    wcoeffs[2, i] = sum(x * poly[:, i] * wts)
+    wcoeffs[2, 3 + i] = sum(y * poly[:, i] * wts)
 
 # Interpolation
 # -------------
 #
 # For this element, there will be multiple points used per DOF, as the functionals that define
-# the element are integrals. We begin by defining a degree 2 quadrature rule on a triangle.
+# the element are integrals. We begin by defining a degree 1 quadrature rule on an interval.
+# This quadrature rule will be used to integrate on the edges of the triangle.
 
 pts, wts = basix.make_quadrature(CellType.interval, 1)
 
