### 2022-04-25 3432fdf
Start documenting polyset order
