+++ doc/web/polyset-order.md
@@ -1,0 +1,179 @@
+# The ordering of Basix's polynomial sets
+When defining finite elements, Basix uses orthonormal polyomials to represent
+a basis of the space spanned by the finite element functions. This page
+details the order in which these polynomials are stored.
+
+In each example below, we give a list of the monomials that appear for the
+first time in each term. The actual orthogonal polynomials will be linear
+combinations of the monomial and all the previous monomials. For example,
+the monomials
+
+\\(1\\),
+\\(x\\),
+\\(x^2\\),
+\\(x^3\\),
+...
+
+are given below for an interval. The orthonormal polynomials on an interval,
+however, are
+
+\\(1\\),
+\\(\\sqrt{3}(2x-1)\\),
+\\(\\sqrt{5}(6x^2-6x+1)\\),
+\\(\\sqrt{6}(20x^3 - 30x^2 + 12x - 1)\\),
+...
+
+
+### Interval
+On an interval, Basix orders its polynomials by lowest degree first, ie
+
+\\(1\\),
+\\(x\\),
+\\(x^2\\),
+\\(x^3\\),
+...
+
+### Triangle
+On a triangle, Basix orders its polynomials by degree (lowest first).
+Polynomials of the same degree are ordered so that higher powers of \\(x\\)
+appear first, ie
+
+\\(1\\),
+\\(x\\),
+\\(y\\),
+\\(x^2\\),
+\\(xy\\),
+\\(y^2\\),
+\\(x^3\\),
+\\(xy^2\\),
+\\(x^2y\\),
+\\(y^3\\),
+...
+
+### Tetrahedron
+On a tetrahedron, Basix orders its polynomials by degree (lowest first).
+Polynomials of the same degree are ordered so that higher powers of \\(x\\)
+appear first then higher powers of \\(y\\), ie
+
+\\(1\\),
+\\(x\\),
+\\(y\\),
+\\(z\\),
+\\(x^2\\),
+\\(xy\\),
+\\(xz\\),
+\\(y^2\\),
+\\(yz\\),
+\\(z^2\\),
+\\(x^3\\),
+\\(x^2y\\),
+\\(x^2z\\),
+\\(xy^2\\),
+\\(xyz\\),
+\\(xz^2\\),
+\\(y^3\\),
+\\(y^2z\\),
+\\(yz^2\\),
+\\(z^3\\),
+...
+
+### Quadrilateral
+On a quadrilateral, we take a tensor product of the polynomials on an
+interval with \\(y\\) as the variable and those with \\(x\\) as the variable.
+The resulting ordering has polynomials containing only \\(y\\) first
+(low-to-high degree), then these polynomials times \\(x\\), then the same
+polynomials times \\(x^2\\), and so on; ie
+
+\\(1\\),
+\\(y\\),
+\\(y^2\\),
+\\(y^3\\),
+\\(x\\),
+\\(xy\\),
+\\(xy^2\\),
+\\(xy^3\\),
+\\(x^2\\),
+\\(x^2y\\),
+\\(x^2y^2\\),
+\\(x^2y^3\\),
+\\(x^3\\),
+\\(x^3y\\),
+\\(x^3y^2\\),
+\\(x^3y^3\\)
+
+### Hexahedron
[…]
