### 2022-04-25 a5be81a
Write test that ensure what demo says about polysets is true
