+++ test/test_polynomials.py
@@ -52,7 +52,8 @@ def evaluate(function, pt):
     [basix.CellType.tetrahedron, [one, z, y, x], 1],
     [basix.CellType.tetrahedron, [one, z, y, x, z**2, y * z, x * z, y**2, x * y, x**2], 2],
     [basix.CellType.tetrahedron, [one, z, y, x, z**2, y * z, x * z, y**2, x * y, x**2,
-                                  z**3, y * z**2, x * z**2, y**2 * z, x * y * z, x**2 * z, y**3, x * y**2, x**2 * y, x**3], 3],
+                                  z**3, y * z**2, x * z**2, y**2 * z, x * y * z, x**2 * z,
+                                  y**3, x * y**2, x**2 * y, x**3], 3],
     [basix.CellType.quadrilateral, [one, y, x, x * y], 1],
     [basix.CellType.quadrilateral, [one, y, y**2, x, x * y, x * y**2, x**2, x**2 * y, x**2 * y**2], 2],
     [basix.CellType.hexahedron, [one, z, y, y * z, x, x * z, x * y, x * y * z], 1],
