### 2022-04-25 5d5e105
only flip order for polynomial numbering idx
