+++ cpp/basix/indexing.h
@@ -19,7 +19,7 @@ constexpr int idx(int p) { return p; }
 /// @param p Index in x
 /// @param q Index in y
 /// @return 1D Index
-constexpr int idx(int p, int q) { return (p + q + 1) * (p + q) / 2 + p; }
+constexpr int idx(int p, int q) { return (p + q + 1) * (p + q) / 2 + q; }
 
 /// Compute indexing in a 3D tetrahedral array compressed into a 1D array
 /// @param p Index in x
@@ -29,6 +29,6 @@ constexpr int idx(int p, int q) { return (p + q + 1) * (p + q) / 2 + p; }
 constexpr int idx(int p, int q, int r)
 {
   return (p + q + r) * (p + q + r + 1) * (p + q + r + 2) / 6
-         + (p + q) * (p + q + 1) / 2 + p;
+         + (q + r) * (q + r + 1) / 2 + r;
 }
 } // namespace basix::indexing
