+++ test/test_polynomials.py
@@ -40,7 +40,7 @@ def evaluate(function, pt):
         return function.subs(x, pt[0]).subs(y, pt[1]).subs(z, pt[2])
 
 
-# TODO: prism, pyramid
+# TODO: pyramid
 @pytest.mark.parametrize("cell_type, functions, degree", [
     [basix.CellType.interval, [one, x], 1],
     [basix.CellType.interval, [one, x, x**2], 2],
@@ -57,6 +57,12 @@ def evaluate(function, pt):
     [basix.CellType.quadrilateral, [one, y, x, x * y], 1],
     [basix.CellType.quadrilateral, [one, y, y**2, x, x * y, x * y**2, x**2, x**2 * y, x**2 * y**2], 2],
     [basix.CellType.hexahedron, [one, z, y, y * z, x, x * z, x * y, x * y * z], 1],
+    [basix.CellType.hexahedron, [i * j * k for i in [one, x, x**2]
+                                 for j in [one, y, y**2] for k in [one, z, z**2]], 2],
+    [basix.CellType.prism, [one, z, y, y * z, x, x * z], 1],
+    [basix.CellType.prism, [one, z, z**2, y, y * z, y * z**2, x, x * z, x * z**2,
+                            y**2, y**2 * z, y**2 * z**2, x * y, x * y * z, x * y * z**2,
+                            x**2, x**2 * z, x**2 * z**2], 2],
 ])
 def test_order(cell_type, functions, degree):
     points, weights = basix.make_quadrature(cell_type, 2 * degree)
