### 2022-04-26 84e247a
Document polyset order (#443)
* Consistent stylisation of eg and ie (following Guardian style guide)

* Start documenting polyset order

* link to docs

* Write test that ensure what demo says about polysets is true

* fix #442

* update docs with reverse order

* flake

* only flip order for polynomial numbering idx

* Prisms
