+++ cpp/basix/polyset.cpp
@@ -133,38 +133,38 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
     {
       for (std::size_t p = 1; p <= n; ++p)
       {
-        auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 0));
+        auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(0, p));
         const double a
             = static_cast<double>(2 * p - 1) / static_cast<double>(p);
         p0 = ((x0 * 2.0 - 1.0) + 0.5 * (x1 * 2.0 - 1.0) + 0.5)
-             * xt::view(P, idx(kx, ky), xt::all(), idx(p - 1, 0)) * a;
+             * xt::view(P, idx(kx, ky), xt::all(), idx(0, p - 1)) * a;
         if (kx > 0)
         {
           p0 += 2 * kx * a
-                * xt::view(P, idx(kx - 1, ky), xt::all(), idx(p - 1, 0));
+                * xt::view(P, idx(kx - 1, ky), xt::all(), idx(0, p - 1));
         }
 
         if (ky > 0)
         {
-          p0 += ky * a * xt::view(P, idx(kx, ky - 1), xt::all(), idx(p - 1, 0));
+          p0 += ky * a * xt::view(P, idx(kx, ky - 1), xt::all(), idx(0, p - 1));
         }
 
         if (p > 1)
         {
           // y^2 terms
-          p0 -= f3 * xt::view(P, idx(kx, ky), xt::all(), idx(p - 2, 0))
+          p0 -= f3 * xt::view(P, idx(kx, ky), xt::all(), idx(0, p - 2))
                 * (a - 1.0);
           if (ky > 0)
           {
             p0 -= ky * ((x1 * 2.0 - 1.0) - 1.0)
-                  * xt::view(P, idx(kx, ky - 1), xt::all(), idx(p - 2, 0))
+                  * xt::view(P, idx(kx, ky - 1), xt::all(), idx(0, p - 2))
                   * (a - 1.0);
           }
 
           if (ky > 1)
           {
             p0 -= ky * (ky - 1)
-                  * xt::view(P, idx(kx, ky - 2), xt::all(), idx(p - 2, 0))
+                  * xt::view(P, idx(kx, ky - 2), xt::all(), idx(0, p - 2))
                   * (a - 1.0);
           }
         }
@@ -172,27 +172,27 @@ void tabulate_polyset_triangle_derivs(xt::xtensor<double, 3>& P, std::size_t n,
 
       for (std::size_t p = 0; p < n; ++p)
       {
-        auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 0));
-        auto p1 = xt::view(P, idx(kx, ky), xt::all(), idx(p, 1));
+        auto p0 = xt::view(P, idx(kx, ky), xt::all(), idx(0, p));
+        auto p1 = xt::view(P, idx(kx, ky), xt::all(), idx(1, p));
         p1 = p0 * ((x1 * 2.0 - 1.0) * (1.5 + p) + 0.5 + p);
         if (ky > 0)
         {
           p1 += 2 * ky * (1.5 + p)
-                * xt::view(P, idx(kx, ky - 1), xt::all(), idx(p, 0));
+                * xt::view(P, idx(kx, ky - 1), xt::all(), idx(0, p));
         }
 
         for (std::size_t q = 1; q < n - p; ++q)
         {
           const auto [a1, a2, a3] = jrc(2 * p + 1, q);
-          xt::view(P, idx(kx, ky), xt::all(), idx(p, q + 1))
[…]
