### 2022-04-26 d90bfcf
Run demos with savefig when building docs
