### 2022-04-26 523282c
Remove images from demo folder (#445)
* remove images from repo

* Run demos with savefig when building docs

* inside if

* didn't mean to commit that line
