### 2022-04-26 2cfca03
write basis functions into file rather than computing
