+++ test/test_nedelec.py
@@ -7,198 +7,122 @@ import numpy
 import pytest
 import sympy
 
-from .test_lagrange import sympy_lagrange
-
 
 def sympy_nedelec(celltype, n):
     x = sympy.Symbol("x")
     y = sympy.Symbol("y")
     z = sympy.Symbol("z")
 
-    from sympy import S
-    topology = basix.topology(celltype)
-    geometry = S(basix.geometry(celltype).astype(int))
-    dummy = [sympy.Symbol("DUMMY1"), sympy.Symbol("DUMMY2"), sympy.Symbol("DUMMY3")]
-
-    funcs = []
     if celltype == basix.CellType.triangle:
-        tdim = 2
-        for i in range(n):
-            for j in range(n - i):
-                for d in range(2):
-                    funcs += [[x**j * y**i if k == d else 0 for k in range(2)]]
-        for i in range(n):
-            funcs += [[x ** (n - 1 - i) * y ** (i + 1),
-                       -x ** (n - i) * y ** i]]
-        mat = numpy.empty((len(funcs), len(funcs)), dtype=object)
-
-        # edge tangents
         if n == 1:
-            edge_basis = [sympy.Integer(1)]
-        else:
-            edge_basis = sympy_lagrange(basix.CellType.interval, n - 1)
-        edge_basis = [a.subs(x, dummy[0]) for a in edge_basis]
-        for i, f in enumerate(funcs):
-            j = 0
-            for edge in topology[1]:
-                edge_geom = [geometry[t, :] for t in edge]
-                tangent = edge_geom[1] - edge_geom[0]
-                norm = sympy.sqrt(sum(i ** 2 for i in tangent))
-                tangent = [i / norm for i in tangent]
-                param = [(1 - dummy[0]) * a + dummy[0] * b for a, b in zip(edge_geom[0], edge_geom[1])]
-
-                for g in edge_basis:
-                    integrand = sum((f_i * v_i) for f_i, v_i in zip(f, tangent))
-
-                    integrand = integrand.subs(x, param[0]).subs(y, param[1])
-
-                    integrand *= g * norm
-
-                    mat[i, j] = integrand.integrate((dummy[0], 0, 1))
-                    j += 1
-
-        # interior dofs
-        if n > 1:
-            if n == 2:
-                face_basis = [sympy.Integer(1)]
-            else:
-                face_basis = sympy_lagrange(basix.CellType.triangle, n - 2)
-            for i, f in enumerate(funcs):
-                j = n * 3
-                for g in face_basis:
-                    for vec in [(1, 0), (0, 1)]:
[…]
