### 2022-04-26 7314262
speed up second kind test
