+++ test/test_nedelec_second_kind.py
@@ -7,139 +7,105 @@ import numpy
 import pytest
 import sympy
 
-from .test_lagrange import sympy_lagrange
-from .test_rt import sympy_rt
-
 
 def sympy_nedelec(celltype, n):
     x = sympy.Symbol("x")
     y = sympy.Symbol("y")
     z = sympy.Symbol("z")
 
-    from sympy import S
-    topology = basix.topology(celltype)
-    geometry = S(basix.geometry(celltype).astype(int))
-    dummy = [sympy.Symbol("DUMMY1"), sympy.Symbol("DUMMY2"), sympy.Symbol("DUMMY3")]
-
-    funcs = []
     if celltype == basix.CellType.triangle:
-        tdim = 2
-        for i in range(n + 1):
-            for j in range(n + 1 - i):
-                for d in range(2):
-                    funcs += [[x**j * y**i if k == d else 0 for k in range(2)]]
-        mat = numpy.empty((len(funcs), len(funcs)), dtype=object)
+        if n == 1:
+            return [[2*y, 4*x],
+                    [-4*y, -2*x],
+                    [-2*y, -4*x - 6*y + 4],
+                    [4*y, 2*x + 6*y - 2],
+                    [-6*x - 4*y + 4, -2*x],
+                    [6*x + 2*y - 2, 4*x]]
+        if n == 2:
+            return [[12*x*y - 3*y, 18*x**2 - 9*x],
+                    [-18*y**2 + 9*y, -12*x*y + 3*x],
+                    [-9*x*y - 3*y**2/2 + 3*y, 3*x**2/2 + 9*x*y - 3*x],
+                    [12*x*y + 12*y**2 - 9*y, 18*x**2 + 48*x*y - 27*x + 30*y**2 - 36*y + 9],
+                    [18*y**2 - 9*y, 12*x*y - 3*x + 30*y**2 - 24*y + 3],
+                    [-9*x*y - 15*y**2/2 + 6*y, 3*x**2/2 - 15*x*y - 15*y**2 + 15*y - 3/2],
+                    [30*x**2 + 48*x*y - 36*x + 18*y**2 - 27*y + 9, 12*x**2 + 12*x*y - 9*x],
+                    [30*x**2 + 12*x*y - 24*x - 3*y + 3, 18*x**2 - 9*x],
+                    [-15*x**2 - 15*x*y + 15*x + 3*y**2/2 - 3/2, -15*x**2/2 - 9*x*y + 6*x],
+                    [-48*x*y - 12*y**2 + 12*y, -12*x**2 - 48*x*y + 12*x],
+                    [48*x*y + 36*y**2 - 36*y, 12*x**2 + 24*x*y - 12*x],
+                    [-24*x*y - 12*y**2 + 12*y, -36*x**2 - 48*x*y + 36*x]]
+        if n == 3:
+            return [
+                [60*x**2*y - 40*x*y + 4*y, 80*x**3 - 80*x**2 + 16*x],
+                [-80*y**3 + 80*y**2 - 16*y, -60*x*y**2 + 40*x*y - 4*x],
+                [-320*x**2*y/9 - 80*x*y**2/9 + 80*x*y/3 + 40*y**3/27 - 8*y/3, 400*x**3/27 + 640*x**2*y/9 - 80*x**2/3 + 220*x*y**2/9 - 40*x*y + 28*x/3],  # noqa: E501
+                [-220*x**2*y/9 - 640*x*y**2/9 + 40*x*y - 400*y**3/27 + 80*y**2/3 - 28*y/3, -40*x**3/27 + 80*x**2*y/9 + 320*x*y**2/9 - 80*x*y/3 + 8*x/3],  # noqa: E501
[…]
