+++ test/test_lagrange.py
@@ -13,71 +13,171 @@ def sympy_lagrange(celltype, n):
     y = sympy.Symbol("y")
     z = sympy.Symbol("z")
 
-    from sympy import S
-    topology = basix.topology(celltype)
-    geometry = S(basix.geometry(celltype).astype(int))
-    pt = []
-    for dim, entities in enumerate(topology):
-        for ent in entities:
-            entity_geom = [geometry[t, :] for t in ent]
-
-            if (dim == 0):
-                pt += [entity_geom[0]]
-            elif (dim == 1):
-                for i in range(n - 1):
-                    pt += [entity_geom[0]
-                           + sympy.Rational(i + 1, n) * (entity_geom[1] - entity_geom[0])]
-            elif (dim == 2):
-                for i in range(n - 2):
-                    for j in range(n - 2 - i):
-                        pt += [entity_geom[0]
-                               + sympy.Rational(i + 1, n) * (entity_geom[2] - entity_geom[0])
-                               + sympy.Rational(j + 1, n) * (entity_geom[1] - entity_geom[0])]
-            elif (dim == 3):
-                for i in range(n - 3):
-                    for j in range(n - 3 - i):
-                        for k in range(n - 3 - i - j):
-                            pt += [entity_geom[0]
-                                   + sympy.Rational(i + 1, n) * (entity_geom[3] - entity_geom[0])
-                                   + sympy.Rational(j + 1, n) * (entity_geom[2] - entity_geom[0])
-                                   + sympy.Rational(k + 1, n) * (entity_geom[1] - entity_geom[0])]
-
-    funcs = []
     if celltype == basix.CellType.interval:
-        for i in range(n + 1):
-            funcs += [x**i]
+        if n == 1:
+            return [1 - x, x]
+        if n == 2:
+            return [2*x**2 - 3*x + 1, 2*x**2 - x, -4*x**2 + 4*x]
+        if n == 3:
+            return [-9*x**3/2 + 9*x**2 - 11*x/2 + 1,
+                    9*x**3/2 - 9*x**2/2 + x,
+                    27*x**3/2 - 45*x**2/2 + 9*x,
+                    -27*x**3/2 + 18*x**2 - 9*x/2]
+        if n == 4:
+            return [32*x**4/3 - 80*x**3/3 + 70*x**2/3 - 25*x/3 + 1,
+                    32*x**4/3 - 16*x**3 + 22*x**2/3 - x,
+                    -128*x**4/3 + 96*x**3 - 208*x**2/3 + 16*x,
+                    64*x**4 - 128*x**3 + 76*x**2 - 12*x,
[…]
