### 2022-04-27 ce7b4fe
add double contravariant mapping test
