### 2022-04-28 #3982 @wence-
First off, many apologies to both the authors and JOSS for taking such a long time to review here. I had already taken a long time, but a break between jobs where I thought I would get this done needed to be more handsoff than I had anticipated. So I am only now getting round to cleaning the slate.

That said, here are my comments on this paper/submission.

From a software/documentation point of view, I have no real qualms. The two minor installation issues I [mentioned above](https://github.com/openjournals/joss-reviews/issues/3982#issuecomment-1012392461) were swiftly fixed. One very minor thing is that the frontpage of [the docs](https://docs.fenicsproject.org/basix/v0.4.0/) doesn't do quite a job at summarising the point of basix as the top of the README on the github project page. Replicating this information by hand is probably a bad idea but perhaps it can be incorporated programmatically in future docs builds (I don't think this is a sufficient issue to requiring fixing for the paper, but can open an issue to track it if that is normal policy).

From the paper point of view, I think that the survey of related software is a little bit weak. I agree that most finite element libraries do not separate the element library from the rest of the software (and hence there are often not publications specifically about tabulation), however, it would be useful to provide a short summary of alternate approaches and how that in `basix` differs. The paper does a good job contrasting with FIAT, but in many ways `basix` provides effectively the same style of interface as FIAT (namely tabulation of the transformation matrix at a set of points). I think it would be good to contrast with approaches in other popular libraries such as DUNE, deal, NGSolve, MFEM. As a few pointers:
- The NGSolve interface to finite elements is described in https://www.asc.tuwien.ac.at/~schoeberl/wiki/publications/ngs-cpp11.pdf
- The DUNE approach [`dune-localfunctions`](https://gitlab.dune-project.org/core/dune-localfunctions) is described in Chapter 8 of https://link.springer.com/book/10.1007/978-3-030-59702-3
- In deal.ii the description in section 3.3 of http://doi.org/10.1145/1268776.1268779 is still relevant I think (augmented with the documentation at https://dealii.org/current/doxygen/deal.II/group__feaccess.html)
- For MFEM I am a little less sure, but I think a combination of https://doi.org/10.1016/j.camwa.2020.06.009 and https://joss.theoj.org/papers/10.21105/joss.02945 cover things (perhaps @jedbrown can confirm)

I agree that you don't want to recreate too much from the longer [arxiv paper](https://arxiv.org/abs/2102.11901), but I think the paper would be improved by a little bit of discussion here.

One thing that I think should probably be addressed is whether there is a development track towards efficient high-order finite elements that expose structure (typically tensor-product of some kind), or if this is a direction the library intends to go at all. I see there is an [open issue](https://github.com/FEniCS/basix/issues/356) on the interpolation part of this.

Thanks!
