### 2022-05-05 #3982 @mscroggs
@wence- Thanks for your comments. We've made some changes to the paper, I'll rebuilt it here once the CI has finished running and I've merged it.

> One very minor thing is that the frontpage of [the docs](https://docs.fenicsproject.org/basix/v0.4.0/) doesn't do quite a job at summarising the point of basix as the top of the README on the github project page. Replicating this information by hand is probably a bad idea but perhaps it can be incorporated programmatically in future docs builds (I don't think this is a sufficient issue to requiring fixing for the paper, but can open an issue to track it if that is normal policy).

I've made the docs builder copy this info when building the website docs (https://github.com/FEniCS/basix/pull/484)

>
> From the paper point of view, I think that the survey of related software is a little bit weak. I agree that most finite element libraries do not separate the element library from the rest of the software (and hence there are often not publications specifically about tabulation), however, it would be useful to provide a short summary of alternate approaches and how that in `basix` differs. The paper does a good job contrasting with FIAT, but in many ways `basix` provides effectively the same style of interface as FIAT (namely tabulation of the transformation matrix at a set of points). I think it would be good to contrast with approaches in other popular libraries such as DUNE, deal, NGSolve, MFEM.

Thanks for the suggested references, this was very helpful. We've added us much information as we can about other libraries based on what they've written in papers and docs (see changes in https://github.com/FEniCS/basix/pull/481).

> I agree that you don't want to recreate too much from the longer [arxiv paper](https://arxiv.org/abs/2102.11901), but I think the paper would be improved by a little bit of discussion here.

We've added a little bit more to the paragraph on this (see changes in https://github.com/FEniCS/basix/pull/481)

> One thing that I think should probably be addressed is whether there is a development track towards efficient high-order finite elements that expose structure (typically tensor-product of some kind), or if this is a direction the library intends to go at all. I see there is an [open issue](https://github.com/FEniCS/basix/issues/356) on the interpolation part of this.

We've added a paragraph on this (see changes in https://github.com/FEniCS/basix/pull/481)
