### 2022-05-05 FEniCS/basix#485 @tisaac
Define DPC at least one place in documentation
Just about all implemented elements that have a commonly used acronym have that acronym expanded somewhere in the documentation, but I could not find such an expansion for DPC anywhere.

I think that a user who has not seen femtable before would not be able to tell you what DPC stands for.  And it seems like a waste of time to direct them to defelement.com for such a simple to understand element.

The most visible part of the interface where this pops up is `enum dpc_variant`: I would recommend expanding on what is meant by DPC there.
