### 2022-05-05 #3982 @tisaac
I find the paper improved since @wence-'s suggestions were incorporated.

I found once instance of jargon that I felt needed clarifiying in the documentation, but I do not consider this an acceptance blocker (FEniCS/basix#485).
