### 2022-05-06 #3982 @mscroggs
@wence- @tisaac Thanks both! I've made both these changes in https://github.com/FEniCS/basix/pull/486
