### 2022-05-18 #3982 @jedbrown
Wonderful! I just notice one minor copy edit.

> The majority of the Basix’s functionality

remove second of "the" or say "Most of Basix's functionality"

After making this change, could you:
