### 2022-05-19 #3982 @mscroggs
@wence- and @tisaac: Thanks for reviewing this!

@jedbrown. The release tag is `joss` (https://github.com/FEniCS/basix/releases/tag/joss). The DOI of the archive is [10.6084/m9.figshare.19794268.v1](https://dx.doi.org/10.6084/m9.figshare.19794268.v1).
