### 2022-05-19 #3982 @mscroggs
Looks like one of the references in the paper has an extra set of curly braces around the DOI, which is causing an issue. https://github.com/FEniCS/basix/pull/511 will hopefully fix this
